% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_shape_or_scale.R
\name{acc_shape_or_scale}
\alias{acc_shape_or_scale}
\title{Compare observed versus expected distributions}
\usage{
acc_shape_or_scale(
  resp_vars,
  study_data,
  label_col,
  item_level = "item_level",
  dist_col,
  guess,
  par1,
  par2,
  end_digits,
  flip_mode = "noflip",
  meta_data = item_level,
  meta_data_v2
)
}
\arguments{
\item{resp_vars}{\link{variable} the name of the continuous measurement variable}

\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{item_level}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{dist_col}{\link{variable attribute} the name of the variable attribute in
meta_data that provides the expected
distribution of a study variable}

\item{guess}{\link{logical} estimate parameters}

\item{par1}{\link{numeric} first parameter of the distribution if applicable}

\item{par2}{\link{numeric} second parameter of the distribution if applicable}

\item{end_digits}{\link{logical} internal use. check for end digits preferences}

\item{flip_mode}{\link{enum} default | flip | noflip | auto. Should the plot be
in default orientation, flipped, not flipped or
auto-flipped. Not all options are always supported.
In general, this con be controlled by
setting the \code{roptions(dataquieR.flip_mode = ...)}. If
called from \code{dq_report}, you can also pass
\code{flip_mode} to all function calls or set them
specifically using \code{specific_args}.}

\item{meta_data}{\link{data.frame} old name for \code{item_level}}

\item{meta_data_v2}{\link{character} path to workbook like metadata file, see
\code{\link{prep_load_workbook_like_file}} for details.
\strong{ALL LOADED DATAFRAMES WILL BE PURGED},
using \code{\link{prep_purge_data_frame_cache}},
if you specify \code{meta_data_v2}.}
}
\value{
a list with:
\itemize{
\item \code{ResultData}: \link{data.frame} underlying the plot
\item \code{SummaryPlot}: \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} probability distribution plot
\item \code{SummaryTable}: \link{data.frame} with the columns \code{Variables} and \code{FLG_acc_ud_shape}
}
}
\description{
This implementation contrasts the empirical distribution of a measurement
variables against assumed distributions. The approach is adapted from the
idea of rootograms (Tukey 1977) which is also applicable for count data
(Kleiber and Zeileis 2016).

\link{Indicator}
}
\section{ALGORITHM OF THIS IMPLEMENTATION:}{
\itemize{
\item This implementation is restricted to data of type float or integer.
\item Missing codes are removed from resp_vars (if defined in the metadata)
\item The user must specify the column of the metadata containing probability
distribution (currently only: normal, uniform, gamma)
\item Parameters of each distribution can be estimated from the data or are
specified by the user
\item A histogram-like plot contrasts the empirical vs. the technical
distribution
}
}

\seealso{
\href{https://dataquality.qihs.uni-greifswald.de/VIN_acc_impl_shape_or_scale.html}{Online Documentation}
}
