% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_api.R
\docType{data}
\name{API_VERSION}
\alias{API_VERSION}
\title{Version of the API}
\format{
An object of class \code{package_version} (inherits from \code{numeric_version}) of length 1.
}
\usage{
API_VERSION
}
\description{
Version of the API
}
\seealso{
\code{.get_internal_api()}
}
\keyword{internal}
