% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_translate.R
\name{as.character.dataquieR_translated}
\alias{as.character.dataquieR_translated}
\title{\code{as.character} implementation for the class \code{dataquieR_translated}}
\usage{
\method{as.character}{dataquieR_translated}(x, ...)
}
\arguments{
\item{x}{\code{dataquieR_translated} object to print}

\item{...}{passed to base::as.character}
}
\value{
character with only the translated entries
}
\description{
\code{dataquieR}'s translated texts featuring access to the language keys, still.
}
\seealso{
base::as.character
}
