% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.interval.R
\name{as.character.interval}
\alias{as.character.interval}
\title{\code{as.character} implementation for the class \code{interval}}
\usage{
\method{as.character}{interval}(x, ...)
}
\arguments{
\item{x}{\code{interval} objects to convert}

\item{...}{not used yet}
}
\value{
interval as character
}
\description{
such objects, for now, only occur in \code{RECCap} rules, so this function
is meant for internal use, mostly -- for now.
}
\seealso{
base::as.character
}
