% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.dataquieR_resultset.R
\name{as.data.frame.dataquieR_resultset}
\alias{as.data.frame.dataquieR_resultset}
\title{Convert a full \code{dataquieR} report to a \code{data.frame}}
\usage{
\method{as.data.frame}{dataquieR_resultset}(x, ...)
}
\arguments{
\item{x}{Deprecated}

\item{...}{Deprecated}
}
\value{
Deprecated
}
\description{
Deprecated
}
