% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.dataquieR_resultset.R
\name{as.list.dataquieR_resultset}
\alias{as.list.dataquieR_resultset}
\title{Convert a full \code{dataquieR} report to a \code{list}}
\usage{
\method{as.list}{dataquieR_resultset}(x, ...)
}
\arguments{
\item{x}{\link[=dq_report]{dataquieR report}}

\item{...}{arguments passed to \link{pipeline_recursive_result}}
}
\value{
a \link{list} with one element per indicator call. Each element is an
encapsulated sub-list as described
in \link{pipeline_recursive_result}
}
\description{
converts a \link[=dq_report]{dataquieR report} to a \link{list}. Intended for use
in pipelines.
}
