% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{as.list.dataquieR_resultset2}
\alias{as.list.dataquieR_resultset2}
\title{inefficient way to convert a report to a list. try \code{\link[=prep_set_backend]{prep_set_backend()}}}
\usage{
\method{as.list}{dataquieR_resultset2}(x, ...)
}
\arguments{
\item{x}{\link{dataquieR_resultset2}}

\item{...}{no used}
}
\value{
\link{list}
}
\description{
inefficient way to convert a report to a list. try \code{\link[=prep_set_backend]{prep_set_backend()}}
}
