% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{ASSOCIATION_METRIC}
\alias{ASSOCIATION_METRIC}
\title{Cross-item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
ASSOCIATION_METRIC
}
\description{
TODO JM
}
\seealso{
\link{meta_data_cross}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{CONTRADICTION_TYPE}},
\code{\link{DATA_PREPARATION}},
\code{\link{GOLDSTANDARD}},
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}},
\code{\link{N_RULES}},
\code{\link{REL_VAL}},
\code{\link{VARIABLE_LIST}},
\code{\link{util_normalize_cross_item}()}
}
\concept{meta_data_cross}
\keyword{datasets}
