% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_dataquieR_result.R
\name{$.dataquieR_result}
\alias{$.dataquieR_result}
\title{Extract elements of a \code{dataquieR} Result Object}
\usage{
\method{$}{dataquieR_result}(x, ...)
}
\arguments{
\item{x}{the \code{dataquieR} result object}

\item{...}{arguments passed to the implementation for lists.}
}
\value{
the element of the \code{dataquieR} result object with all messages
still attached
}
\description{
Extract elements of a \code{dataquieR} Result Object
}
\seealso{
\link[base:Extract]{base::Extract}
}
\keyword{internal}
