% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{$.dataquieR_resultset2}
\alias{$.dataquieR_resultset2}
\alias{cash-.dataquieR_resultset2}
\title{Access single results from a \link{dataquieR_resultset2} report}
\usage{
\method{$}{dataquieR_resultset2}(x, el)
}
\arguments{
\item{x}{the report}

\item{el}{the index}
}
\value{
the \code{dataquieR} result object
}
\description{
Access single results from a \link{dataquieR_resultset2} report
}
