% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{$<-.dataquieR_resultset2}
\alias{$<-.dataquieR_resultset2}
\alias{cash-set-.dataquieR_resultset2.Rd}
\title{Write single results from a \link{dataquieR_resultset2} report}
\usage{
\method{$}{dataquieR_resultset2}(x, el) <- value
}
\arguments{
\item{x}{the report}

\item{el}{the index}

\item{value}{the single result}
}
\value{
the \code{dataquieR} result object
}
\description{
Write single results from a \link{dataquieR_resultset2} report
}
