% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{cause_label_df}
\alias{cause_label_df}
\alias{missing_matchtable}
\alias{CODE_VALUE}
\alias{CODE_LABEL}
\alias{CODE_CLASS}
\alias{CODE_INTERPRET}
\title{Data frame with labels for missing- and jump-codes}
\description{
\link{data.frame} with the following columns:
\itemize{
\item \code{CODE_VALUE}: \link{numeric} | \link{DATETIME} Missing code
(the number or date representing a missing)
\item \code{CODE_LABEL}: \link{character} a label for the missing code
\item \code{CODE_CLASS}: \link{enum} JUMP | MISSING. Class of the missing code.
\item \code{CODE_INTERPRET} \link{enum} I | P | PL | R | BO | NC | O | UH | UO | NE.
Class of the missing code according to
\href{https://aapor.org/standards-and-ethics/standard-definitions/}{\code{AAPOR}}.
\item \code{resp_vars}: \link{character} optional, if a missing code is specific for some
variables, it is listed for each such variable
with one entry in \code{resp_vars}, If \code{NA}, the
code is assumed shared among all variables.
For v1.0 metadata, you need to refer to
\code{VAR_NAMES} here.
}
}
\seealso{
\href{https://dataquality.qihs.uni-greifswald.de/Item_Level_Metadata.html#MISSING_LIST_TABLE}{Online}
}
