% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{CODE_CLASSES}
\alias{CODE_CLASSES}
\title{types of value codes}
\format{
An object of class \code{list} of length 3.
}
\usage{
CODE_CLASSES
}
\description{
types of value codes
}
\keyword{datasets}
