% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{CODE_LIST_TABLE}
\alias{CODE_LIST_TABLE}
\title{Default Name of the Table featuring Code Lists}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
CODE_LIST_TABLE

CODE_LIST_TABLE
}
\description{
Default Name of the Table featuring Code Lists

Metadata sheet name containing VALUE_LABEL_TABLES
This metadata sheet can contain both value labels of several
VALUE_LABEL_TABLE and also Missing and JUMP tables
}
\keyword{datasets}
