% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{CODE_ORDER}
\alias{CODE_ORDER}
\title{Only existence is checked, order not yet used}
\format{
An object of class \code{character} of length 1.
}
\usage{
CODE_ORDER
}
\description{
Only existence is checked, order not yet used
}
\keyword{datasets}
