% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{COMPUTATION_RULE}
\alias{COMPUTATION_RULE}
\title{Cross-item level metadata attribute name}
\usage{
COMPUTATION_RULE
}
\description{
Cross-item level metadata attribute name
}
\seealso{
\link{meta_data_computation}
}
\concept{meta_data_computation}
