% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{COMPUTED_VARIABLE_ROLES}
\alias{COMPUTED_VARIABLE_ROLES}
\title{\code{SSI} related Cross-item level metadata attribute names
Computed Variable roles can be one of the following:}
\description{
\itemize{
\item \code{MAXIMUM_LONG_STRING} Social Science: Computed Indicator Variable,
maximum long string
\item \code{IRV} Social Science: Computed Indicator Variable, \code{IRV}
\item \code{TOTRESPT} Social Science: Computed Indicator Variable, \code{TOTRESPT}
\item \code{RESPT_PER_ITEM} Social Science: Computed Indicator Variable, \code{RESPT_PER_ITEM}
\item \code{RELCOMPL_SPEED} Social Science: Computed Indicator Variable, \code{RELCOMPL_SPEED}
\item \code{MISS_RESP} Social Science: Computed Indicator Variable, \code{MISS_RESP}
\item \code{NA} Social Science: Computed Indicator Variable -- N/A
}
}
\seealso{
\link{VARIABLE_ROLES}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_METRIC}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{CONTRADICTION_TYPE}},
\code{\link{DATA_PREPARATION}},
\code{\link{GOLDSTANDARD}},
\code{\link{IRV}},
\code{\link{MAHALANOBIS_THRESHOLD}},
\code{\link{MAXIMUM_LONG_STRING}},
\code{\link{MISS_RESP}},
\code{\link{MULTIVARIATE_OUTLIER_CHECK}},
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}},
\code{\link{RELCOMPL_SPEED}},
\code{\link{REL_VAL}},
\code{\link{RESPT_PER_ITEM}},
\code{\link{SCALE_ACRONYM}},
\code{\link{SCALE_NAME}},
\code{\link{TOTRESPT}},
\code{\link{VARIABLE_LIST}},
\code{\link{VARIABLE_LIST_ORDER}},
\code{\link{meta_data_computation}},
\code{\link{meta_data_cross}}

Other SSI: 
\code{\link{IRV}},
\code{\link{MAXIMUM_LONG_STRING}},
\code{\link{MISS_RESP}},
\code{\link{RELCOMPL_SPEED}},
\code{\link{RESPT_PER_ITEM}},
\code{\link{TOTRESPT}}
}
\concept{SSI}
\concept{meta_data_cross}
