% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_contradiction_rules.R
\docType{data}
\name{contradiction_functions}
\alias{contradiction_functions}
\title{contradiction_functions}
\format{
An object of class \code{list} of length 11.
}
\usage{
contradiction_functions
}
\description{
Detect abnormalities help functions
}
\details{
2 variables:
\itemize{
\item \code{A_not_equal_B}, if \code{A != B}
\item \code{A_greater_equal_B}, if \code{A >= B}
\item \code{A_greater_than_B}, if \code{A > B}
\item \code{A_less_than_B}, if \code{A < B}
\item \code{A_less_equal_B}, if \code{A <= B}
\item \code{A_present_not_B}, if \code{A & is.na(B)}
\item \code{A_present_and_B}, if \code{A & !(is.na(B))}
\item \code{A_present_and_B_levels}, if \verb{A & B  \%in\% \{set of levels\}}
\item \code{A_levels_and_B_gt_value}, if \verb{A \%in\% \{set of levels\} & B >  value}
\item \code{A_levels_and_B_lt_value}, if \verb{A \%in\% \{set of levels\} & B <  value}
\item \code{A_levels_and_B_levels}, if
\verb{A \%in\% \{set of levels\} & B \%in\% \{set of levels\}}
}
}
\keyword{datasets}
