% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_contradiction_rules.R
\name{contradiction_functions_descriptions}
\alias{contradiction_functions_descriptions}
\title{description of the contradiction functions}
\usage{
contradiction_functions_descriptions
}
\description{
description of the contradiction functions
}
