% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{CONTRADICTION_TYPE}
\alias{CONTRADICTION_TYPE}
\title{Cross-item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
CONTRADICTION_TYPE
}
\description{
Specifies the type of a contradiction. According to the data quality
concept, there are logical and empirical contradictions, see
\href{https://dataquality.qihs.uni-greifswald.de/VIN_con_impl_contradictions_redcap.html}{online vignette}
}
\seealso{
\link{meta_data_cross}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_METRIC}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{COMPUTED_VARIABLE_ROLES}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{DATA_PREPARATION}},
\code{\link{GOLDSTANDARD}},
\code{\link{IRV}},
\code{\link{MAHALANOBIS_THRESHOLD}},
\code{\link{MAXIMUM_LONG_STRING}},
\code{\link{MISS_RESP}},
\code{\link{MULTIVARIATE_OUTLIER_CHECK}},
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}},
\code{\link{RELCOMPL_SPEED}},
\code{\link{REL_VAL}},
\code{\link{RESPT_PER_ITEM}},
\code{\link{SCALE_ACRONYM}},
\code{\link{SCALE_NAME}},
\code{\link{TOTRESPT}},
\code{\link{VARIABLE_LIST}},
\code{\link{VARIABLE_LIST_ORDER}},
\code{\link{meta_data_computation}},
\code{\link{meta_data_cross}}
}
\concept{meta_data_cross}
\keyword{datasets}
