% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{DATA_PREPARATION}
\alias{DATA_PREPARATION}
\title{Cross-item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
DATA_PREPARATION
}
\description{
For contradiction rules, the required pre-processing steps that can be given.
TODO JM: MISSING_LABEL will not work for non-factor variables
}
\details{
LABEL LIMITS MISSING_NA MISSING_LABEL MISSING_INTERPRET
}
\seealso{
\link{meta_data_cross}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_METRIC}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{CONTRADICTION_TYPE}},
\code{\link{GOLDSTANDARD}},
\code{\link{MULTIVARIATE_OUTLIER_CHECK}},
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}},
\code{\link{N_RULES}},
\code{\link{REL_VAL}},
\code{\link{VARIABLE_LIST}},
\code{\link{meta_data_cross}},
\code{\link{util_normalize_cross_item}()}
}
\concept{meta_data_cross}
\keyword{datasets}
