% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{DATA_TYPES}
\alias{DATA_TYPES}
\alias{integer}
\alias{string}
\alias{float}
\alias{datetime}
\alias{numeric}
\alias{enum}
\alias{FLOAT}
\alias{INTEGER}
\alias{STRING}
\alias{DATETIME}
\alias{variable}
\title{Data Types}
\format{
An object of class \code{list} of length 4.
}
\usage{
DATA_TYPES
}
\description{
\subsection{Data Types of Study Data}{

In the metadata, the following entries are allowed for the
\link{variable attribute} \link{DATA_TYPE}:
}
}
\details{
\itemize{
\item \code{integer} for integer numbers
\item \code{string} for text/string/character data
\item \code{float} for decimal/floating point numbers
\item \code{datetime} for timepoints
}
\subsection{Data Types of Function Arguments}{

As function arguments, \link{dataquieR} uses additional type specifications:
\itemize{
\item \code{numeric} is a numerical value (\link{float} or \link{integer}), but it is not an
allowed \code{DATA_TYPE} in the metadata. However, some functions may accept
\code{float} or \code{integer} for specific function arguments. This is, where we
use the term \code{numeric}.
\item \code{enum} allows one element out of a set of allowed options similar to
\link{match.arg}
\item \code{variable} Function arguments of this type expect a character scalar that
specifies one variable using the variable identifier given in
the meta data attribute \code{VAR_NAMES} or, if \code{label_col} is set,
given in the meta data attribute given in that argument.
Labels can easily be translated using \link{prep_map_labels}
\item \verb{variable list} Function arguments of this type expect a character vector
that specifies variables using the variable identifiers
given in the meta data attribute \code{VAR_NAMES} or,
if \code{label_col} is set, given in the meta data attribute
given in that argument. Labels can easily be translated
using \link{prep_map_labels}
}
}
}
\seealso{
\link{integer} \link{string}
}
\keyword{datasets}
\alias{variable list}
