% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{DATA_TYPES_OF_R_TYPE}
\alias{DATA_TYPES_OF_R_TYPE}
\title{All available data types, mapped from their respective
R types}
\usage{
DATA_TYPES_OF_R_TYPE
}
\description{
All available data types, mapped from their respective
R types
}
\seealso{
\code{\link{prep_dq_data_type_of}}
}
