% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_exception_classes.R
\name{dataquieR.applicability_problem}
\alias{dataquieR.applicability_problem}
\title{An exception class assigned for exceptions caused by trying to apply
a non-applicable indicator function}
\usage{
dataquieR.applicability_problem
}
\description{
Amending metadata could make the function running, e.g., a test
for missingness without any declared missing codes
}
