% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_options.R
\docType{data}
\name{dataquieR.grading_formats}
\alias{dataquieR.grading_formats}
\title{Name of the \link{data.frame} featuring a format for grading-values}
\description{
TODO
}
\seealso{
Other options: 
\code{\link{dataquieR}},
\code{\link{dataquieR.CONDITIONS_LEVEL_TRHESHOLD}},
\code{\link{dataquieR.CONDITIONS_WITH_STACKTRACE}},
\code{\link{dataquieR.ELEMENT_MISSMATCH_CHECKTYPE}},
\code{\link{dataquieR.ERRORS_WITH_CALLER}},
\code{\link{dataquieR.GAM_for_LOESS}},
\code{\link{dataquieR.MAX_LABEL_LEN}},
\code{\link{dataquieR.MAX_VALUE_LABEL_LEN}},
\code{\link{dataquieR.MESSAGES_WITH_CALLER}},
\code{\link{dataquieR.MULTIVARIATE_OUTLIER_CHECK}},
\code{\link{dataquieR.VALUE_LABELS_htmlescaped}},
\code{\link{dataquieR.WARNINGS_WITH_CALLER}},
\code{\link{dataquieR.acc_loess.exclude_constant_subgroups}},
\code{\link{dataquieR.acc_loess.mark_time_points}},
\code{\link{dataquieR.acc_loess.min_bw}},
\code{\link{dataquieR.acc_loess.min_proportion}},
\code{\link{dataquieR.acc_loess.plot_format}},
\code{\link{dataquieR.acc_loess.plot_observations}},
\code{\link{dataquieR.acc_margins_num}},
\code{\link{dataquieR.acc_margins_sort}},
\code{\link{dataquieR.acc_multivariate_outlier.scale}},
\code{\link{dataquieR.col_con_con_empirical}},
\code{\link{dataquieR.col_con_con_logical}},
\code{\link{dataquieR.debug}},
\code{\link{dataquieR.des_summary_hard_lim_remove}},
\code{\link{dataquieR.dontwrapresults}},
\code{\link{dataquieR.fix_column_type_on_read}},
\code{\link{dataquieR.flip_mode}},
\code{\link{dataquieR.force_item_specific_missing_codes}},
\code{\link{dataquieR.force_label_col}},
\code{\link{dataquieR.grading_rulesets}},
\code{\link{dataquieR.guess_missing_codes}},
\code{\link{dataquieR.lang}},
\code{\link{dataquieR.max_group_var_levels_in_plot}},
\code{\link{dataquieR.max_group_var_levels_with_violins}},
\code{\link{dataquieR.min_obs_per_group_var_in_plot}},
\code{\link{dataquieR.non_disclosure}},
\code{\link{dataquieR.progress_fkt}},
\code{\link{dataquieR.progress_msg_fkt}},
\code{\link{dataquieR.scale_level_heuristics_control_binaryrecodelimit}},
\code{\link{dataquieR.scale_level_heuristics_control_metriclevels}},
\code{\link{dataquieR.testdebug}}
}
\concept{options}
