% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_exception_classes.R
\name{dataquieR.intrinsic_applicability_problem}
\alias{dataquieR.intrinsic_applicability_problem}
\title{An exception class assigned for exceptions caused by trying to apply
a non-applicable indicator function, which is not caused by deficient
metadata}
\usage{
dataquieR.intrinsic_applicability_problem
}
\description{
Also amending meta data could not make the function running, e.g., a test
for numbers applied to a character.
}
