% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_options.R
\docType{data}
\name{dataquieR.study_data_cache_metrics_env_default}
\alias{dataquieR.study_data_cache_metrics_env_default}
\title{Default space for some metrics during report computation}
\format{
An object of class \code{environment} of length 0.
}
\usage{
dataquieR.study_data_cache_metrics_env_default
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other study_data_cache: 
\code{\link{dataquieR.ignore_empty_vars}},
\code{\link{dataquieR.precomputeStudyData}},
\code{\link{dataquieR.print_block_load_factor}},
\code{\link{dataquieR.study_data_cache_max}},
\code{\link{dataquieR.study_data_cache_metrics}},
\code{\link{dataquieR.study_data_cache_metrics_env}},
\code{\link{dataquieR.study_data_cache_quick_fill}}
}
\concept{study_data_cache}
\keyword{datasets}
