% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset.R
\name{dataquieR_resultset}
\alias{dataquieR_resultset}
\title{Internal constructor for the internal class \link{dataquieR_resultset}.}
\usage{
dataquieR_resultset(...)
}
\arguments{
\item{...}{properties stored in the object}
}
\value{
an object of the class \link{dataquieR_resultset}.
}
\description{
creates an object of the class \link{dataquieR_resultset}.
}
\details{
The class features the following methods:
\itemize{
\item \link{as.data.frame.dataquieR_resultset}
\item \link{as.list.dataquieR_resultset}
\item \link{print.dataquieR_resultset}
\item \link{summary.dataquieR_resultset}`
}
}
\seealso{
\link{dq_report}
}
