% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\docType{class}
\name{dataquieR_resultset2-class}
\alias{dataquieR_resultset2-class}
\alias{dataquieR_resultset2}
\alias{.dataquieR_resultset2}
\title{Class \link{dataquieR_resultset2}.}
\description{
Class \link{dataquieR_resultset2}.
}
\seealso{
\link{dq_report2}
}
