% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{dataquieR_resultset2}
\alias{dataquieR_resultset2}
\alias{.dataquieR_resultset2}
\title{Internal constructor for the internal class \link{dataquieR_resultset2}.}
\usage{
dataquieR_resultset2(...)
}
\arguments{
\item{...}{properties stored in the object}
}
\value{
an object of the class \link{dataquieR_resultset2}.
}
\description{
creates an object of the class \link{dataquieR_resultset2}.
}
\details{
The class features the following methods:
\itemize{
\item [\verb{[.dataquieR_resultset2}]
\item \code{\link{dim.dataquieR_resultset2}}
\item \code{\link{dimnames.dataquieR_resultset2}}
\item \code{\link{prep_extract_summary.dataquieR_resultset2}}
\item \code{\link{print.dataquieR_resultset2}}
\item \code{\link{resnames.dataquieR_resultset2}}
\item \code{\link{summary.dataquieR_resultset2}}
}
}
\seealso{
\link{dq_report2}
}
\keyword{internal}
