% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset_verify.R
\name{dataquieR_resultset_verify}
\alias{dataquieR_resultset_verify}
\title{Verify an object of class dataquieR_resultset}
\usage{
dataquieR_resultset_verify(list_to_verify)
}
\arguments{
\item{list_to_verify}{object to be checked}
}
\value{
\code{invisible(TRUE)} -- stops on errors.
}
\description{
stops on errors
}
