% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection.R
\docType{data}
\name{Descriptor}
\alias{Descriptor}
\title{Descriptor Function}
\format{
An object of class \code{character} of length 1.
}
\usage{
Descriptor
}
\description{
A function that returns some figure or table to assess data quality, but it
does not return a value correlating with the magnitude of a data quality
problem. It's the opposite of an \link{Indicator}.

The object \code{Descriptor} only contains the name used internally to tag
such functions.
}
\seealso{
\link{Indicator}
}
\keyword{internal}
