% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{DF_ELEMENT_COUNT}
\alias{DF_ELEMENT_COUNT}
\title{Data frame level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
DF_ELEMENT_COUNT
}
\description{
Number of expected data elements in a data frame.
\link{numeric}. Check only conducted if number entered
}
\seealso{
\link{meta_data_dataframe}
}
\keyword{datasets}
