% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{DF_ID_REF_TABLE}
\alias{DF_ID_REF_TABLE}
\title{Data frame level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
DF_ID_REF_TABLE
}
\description{
The name of the data frame containing the reference IDs to be
compared with the IDs in the study data set.
}
\seealso{
\link{meta_data_dataframe}
}
\keyword{datasets}
