% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{DF_ID_VARS}
\alias{DF_ID_VARS}
\title{Data frame level metadata attribute name}
\usage{
DF_ID_VARS
}
\description{
All variables that are to be used as one single ID
variable (combined key) in a data frame.
}
\seealso{
\link{meta_data_dataframe}
}
