% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{DF_NAME}
\alias{DF_NAME}
\title{Data frame level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
DF_NAME
}
\description{
Name of the data frame
}
\seealso{
\link{meta_data_dataframe}
}
\keyword{datasets}
