% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{DF_RECORD_COUNT}
\alias{DF_RECORD_COUNT}
\title{Data frame level metadata attribute name}
\usage{
DF_RECORD_COUNT
}
\description{
Number of expected data records in a data frame.
\link{numeric}. Check only conducted if number entered
}
\seealso{
\link{meta_data_dataframe}
}
