% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{DF_UNIQUE_ID}
\alias{DF_UNIQUE_ID}
\title{Data frame level metadata attribute name}
\usage{
DF_UNIQUE_ID
}
\description{
Defines expectancies on the uniqueness of the IDs
across the rows of a data frame, or the number of times some ID can be repeated.
}
\seealso{
\link{meta_data_dataframe}
}
