% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{DF_UNIQUE_ROWS}
\alias{DF_UNIQUE_ROWS}
\title{Data frame level metadata attribute name}
\usage{
DF_UNIQUE_ROWS
}
\description{
Specifies whether identical data is permitted across rows in a
data frame (excluding ID variables)
}
\seealso{
\link{meta_data_dataframe}
}
