% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.dataquieR_resultset2.R
\name{dim.dataquieR_resultset2}
\alias{dim.dataquieR_resultset2}
\title{Get the dimensions of a \code{dq_report2} result}
\usage{
\method{dim}{dataquieR_resultset2}(x)
}
\arguments{
\item{x}{a \code{dataquieR_resultset2} result}
}
\value{
dimensions
}
\description{
Get the dimensions of a \code{dq_report2} result
}
