% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{dimensions}
\alias{dimensions}
\title{Names of DQ dimensions}
\format{
An object of class \code{character} of length 3.
}
\usage{
dimensions
}
\value{
Only a definition, not a function, so no return value
}
\description{
a vector of data quality dimensions. The supported
dimensions are Completeness, Consistency and Accuracy.
}
\seealso{
\href{https://dataquality.ship-med.uni-greifswald.de/DQconceptNew.html}{Data Quality Concept}
}
\keyword{datasets}
