% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames_dataquieR_resultset2.R
\name{dimnames.dataquieR_resultset2}
\alias{dimnames.dataquieR_resultset2}
\title{Names of a \code{dataquieR} report object (v2.0)}
\usage{
\method{dimnames}{dataquieR_resultset2}(x)
}
\arguments{
\item{x}{the result object}
}
\value{
the names
}
\description{
Names of a \code{dataquieR} report object (v2.0)
}
