% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{dims}
\alias{dims}
\title{Dimension Titles for Prefixes}
\usage{
dims
}
\description{
order does matter, because it defines the order in the \code{dq_report2}.
}
\seealso{
\code{util_html_for_var()}

\code{util_html_for_dims()}
}
