% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{DISTRIBUTIONS}
\alias{DISTRIBUTIONS}
\title{All available probability distributions for
\link{acc_shape_or_scale}}
\usage{
DISTRIBUTIONS
}
\description{
\itemize{
\item \code{uniform} For uniform distribution
\item \code{normal} For Gaussian distribution
\item \code{gamma} For a gamma distribution
}
}
