% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection.R
\docType{data}
\name{..indicator_or_descriptor}
\alias{..indicator_or_descriptor}
\title{Holds Indicator .// Descriptor assignments from the manual at run-time}
\format{
An object of class \code{environment} of length 0.
}
\usage{
..indicator_or_descriptor
}
\description{
Holds Indicator .// Descriptor assignments from the manual at run-time
}
\keyword{internal}
