% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection.R
\docType{data}
\name{..manual}
\alias{..manual}
\title{Holds parts of the manual at run-time}
\format{
An object of class \code{environment} of length 0.
}
\usage{
..manual
}
\description{
Holds parts of the manual at run-time
}
\keyword{internal}
