% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{.access_dq_rs2}
\alias{.access_dq_rs2}
\title{Access elements from a \code{dataquieR_resultset2}}
\usage{
.access_dq_rs2(x, els)
}
\arguments{
\item{x}{the \code{dataquieR_resultset2}}

\item{els}{the selector (character, number or logical)}
}
\value{
the sub-list of \code{x}
}
\description{
does so, but similar to \code{[} for lists.
}
\keyword{internal}
