% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_api.R
\name{.get_internal_api}
\alias{.get_internal_api}
\title{Get Access to Utility Functions}
\usage{
.get_internal_api(fkt, version = API_VERSION, or_newer = TRUE)
}
\arguments{
\item{fkt}{function name}

\item{version}{version number to get}
}
\value{
an API object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
