% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOT.template_function_indicator.R
\name{.template_function_indicator}
\alias{.template_function_indicator}
\title{\code{Roxygen}-Template for indicator functions}
\usage{
.template_function_indicator(
  resp_vars,
  study_data,
  label_col,
  item_level,
  meta_data,
  meta_data_v2,
  meta_data_dataframe,
  meta_data_segment,
  dataframe_level,
  segment_level
)
}
\arguments{
\item{resp_vars}{\link{variable} the names of the measurement variables, if
missing or \code{NULL}, all variables will be checked}

\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{item_level}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{meta_data}{\link{data.frame} old name for \code{item_level}}

\item{meta_data_v2}{\link{character} path to workbook like metadata file, see
\code{\link{prep_load_workbook_like_file}} for details.
\strong{ALL LOADED DATAFRAMES WILL BE PURGED},
using \code{\link{prep_purge_data_frame_cache}},
if you specify \code{meta_data_v2}.}

\item{meta_data_dataframe}{\link{data.frame} the data frame that contains the
metadata for the data frame level}

\item{meta_data_segment}{\link{data.frame} -- optional: Segment level metadata}

\item{dataframe_level}{\link{data.frame} alias for \code{meta_data_dataframe}}

\item{segment_level}{\link{data.frame} alias for \code{meta_data_segment}}
}
\value{
\code{invisible(NULL)}
}
\description{
\code{Roxygen}-Template for indicator functions
}
\keyword{internal}
