% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_meta_data_v1_to_item_level_meta_data.R
\name{.util_internal_normalize_meta_data}
\alias{.util_internal_normalize_meta_data}
\title{Make normalizations of v2.0 item_level metadata.}
\usage{
.util_internal_normalize_meta_data(meta_data = "item_level", verbose = TRUE)
}
\arguments{
\item{meta_data}{\link{data.frame} the old item-level-metadata}

\item{verbose}{\link{logical} display all estimated decisions}
}
\description{
Requires referred missing-tables being available by
\code{prep_get_data_frame}.
}
