% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_correct_variable_use.R
\name{.variable_arg_roles}
\alias{.variable_arg_roles}
\title{Variable-argument roles}
\usage{
.variable_arg_roles
}
\description{
A Variable-argument role is the intended use of an argument of a indicator
function -- an argument that refers variables.
In general for the table .variable_arg_roles, the suffix _var means one
variable allowed,
while _vars means more than one. The default  sets of arguments
for \code{util_correct_variable_use}/\code{util_correct_variable_use2} are defined
from the point of usage, e.g. if it could be, that NAs are in
the list of variable names, the function should be able to remove certain
response variables
from the output and not disallow them by setting \code{allow_na} to \code{FALSE}.
}
\seealso{
\code{util_correct_variable_use()}

\code{util_correct_variable_use2()}
}
\keyword{internal}
