% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_create_lean_ggplot.R, R/util_lazy_ggplot.R
\name{ggplotGrob.dq_lazy_ggplot}
\alias{ggplotGrob.dq_lazy_ggplot}
\alias{ggplotly.dq_lazy_ggplot_s7}
\alias{plotly_build.dq_lazy_ggplot_s7}
\alias{ggplotly.dq_lazy_ggplot}
\alias{plotly_build.dq_lazy_ggplot}
\alias{dq_lazy_ggplot_methods}
\alias{ggplotGrob.dq_lazy_ggplot_s7}
\title{S3/S7 methods for lazy ggplot objects}
\usage{
ggplotGrob.dq_lazy_ggplot(x, ...)

ggplotly.dq_lazy_ggplot_s7(p, ...)

plotly_build.dq_lazy_ggplot_s7(p, ...)

ggplotly.dq_lazy_ggplot(p, ...)

plotly_build.dq_lazy_ggplot(p, ...)

ggplotGrob.dq_lazy_ggplot_s7(x, ...)
}
\arguments{
\item{x, p}{A \code{dq_lazy_ggplot} object.}

\item{...}{Further arguments passed on to the underlying generic.}
}
\value{
The return value is the same as for the corresponding generic:
\itemize{
\item \code{ggplotGrob()} returns a \code{gtable} object.
\item \code{ggplotly()} returns a \code{plotly} object.
\item \code{plotly_build()} returns a \code{plotly_proxy} or similar.
}
}
\description{
These S3/S7 methods make \code{dq_lazy_ggplot}/\code{dq_lazy_ggplot_s7}
objects work smoothly with
functions from \pkg{ggplot2} and \pkg{plotly}. They simply materialize
the underlying \code{ggplot} object and then delegate to the respective
generic.
}
\seealso{
\code{\link[ggplot2]{ggplotGrob}},
\code{plotly::ggplotly}
\code{plotly::plotly_build}
}
