% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels.ReportSummaryTable.R
\name{droplevels.ReportSummaryTable}
\alias{droplevels.ReportSummaryTable}
\title{Remove unused levels from \code{ReportSummaryTable}}
\usage{
\method{droplevels}{ReportSummaryTable}(x, ...)
}
\arguments{
\item{x}{\code{ReportSummaryTable} an object from which to drop unused factor
levels.}

\item{...}{no used.}
}
\value{
\code{ReportSummaryTable} with all (\code{NA} or \code{0})-columns removed
}
\description{
Remove unused levels from \code{ReportSummaryTable}
}
