% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pairs_matrix.R
\name{grid.draw.util_pairs_ggplot_panels}
\alias{grid.draw.util_pairs_ggplot_panels}
\title{\code{grid.draw} method for \code{util_pairs_ggplot_panels} objects}
\usage{
\method{grid.draw}{util_pairs_ggplot_panels}(x, ...)
}
\arguments{
\item{x}{An object of class \code{util_pairs_ggplot_panels}.}

\item{...}{Ignored.}
}
\description{
\code{grid.draw} method for \code{util_pairs_ggplot_panels} objects
}
