% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependency_clipboard.R
\name{html_dependency_clipboard}
\alias{html_dependency_clipboard}
\title{HTML Dependency for report headers in \code{clipboard}}
\usage{
html_dependency_clipboard()
}
\value{
the dependency
}
\description{
HTML Dependency for report headers in \code{clipboard}
}
