% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependency_dataquieR.R
\name{html_dependency_dataquieR}
\alias{html_dependency_dataquieR}
\title{HTML Dependency for \code{dataquieR}}
\usage{
html_dependency_dataquieR(iframe = FALSE)
}
\arguments{
\item{iframe}{\link{logical}\code{(1)} if \code{TRUE}, create the dependency used in figure
\code{iframes}.}
}
\value{
the dependency
}
\description{
generate all dependencies used in static \code{dataquieR} reports
}
