% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependency_jspdf.R
\name{html_dependency_jspdf}
\alias{html_dependency_jspdf}
\title{HTML dependency for \code{jsPDF}}
\usage{
html_dependency_jspdf()
}
\value{
An \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} object
}
\description{
Provides \code{jsPDF} for use in \code{Shiny} or \code{RMarkdown} via \code{htmltools}.
}
