% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependency_report_dt.R
\name{html_dependency_report_dt}
\alias{html_dependency_report_dt}
\title{HTML Dependency for report headers in \code{DT::datatable}}
\usage{
html_dependency_report_dt()
}
\value{
the dependency
}
\description{
HTML Dependency for report headers in \code{DT::datatable}
}
