% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependency_tippy.R
\name{html_dependency_tippy}
\alias{html_dependency_tippy}
\title{HTML Dependency for \code{tippy}}
\usage{
html_dependency_tippy()
}
\value{
the dependency
}
\description{
HTML Dependency for \code{tippy}
}
