% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependency_vert_dt.R
\name{html_dependency_vert_dt}
\alias{html_dependency_vert_dt}
\title{HTML Dependency for vertical headers in \code{DT::datatable}}
\usage{
html_dependency_vert_dt()
}
\value{
the dependency
}
\description{
HTML Dependency for vertical headers in \code{DT::datatable}
}
