% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_duplicate_content.R
\name{int_duplicate_content}
\alias{int_duplicate_content}
\title{Check for duplicated content}
\usage{
int_duplicate_content(level = c("dataframe", "segment"), ...)
}
\arguments{
\item{level}{\link{character} a character vector indicating whether the assessment should be conducted at the study level (level = "dataframe") or at the segment level (level = "segment").}

\item{...}{Depending on \code{level}, passed to either
\link{util_int_duplicate_content_segment} or
\link{util_int_duplicate_content_dataframe}}
}
\value{
a \link{list}. Depending on \code{level}, see
\link{util_int_duplicate_content_segment} or
\link{util_int_duplicate_content_dataframe} for a description of the outputs.
}
\description{
This function tests for duplicates entries in the data set. It is possible to
check duplicated entries by study segments or to consider only selected
segments.
}
