% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_exception_classes.R
\docType{data}
\name{LONG_LABEL_EXCEPTION}
\alias{LONG_LABEL_EXCEPTION}
\title{An exception class assigned for exceptions caused by long
variable labels}
\format{
An object of class \code{character} of length 1.
}
\usage{
LONG_LABEL_EXCEPTION
}
\description{
An exception class assigned for exceptions caused by long
variable labels
}
\keyword{internal}
