% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_env.R
\name{menu_env-menu}
\alias{menu_env-menu}
\title{Generate the menu for a report}
\arguments{
\item{pages}{encapsulated \code{list} with report pages as \code{tagList} objects,
its names are the desired file names}
}
\value{
the html-\code{taglist} for the menu
}
\description{
Generate the menu for a report
}
\keyword{internal}
