% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_env.R
\docType{data}
\name{menu_env}
\alias{menu_env}
\alias{.menu_env}
\title{\code{.menu_env} -- an environment for HTML menu creation}
\format{
An object of class \code{environment} of length 3.
}
\usage{
.menu_env
}
\description{
used by the dq_report2-pipeline
}
\keyword{datasets}
