% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_env.R
\name{menu_env_drop_down}
\alias{menu_env_drop_down}
\title{Creates a drop-down menu}
\arguments{
\item{title}{name of the entry in the main menu}

\item{menu_description}{description, displayed, if the main
menu entry itself is clicked}

\item{...}{the sub-menu-entries}

\item{id}{id for the entry, defaults to modified title}
}
\value{
html div object
}
\description{
Creates a drop-down menu
}
