% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_env.R
\name{menu_env_menu_entry}
\alias{menu_env_menu_entry}
\title{Create a single menu entry}
\arguments{
\item{title}{of the entry}

\item{id}{linked \code{href}, defaults to modified title. can be a word, then
a single-page-link with an anchor tag is created.}

\item{...}{additional arguments for the menu link}
}
\value{
html-a-tag object
}
\description{
Create a single menu entry
}
