% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{meta_data}
\alias{meta_data}
\title{Data frame with metadata about the study data on variable level}
\description{
Variable level metadata.
}
\seealso{
\href{https://dataquality.qihs.uni-greifswald.de/VIN_Annotation_of_Metadata.html}{further details on variable level metadata.}

\link{meta_data_segment}

\link{meta_data_dataframe}
}
