% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{meta_data_dataframe}
\alias{meta_data_dataframe}
\title{Well known columns on the \code{meta_data_dataframe} sheet}
\description{
Meta data describing data delivered on one data frame/table sheet,
e.g., a full questionnaire, not its items.
}
