% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_env.R
\docType{data}
\name{meta_data_env}
\alias{meta_data_env}
\alias{.meta_data_env}
\title{\code{.meta_data_env} -- an environment for easy metadata access}
\format{
An object of class \code{environment} of length 6.
}
\usage{
.meta_data_env
}
\description{
used by the dq_report2-pipeline
}
\seealso{
\link{meta_data_env_id_vars} \link{meta_data_env_co_vars}
\link{meta_data_env_time_vars} \link{meta_data_env_group_vars}
}
\keyword{datasets}
