% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_env.R
\name{meta_data_env_co_vars}
\alias{meta_data_env_co_vars}
\title{Extract co-variables for a given item}
\arguments{
\item{entity}{vector of item-identifiers}
}
\value{
a vector with co-variables for each entity-entry, having the
\code{explode} attribute set to \code{FALSE}
}
\description{
Extract co-variables for a given item
}
\seealso{
\link{meta_data_env}
}
\keyword{internal}
