% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_env.R
\name{meta_data_env_group_vars}
\alias{meta_data_env_group_vars}
\title{Extract group variables for a given item}
\arguments{
\item{entity}{vector of item-identifiers}
}
\value{
a vector with possible group-variables (can be more than  one per
item) for each entity-entry, having the \code{explode} attribute
set to \code{TRUE}
}
\description{
Extract group variables for a given item
}
\seealso{
\link{meta_data_env}
}
\keyword{internal}
