% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_env.R
\name{meta_data_env_time_vars}
\alias{meta_data_env_time_vars}
\title{Extract measurement time variable for a given item}
\arguments{
\item{entity}{vector of item-identifiers}
}
\value{
a vector with time-variables (usually one per item) for each
entity-entry, having the \code{explode} attribute set to \code{TRUE}
}
\description{
Extract measurement time variable for a given item
}
\seealso{
\link{meta_data_env}
}
\keyword{internal}
