% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{meta_data_segment}
\alias{meta_data_segment}
\title{Well known columns on the \code{meta_data_segment} sheet}
\description{
Metadata describing study segments, e.g., a full questionnaire, not its
items.
}
