% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{MISSING_CODE_RULES}
\alias{MISSING_CODE_RULES}
\title{Name of the sheet with rules to introduce missing codes in the pipeline}
\usage{
MISSING_CODE_RULES
}
\description{
Name of the sheet with rules to introduce missing codes in the pipeline
}
