% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{MULTIVARIATE_OUTLIER_CHECK}
\alias{MULTIVARIATE_OUTLIER_CHECK}
\title{Cross-item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
MULTIVARIATE_OUTLIER_CHECK
}
\description{
Select, whether to compute \link{acc_multivariate_outlier}.
}
\details{
You can leave the cell empty, then the depends on the setting of the
\code{option} \link{dataquieR.MULTIVARIATE_OUTLIER_CHECK}. If this column is missing,
all this is the same as having all cells empty and
\code{dataquieR.MULTIVARIATE_OUTLIER_CHECK} set to \code{"auto"}.

See also \code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}}.
}
\seealso{
\link{meta_data_cross}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_METRIC}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{COMPUTED_VARIABLE_ROLES}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{CONTRADICTION_TYPE}},
\code{\link{DATA_PREPARATION}},
\code{\link{GOLDSTANDARD}},
\code{\link{IRV}},
\code{\link{MAHALANOBIS_THRESHOLD}},
\code{\link{MAXIMUM_LONG_STRING}},
\code{\link{MISS_RESP}},
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}},
\code{\link{RELCOMPL_SPEED}},
\code{\link{REL_VAL}},
\code{\link{RESPT_PER_ITEM}},
\code{\link{SCALE_ACRONYM}},
\code{\link{SCALE_NAME}},
\code{\link{TOTRESPT}},
\code{\link{VARIABLE_LIST}},
\code{\link{VARIABLE_LIST_ORDER}},
\code{\link{meta_data_computation}},
\code{\link{meta_data_cross}}
}
\concept{meta_data_cross}
\keyword{datasets}
