% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{MULTIVARIATE_OUTLIER_CHECKTYPE}
\alias{MULTIVARIATE_OUTLIER_CHECKTYPE}
\title{Cross-item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
MULTIVARIATE_OUTLIER_CHECKTYPE
}
\description{
Select, which outlier criteria to compute, see \link{acc_multivariate_outlier}.
}
\details{
You can leave the cell empty, then, all checks will apply. If you enter
a set of methods, the maximum for \link{N_RULES} changes. See also
\code{\link{UNIVARIATE_OUTLIER_CHECKTYPE}}.
}
\seealso{
\link{meta_data_cross}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_METRIC}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{CONTRADICTION_TYPE}},
\code{\link{DATA_PREPARATION}},
\code{\link{GOLDSTANDARD}},
\code{\link{N_RULES}},
\code{\link{REL_VAL}},
\code{\link{VARIABLE_LIST}},
\code{\link{util_normalize_cross_item}()}
}
\concept{meta_data_cross}
\keyword{datasets}
