% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{N_RULES}
\alias{N_RULES}
\title{Cross-item and item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
N_RULES
}
\description{
Select, how many violated outlier criteria make an observation an outlier,
see \link{acc_multivariate_outlier}.
}
\details{
You can leave the cell empty, then, all applied checks must deem an
observation an outlier to have it flagged. See
\link{UNIVARIATE_OUTLIER_CHECKTYPE} and
\link{MULTIVARIATE_OUTLIER_CHECKTYPE} for the selected outlier criteria.
}
\seealso{
\link{meta_data_cross}

\link{meta_data}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_METRIC}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{CONTRADICTION_TYPE}},
\code{\link{DATA_PREPARATION}},
\code{\link{GOLDSTANDARD}},
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}},
\code{\link{REL_VAL}},
\code{\link{VARIABLE_LIST}},
\code{\link{util_normalize_cross_item}()}
}
\concept{meta_data_cross}
\keyword{datasets}
