% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_translate.R
\name{names<-.dataquieR_translated}
\alias{names<-.dataquieR_translated}
\title{\code{names} implementation for the class \code{dataquieR_translated}}
\usage{
\method{names}{dataquieR_translated}(x) <- value
}
\arguments{
\item{x}{\code{dataquieR_translated} object}

\item{value}{the names to assign}
}
\value{
names of the underlying character vector
}
\description{
\code{dataquieR}'s translated texts featuring access to the language keys, still.
this function returns the language keys.
}
\details{
only \code{setNames(nm = x)} is allowed for convenience. Any other assignment
would mean to change the language keys, so this is not allowed.
}
\seealso{
base::as.character
}
