% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nres.R
\name{nres}
\alias{nres}
\title{return the number of result slots in a report}
\usage{
nres(x)
}
\arguments{
\item{x}{the \code{dataquieR} report (v2.0)}
}
\value{
the number of used result slots
}
\description{
return the number of result slots in a report
}
