% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_recursive_result.R
\name{pipeline_recursive_result}
\alias{pipeline_recursive_result}
\title{Convert a pipeline result data frame to named encapsulated lists}
\usage{
pipeline_recursive_result(...)
}
\arguments{
\item{...}{Deprecated}
}
\value{
Deprecated
}
\description{
Deprecated
}
