% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_vectorized.R
\name{pipeline_vectorized}
\alias{pipeline_vectorized}
\title{Call (nearly) one "Accuracy" function with many parameterizations at once
automatically}
\usage{
pipeline_vectorized(...)
}
\arguments{
\item{...}{Deprecated}
}
\value{
Deprecated
}
\description{
Deprecated
}
