% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dataquieR_summary.R
\name{plot.dataquieR_summary}
\alias{plot.dataquieR_summary}
\title{Plot a \code{dataquieR} summary}
\usage{
\method{plot}{dataquieR_summary}(
  x,
  y,
  ...,
  filter,
  dont_plot = FALSE,
  stratify_by,
  vars_to_include = "study",
  disable_plotly = FALSE,
  hierarchy,
  folder_of_report = NULL,
  var_uniquenames = NULL
)
}
\arguments{
\item{x}{the \code{dataquieR} summary, see \code{\link[=summary]{summary()}} and \code{\link[=dq_report2]{dq_report2()}}}

\item{y}{not yet used}

\item{...}{not yet used}

\item{filter}{if given, this filters the summary, e.g.,
\code{filter = call_names == "com_qualified_item_missingness"}}

\item{dont_plot}{suppress the actual plotting, just return a printable
object derived from \code{x}}

\item{stratify_by}{column to stratify the summary, may be one string.}

\item{vars_to_include}{\link{character} study | \code{ssi}. variables are study
variables or computed social science
indicator variables.}

\item{disable_plotly}{\link{logical} do not use \code{plotly}, even if installed}

\item{hierarchy}{not yet defined, but if an argument is given, a
sunburst chart is displayed, currently, only \code{DQ_OBS}
can be used a the hierarchy.}

\item{folder_of_report}{a named vector with the location of variable and
\code{call_names}}

\item{var_uniquenames}{a data frame with the original variable names and the
unique names in case of reports created with dq_report_by
containing the same variable in several reports
(e.g., creation of reports by sex)}
}
\value{
invisible html object
}
\description{
Plot a \code{dataquieR} summary
}
