% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_apply_coding.R
\name{prep_apply_coding}
\alias{prep_apply_coding}
\title{Re-Code labels with their respective codes according to the \code{meta_data}}
\usage{
prep_apply_coding(study_data, meta_data = "item_level")
}
\arguments{
\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data}
}
\value{
\link{data.frame} modified study data with labels replaced by the codes
}
\description{
Re-Code labels with their respective codes according to the \code{meta_data}
}
