% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_check_for_dataquieR_updates.R
\name{prep_check_for_dataquieR_updates}
\alias{prep_check_for_dataquieR_updates}
\title{Check for package updates}
\usage{
prep_check_for_dataquieR_updates(beta = FALSE, deps = TRUE)
}
\arguments{
\item{beta}{\link{logical} check for beta version too}

\item{deps}{\link{logical} check for missing (optional) dependencies}
}
\value{
\code{invisible(NULL)}
}
\description{
Check for package updates
}
