% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_create_meta_data_file.R
\name{prep_create_meta_data_file}
\alias{prep_create_meta_data_file}
\title{Instantiate a new metadata file}
\usage{
prep_create_meta_data_file(
  file_name,
  study_data,
  open = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{file_name}{\link{character} file path to write to}

\item{study_data}{\link{data.frame} optional, study data to guess metadata from}

\item{open}{\link{logical} open the file after creation}

\item{overwrite}{\link{logical} overwrite \code{file}, if exists}
}
\value{
\code{invisible(NULL)}
}
\description{
Instantiate a new metadata file
}
