% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_storr_factory.R
\name{prep_create_storr_factory}
\alias{prep_create_storr_factory}
\title{Create a factory function for \code{storr} objects for backing
a \link{dataquieR_resultset2}}
\usage{
prep_create_storr_factory(db_dir = tempfile(), namespace = "objects")
}
\arguments{
\item{db_dir}{\link{character} path to the directory for the back-end, if
one is created on the fly.}

\item{namespace}{\link{character} namespace for the report, so that one back-end
can back several reports

the returned function will try to create a \code{storr} object using a temporary
folder or the folder in \code{db_dir}, if specified. The database will either
be the \code{storr_rds}.}
}
\value{
\code{storr} object or \code{NULL}, if package \code{storr} is not available
}
\description{
Create a factory function for \code{storr} objects for backing
a \link{dataquieR_resultset2}
}
