% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_deparse_assignments.R
\name{prep_deparse_assignments}
\alias{prep_deparse_assignments}
\title{Convert two vectors from a code-value-table to a key-value list}
\usage{
prep_deparse_assignments(codes, labels, split_char = SPLIT_CHAR)
}
\arguments{
\item{codes}{\link{integer} codes}

\item{labels}{\link{character} labels, same length as codes}

\item{split_char}{\link{character} split character character to split
code assignments}
}
\value{
a vector with assignment strings for each row of
\code{cbind(codes, labels)}
}
\description{
Convert two vectors from a code-value-table to a key-value list
}
