% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_setup_rstudio_job.R
\name{prep_deregister_progress_hook}
\alias{prep_deregister_progress_hook}
\title{De-register a hook function for progresses in computation/rendering}
\usage{
prep_deregister_progress_hook(handle, verbose = TRUE)
}
\arguments{
\item{handle}{\link{character} the handle}

\item{verbose}{\link{logical} message, if \code{handle} has currently no registration}
}
\value{
\link{logical} \code{invisible(TRUE)} on success
}
\description{
De-register a hook function for progresses in computation/rendering
}
