% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_extract_classes_by_functions.R
\name{prep_extract_classes_by_functions}
\alias{prep_extract_classes_by_functions}
\title{Extract old function based summary from data quality results}
\usage{
prep_extract_classes_by_functions(r)
}
\arguments{
\item{r}{\link{dq_report2}}
}
\value{
\link{data.frame} long format, compatible with \code{\link[=prep_summary_to_classes]{prep_summary_to_classes()}}
}
\description{
Extract old function based summary from data quality results
}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()}
}
\concept{summary_functions}
