% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_extract_summary.dataquieR_resultset2.R
\name{prep_extract_summary.dataquieR_resultset2}
\alias{prep_extract_summary.dataquieR_resultset2}
\title{Extract report summary from reports}
\usage{
\method{prep_extract_summary}{dataquieR_resultset2}(r, ...)
}
\arguments{
\item{r}{\link{dq_report2} a \link{dq_report2} report}

\item{...}{not used}
}
\value{
\link{list} with two slots \code{Data} and \code{Table} with \link{data.frame}s
featuring all metrics columns
from the report \code{r}, the \link{STUDY_SEGMENT} and the \link{VAR_NAMES}.
In case of \code{Data}, the columns are formatted nicely but still
with the standardized column names -- use
\code{util_translate_indicator_metrics()} to rename them nicely. In
case of \code{Table}, just as they are.
}
\description{
Extract report summary from reports
}
\seealso{
\code{\link[=prep_combine_report_summaries]{prep_combine_report_summaries()}}

Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()}
}
\concept{summary_functions}
