% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_extract_summary.R
\name{prep_extract_summary}
\alias{prep_extract_summary}
\title{Extract summary from data quality results}
\usage{
prep_extract_summary(r, ...)
}
\arguments{
\item{r}{\link{dq_report2} or \link{dataquieR_result} object}

\item{...}{further arguments, maybe needed for some implementations}
}
\value{
\link{list} with two slots \code{Data} and \code{Table} with \link{data.frame}s
featuring all metrics columns
from the report or result in \code{x},
the \link{STUDY_SEGMENT} and the \link{VAR_NAMES}.
In case of \code{Data}, the columns are formatted nicely but still
with the standardized column names -- use
\code{util_translate_indicator_metrics()} to rename them nicely. In
case of \code{Table}, just as they are.
}
\description{
Generic function, currently supports \link{dq_report2} and \link{dataquieR_result}
}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()}
}
\concept{summary_functions}
