% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_get_study_data_segment.R
\name{prep_get_study_data_segment}
\alias{prep_get_study_data_segment}
\title{Get data frame for a given segment}
\usage{
prep_get_study_data_segment(
  segment,
  study_data,
  item_level = "item_level",
  meta_data = item_level,
  meta_data_v2,
  segment_level,
  meta_data_segment = "segment_level"
)
}
\arguments{
\item{segment}{\link{character} name of the segment to return data for}

\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{item_level}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{meta_data}{\link{data.frame} old name for \code{item_level}}

\item{meta_data_v2}{\link{character} path to workbook like metadata file, see
\code{\link{prep_load_workbook_like_file}} for details.
\strong{ALL LOADED DATAFRAMES WILL BE PURGED},
using \code{\link{prep_purge_data_frame_cache}},
if you specify \code{meta_data_v2}.}

\item{segment_level}{\link{data.frame} alias for \code{meta_data_segment}}

\item{meta_data_segment}{\link{data.frame} -- optional: Segment level metadata}
}
\value{
\link{data.frame} the data for the segment
}
\description{
Get data frame for a given segment
}
