% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_get_user_name.R
\name{prep_get_user_name}
\alias{prep_get_user_name}
\title{Return the logged-in User's Full Name}
\usage{
prep_get_user_name()
}
\value{
\link{character} the user's name
}
\description{
If \code{whoami} is not installed, the user name from
\code{Sys.info()} is returned.
}
\details{
Can be overridden by options or environment:

\code{options(FULLNAME = "Stephan Struckmann")}

\code{Sys.setenv(FULLNAME = "Stephan Struckmann")}
}
