% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_get_variant.R
\name{prep_get_variant}
\alias{prep_get_variant}
\title{Get machine variant for snapshot tests}
\usage{
prep_get_variant()
}
\value{
\link{character} the variant
}
\description{
Get machine variant for snapshot tests
}
\keyword{internal}
