% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_robust_guess_data_type.R
\name{prep_guess_encoding}
\alias{prep_guess_encoding}
\title{Guess encoding of text or text files}
\usage{
prep_guess_encoding(x, file)
}
\arguments{
\item{x}{\link{character} string to guess encoding for}

\item{file}{\link{character} file to guess encoding for}
}
\value{
encoding
}
\description{
Guess encoding of text or text files
}
