% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_translate.R
\name{prep_is_translated}
\alias{prep_is_translated}
\title{Detect if an object is a \code{dataquieR_translated} object}
\usage{
prep_is_translated(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
\code{TRUE}, if \code{x} is a \code{dataquieR_translated} object.
}
\description{
Detect if an object is a \code{dataquieR_translated} object
}
