% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_list_dataframes.R
\name{prep_list_dataframes}
\alias{prep_list_dataframes}
\title{List Loaded Data Frames}
\usage{
prep_list_dataframes()
}
\value{
names of all loaded data frames
}
\description{
List Loaded Data Frames
}
\seealso{
Other data-frame-cache: 
\code{\link{prep_add_data_frames}()},
\code{\link{prep_get_data_frame}()},
\code{\link{prep_load_folder_with_metadata}()},
\code{\link{prep_load_workbook_like_file}()},
\code{\link{prep_purge_data_frame_cache}()}
}
\concept{data-frame-cache}
