% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_save_report.R
\name{prep_load_report}
\alias{prep_load_report}
\title{Load a \code{dq_report2}}
\usage{
prep_load_report(file)
}
\arguments{
\item{file}{\link{character} the file name to load from}
}
\value{
\link{dataquieR_resultset2} the report
}
\description{
Load a \code{dq_report2}
}
