% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_merge_study_data.R
\name{prep_merge_study_data}
\alias{prep_merge_study_data}
\title{Merge a list of study data frames to one (sparse) study data frame}
\usage{
prep_merge_study_data(study_data_list)
}
\arguments{
\item{study_data_list}{\link{list} the list}
}
\value{
\link{data.frame} \link{study_data}
}
\description{
Merge a list of study data frames to one (sparse) study data frame
}
