% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lazy_ggplot.R
\name{prep_realize_ggplot}
\alias{prep_realize_ggplot}
\title{Materialize a lazy \code{ggplot}}
\usage{
prep_realize_ggplot(x)
}
\arguments{
\item{x}{a \code{dq_lazy_ggplot} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Evaluate the stored expression in its lean environment and cache
the resulting \code{ggplot} object in the current R session, if enabled
using the option \link{dataquieR.lazy_plots_cache}.
}
