% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_setup_rstudio_job.R
\name{prep_register_progress_hook}
\alias{prep_register_progress_hook}
\title{Register a hook function for progresses in computation/rendering}
\usage{
prep_register_progress_hook(type = c("progress", "init", "msg"), hook)
}
\arguments{
\item{type}{\link{character} what event}

\item{hook}{\link{function} hook function}
}
\value{
\link{character} a handle for de-registering, \code{invisible}
}
\description{
The order hooks are called is not defined.
}
