% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/prep_render_pie_chart_from_summaryclasses_plotly.R
\name{prep_render_pie_chart_from_summaryclasses_plotly}
\alias{prep_render_pie_chart_from_summaryclasses_plotly}
\title{Create a \code{plotly} pie chart}
\usage{
prep_render_pie_chart_from_summaryclasses_plotly(
  data,
  meta_data = "item_level"
)
}
\arguments{
\item{data}{data as returned by \code{prep_summary_to_classes} but
summarized by one column (currently, we support
\code{indicator_metric}, \code{call_names},
\code{STUDY_SEGMENT}, and \code{VAR_NAMES})}

\item{meta_data}{\link{meta_data}}
}
\value{
a \code{htmltools} compatible object
}
\description{
Create a \code{plotly} pie chart
}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_summary_to_classes}()}
}
\concept{summary_functions}
