% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_save_report.R
\name{prep_save_report}
\alias{prep_save_report}
\title{Save a \code{dq_report2}}
\usage{
prep_save_report(report, file, compression_level = 3)
}
\arguments{
\item{report}{\link{dataquieR_resultset2} the report}

\item{file}{\link{character} the file name to write to}

\item{compression_level}{\link{integer}  from=0 to=9. Compression level.
9 is very slow.}
}
\value{
\code{invisible(NULL)}
}
\description{
Save a \code{dq_report2}
}
