% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scalelevel_from_data_and_metadata.R
\name{prep_scalelevel_from_data_and_metadata}
\alias{prep_scalelevel_from_data_and_metadata}
\title{Heuristics to amend a SCALE_LEVEL column and a UNIT column in the metadata}
\usage{
prep_scalelevel_from_data_and_metadata(
  resp_vars = lifecycle::deprecated(),
  study_data,
  item_level = "item_level",
  label_col = LABEL,
  meta_data = item_level,
  meta_data_v2
)
}
\arguments{
\item{resp_vars}{\link{variable list} deprecated, the function always addresses
all variables.}

\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{item_level}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{meta_data}{\link{data.frame} old name for \code{item_level}}

\item{meta_data_v2}{\link{character} path to workbook like metadata file, see
\code{\link{prep_load_workbook_like_file}} for details.
\strong{ALL LOADED DATAFRAMES WILL BE PURGED},
using \code{\link{prep_purge_data_frame_cache}},
if you specify \code{meta_data_v2}.}
}
\value{
\link{data.frame} modified metadata
}
\description{
...if missing
}
\examples{
\dontrun{
  prep_load_workbook_like_file("meta_data_v2")
  prep_scalelevel_from_data_and_metadata(study_data = "study_data")
}
}
