% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{prep_set_backend}
\alias{prep_set_backend}
\title{Change the back-end of a report}
\usage{
prep_set_backend(r, storr_factory = NULL, amend = FALSE)
}
\arguments{
\item{r}{\link{dataquieR_resultset2} the report}

\item{storr_factory}{\code{storr} the \code{storr} storage or \code{NULL}, to move
the report fully back into the RAM.}

\item{amend}{\link{logical} if there is already data in.\code{storr_factory},
use it anyways -- unsupported, so far!}
}
\value{
\link{dataquieR_resultset2} but now with the desired back-end
}
\description{
with this function, you can move a report from/to a \code{storr} storage.
}
