% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_summary_to_classes.R
\name{prep_summary_to_classes}
\alias{prep_summary_to_classes}
\title{Classify metrics from a report summary table}
\usage{
prep_summary_to_classes(report_summary)
}
\arguments{
\item{report_summary}{\code{\link[=list]{list()}} as returned by
\code{\link[=prep_extract_summary]{prep_extract_summary()}}}
}
\value{
\link{data.frame} classes for the report summary table, long format
}
\description{
Classify metrics from a report summary table
}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()}
}
\concept{summary_functions}
