% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_title_escape.R
\name{prep_title_escape}
\alias{prep_title_escape}
\title{Prepare a label as part of a title text for \code{RMD} files}
\usage{
prep_title_escape(s, html = FALSE)
}
\arguments{
\item{s}{the label}

\item{html}{prepare the label for direct \code{HTML} output instead of \code{RMD}}
}
\value{
the escaped label
}
\description{
Prepare a label as part of a title text for \code{RMD} files
}
