% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_undisclose.R
\name{prep_undisclose}
\alias{prep_undisclose}
\title{Remove data disclosing details}
\usage{
prep_undisclose(x)
}
\arguments{
\item{x}{an object to un-disclose, a}
}
\value{
undisclosed object
}
\description{
new function: no warranty, so far.
}
