% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_split_val_tab.R
\name{prep_unsplit_val_tabs}
\alias{prep_unsplit_val_tabs}
\title{Combine all missing and value lists to one big table}
\usage{
prep_unsplit_val_tabs(meta_data = "item_level", val_tab = NULL)
}
\arguments{
\item{meta_data}{\link{data.frame} item level meta data to be used, defaults to
\code{"item_level"}}

\item{val_tab}{\link{character} name of the table being created: This table will
be added to the data frame cache (or overwritten). If \code{NULL}, the table
will only be returned}
}
\value{
\link{data.frame} the combined table
}
\description{
Combine all missing and value lists to one big table
}
