% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_valuelabels_from_data.R
\name{prep_valuelabels_from_data}
\alias{prep_valuelabels_from_data}
\title{Get value labels from data}
\usage{
prep_valuelabels_from_data(resp_vars = colnames(study_data), study_data)
}
\arguments{
\item{resp_vars}{\link{variable} names of the variables to fetch the value labels
from the data}

\item{study_data}{\link{data.frame} the data frame that contains the measurements}
}
\value{
a \link{list} with:
\itemize{
\item \code{VALUE_LABELS}: vector of value labels and modified study data
\item \code{ModifiedStudyData}: study data with factors as integers
}
}
\description{
Detects factors and converts them to compatible metadata/study data.
}
\examples{
\dontrun{
dataquieR::prep_datatype_from_data(iris)
}
}
