% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_resultset.R
\name{print.dataquieR_resultset}
\alias{print.dataquieR_resultset}
\title{Generate a RMarkdown-based report from a \link{dataquieR} report}
\usage{
\method{print}{dataquieR_resultset}(...)
}
\arguments{
\item{...}{deprecated}
}
\value{
deprecated
}
\description{
Generate a RMarkdown-based report from a \link{dataquieR} report
}
