% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_resultset2.R
\name{print.dataquieR_resultset2}
\alias{print.dataquieR_resultset2}
\title{Generate a HTML-based report from a \link{dataquieR} report}
\usage{
\method{print}{dataquieR_resultset2}(
  x,
  dir,
  view = TRUE,
  disable_plotly = FALSE,
  block_load_factor = 4,
  advanced_options = list(),
  dashboard = NA,
  ...
)
}
\arguments{
\item{x}{\link[=dq_report2]{dataquieR report v2}.}

\item{dir}{\link{character} directory to store the rendered report's files,
a temporary one,
if omitted. Directory will be created, if missing, files may be
overwritten inside that directory}

\item{view}{\link{logical} display the report}

\item{disable_plotly}{\link{logical} do not use \code{plotly}, even if installed}

\item{block_load_factor}{\link{numeric} multiply size of parallel compute blocks
by this factor.}

\item{advanced_options}{\link{list} options to set during report computation,
see \code{\link[=options]{options()}}}

\item{dashboard}{\link{logical} dashboard mode: \code{TRUE}: create a dashboard only,
\code{FALSE}: don't create a dashboard at all,
\code{NA} or missing: create a "normal" report with
a dashboard included.}

\item{...}{additional arguments:}
}
\value{
file names of the generated report's HTML files
}
\description{
Generate a HTML-based report from a \link{dataquieR} report
}
