% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_translate.R
\name{print.dataquieR_translated}
\alias{print.dataquieR_translated}
\title{\code{print} implementation for the class \code{dataquieR_translated}}
\usage{
\method{print}{dataquieR_translated}(x, ...)
}
\arguments{
\item{x}{\code{dataquieR_translated} object to print}

\item{...}{passed to base::print}
}
\value{
as print
}
\description{
\code{dataquieR}'s translated texts featuring access to the language keys, still.
}
\seealso{
base::print
}
