% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.interval.R
\name{print.interval}
\alias{print.interval}
\title{print implementation for the class \code{interval}}
\usage{
\method{print}{interval}(x, ...)
}
\arguments{
\item{x}{\code{interval} objects to print}

\item{...}{not used yet}
}
\value{
the printed object
}
\description{
such objects, for now, only occur in \code{RECCap} rules, so this function
is meant for internal use, mostly -- for now.
}
\seealso{
base::print
}
