% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_resultset2.R
\name{print.list}
\alias{print.list}
\title{print a list of \code{dataquieR_result} objects}
\usage{
\method{print}{list}(x, ...)
}
\arguments{
\item{x}{\code{\link[=list]{list()}} only, if all elements inherit from \link{dataquieR_result},
this implementation runs}

\item{...}{passed to other implementations}
}
\value{
undefined
}
\description{
print a list of \code{dataquieR_result} objects
}
