% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_as_numeric_with_unit.R
\name{print.numeric_with_unit}
\alias{print.numeric_with_unit}
\title{Print a number with unit}
\usage{
\method{print}{numeric_with_unit}(x, ...)
}
\arguments{
\item{x}{number with unit}

\item{...}{not used}
}
\value{
\code{invisible(x)}
}
\description{
Print a number with unit
}
