% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_result.R
\name{print.Slot}
\alias{print.Slot}
\title{Print a \code{Slot} object}
\usage{
\method{print}{Slot}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{not used}
}
\value{
calls the next print method
}
\description{
displays all warnings and stuff. then it prints \code{x}.
}
