% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_result.R
\name{print.StudyDataSlot}
\alias{print.StudyDataSlot}
\title{Print a \code{StudyDataSlot} object}
\usage{
\method{print}{StudyDataSlot}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{not used}
}
\value{
see print
}
\description{
Print a \code{StudyDataSlot} object
}
