% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pairs_matrix.R
\name{print.util_pairs_ggplot_panels}
\alias{print.util_pairs_ggplot_panels}
\title{Print method for \code{util_pairs_ggplot_panels} objects}
\usage{
\method{print}{util_pairs_ggplot_panels}(x, ...)
}
\arguments{
\item{x}{An object of class \code{util_pairs_ggplot_panels}.}

\item{...}{Ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print method for \code{util_pairs_ggplot_panels} objects
}
