% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind.ReportSummaryTable}
\alias{rbind.ReportSummaryTable}
\title{Combine \code{ReportSummaryTable} outputs}
\usage{
\method{rbind}{ReportSummaryTable}(...)
}
\arguments{
\item{...}{\code{ReportSummaryTable} objects to combine.}
}
\description{
Using this \code{rbind} implementation, you can combine different
heatmap-like results of the class \code{ReportSummaryTable}.
}
\seealso{
\link[base:cbind]{base::rbind.data.frame}
}
