% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resnames.dataquieR_resultset2.R
\name{resnames.dataquieR_resultset2}
\alias{resnames.dataquieR_resultset2}
\title{Return names of result slots (e.g., 3rd dimension of dataquieR results)}
\usage{
\method{resnames}{dataquieR_resultset2}(x)
}
\arguments{
\item{x}{the objects}
}
\value{
character vector with names
}
\description{
Return names of result slots (e.g., 3rd dimension of dataquieR results)
}
