% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resnames.R
\name{resnames}
\alias{resnames}
\title{Return names of result slots (e.g., 3rd dimension of dataquieR results)}
\usage{
resnames(x)
}
\arguments{
\item{x}{the objects}
}
\value{
character vector with names
}
\description{
Return names of result slots (e.g., 3rd dimension of dataquieR results)
}
