% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{SEGMENT_ID_TABLE}
\alias{SEGMENT_ID_TABLE}
\title{Deprecated segment level metadata attribute name}
\usage{
SEGMENT_ID_TABLE
}
\description{
The name of the data frame containing the reference IDs to be compared
with the IDs in the targeted segment.
}
\details{
Please use \link{SEGMENT_ID_REF_TABLE}
}
