% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{SEGMENT_ID_VARS}
\alias{SEGMENT_ID_VARS}
\title{Segment level metadata attribute name}
\usage{
SEGMENT_ID_VARS
}
\description{
All variables that are to be used as one single ID
variable (combined key) in a segment.
}
\seealso{
\link{meta_data_segment}
}
