% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{SEGMENT_PART_VARS}
\alias{SEGMENT_PART_VARS}
\title{Segment level metadata attribute name}
\usage{
SEGMENT_PART_VARS
}
\description{
The name of the segment participation status variable
}
\seealso{
\link{meta_data_segment}
}
