% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{SEGMENT_RECORD_CHECK}
\alias{SEGMENT_RECORD_CHECK}
\title{Segment level metadata attribute name}
\usage{
SEGMENT_RECORD_CHECK
}
\description{
The type of check to be conducted when comparing the reference ID
table with the IDs in a segment.
}
\seealso{
\link{meta_data_segment}
}
