% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{SEGMENT_RECORD_COUNT}
\alias{SEGMENT_RECORD_COUNT}
\title{Segment level metadata attribute name}
\usage{
SEGMENT_RECORD_COUNT
}
\description{
Number of expected data records in each segment.
\link{numeric}. Check only conducted if number entered
}
\seealso{
\link{meta_data_segment}
}
