% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{SEGMENT_UNIQUE_ID}
\alias{SEGMENT_UNIQUE_ID}
\title{Segment level metadata attribute name}
\usage{
SEGMENT_UNIQUE_ID
}
\description{
Segment level metadata attribute name
}
\seealso{
\link{DF_UNIQUE_ID}

\link{meta_data_segment}
}
