% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{SPLIT_CHAR}
\alias{SPLIT_CHAR}
\title{Character used  by default as a separator in metadata such as
missing codes}
\usage{
SPLIT_CHAR
}
\description{
This 1 character is according to our metadata concept "|".
}
