% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{study_data}
\alias{study_data}
\title{Data frame with the study data whose quality is being assessed}
\description{
Study data is expected in wide format. If should contain all variables
for all segments in one large table, even, if some variables are not
measured for all observational utils (study participants).
}
