% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_dataquieR_resultset2.R
\name{[.dataquieR_resultset2}
\alias{[.dataquieR_resultset2}
\title{Get a subset of a \code{dataquieR} \code{dq_report2} report}
\usage{
\method{[}{dataquieR_resultset2}(x, row, col, res, drop = FALSE)
}
\arguments{
\item{x}{the report}

\item{row}{the variable names, must be unique}

\item{col}{the function-call-names, must be unique}

\item{res}{the result slot, must be unique}

\item{drop}{drop, if length is 1}
}
\value{
a list with results, depending on \code{drop} and the number of results,
the list may contain all requested results in sub-lists. The order
of the results follows the order of the row/column/result-names given
}
\description{
Get a subset of a \code{dataquieR} \code{dq_report2} report
}
