% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{[[<-.dataquieR_resultset2}
\alias{[[<-.dataquieR_resultset2}
\title{Set a single result from a \verb{dataquieR 2} report}
\usage{
\method{[[}{dataquieR_resultset2}(x, el) <- value
}
\arguments{
\item{x}{the report}

\item{el}{the index}

\item{value}{the single result}
}
\value{
the \code{dataquieR} result object
}
\description{
Set a single result from a \verb{dataquieR 2} report
}
