% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{[<-.dataquieR_resultset2}
\alias{[<-.dataquieR_resultset2}
\title{Write to a report}
\usage{
\method{[}{dataquieR_resultset2}(x, ...) <- value
}
\arguments{
\item{x}{a `dataquieR_resultset2}

\item{...}{if this contains only one entry and this entry is not named
or its name is \code{els}, then, the report will be accessed in
list mode.}

\item{value}{new value to write}
}
\value{
nothing, stops
}
\description{
Overwriting of elements only list-wise supported
}
