% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dataquieR_resultset.R
\name{summary.dataquieR_resultset}
\alias{summary.dataquieR_resultset}
\title{Summarize a \link{dataquieR} report}
\usage{
\method{summary}{dataquieR_resultset}(...)
}
\arguments{
\item{...}{Deprecated}
}
\value{
Deprecated
}
\description{
Deprecated
}
