% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dataquieR_resultset2.R
\name{summary.dataquieR_resultset2}
\alias{summary.dataquieR_resultset2}
\title{Generate a report summary table}
\usage{
\method{summary}{dataquieR_resultset2}(
  object,
  aspect = c("applicability", "error", "issue"),
  FUN = util_get_html_cell_for_result,
  collapse = "\\n<br />\\n",
  ...
)
}
\arguments{
\item{object}{a square result set}

\item{aspect}{an aspect/problem category of results}

\item{FUN}{function to apply to the cells of the result table}

\item{collapse}{passed to \code{FUN}}

\item{...}{not used}
}
\value{
a summary of a \code{Square2} report
}
\description{
Generate a report summary table
}
\examples{
\dontrun{
  util_html_table(summary(report, aspect = "error", FUN = util_get_html_cell_for_result),
       filter = "top", options = list(scrollCollapse = TRUE, scrollY = "75vh"),
       is_matrix_table = TRUE, rotate_headers = TRUE, output_format = "HTML"
  )
}
}
