% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_geom_pointrange_robust.R
\name{to_basic.GeomPointrangeRobust}
\alias{to_basic.GeomPointrangeRobust}
\title{Internally used point-range}
\usage{
to_basic.GeomPointrangeRobust(data, prestats_data, layout, params, p, ...)
}
\arguments{
\item{data}{the data returned by \code{ggplot2::ggplot_build()}}

\item{prestats_data}{the data before statistics are computed.}

\item{layout}{the panel layout.}

\item{params}{parameters for the geom, statistic, and 'constant' aesthetics}

\item{p}{a ggplot2 object (the conversion may depend on scales, for instance).}

\item{...}{currently ignored}
}
\description{
Internally used point-range
}
