% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{UNIT_IS_COUNT}
\alias{UNIT_IS_COUNT}
\title{Is a unit a count according to \code{\link[units:valid_udunits]{units::valid_udunits()}}}
\description{
see column \code{def}, therein
}
\details{
like \verb{\%}, \code{ppt}, \code{ppm}
}
\seealso{
Other UNITS: 
\code{\link{UNITS}},
\code{\link{UNIT_PREFIXES}},
\code{\link{UNIT_PREFIX_FACTORS}},
\code{\link{UNIT_SOURCES}},
\code{\link{WELL_KNOWN_META_VARIABLE_NAMES}}
}
\concept{UNITS}
