% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{UNIT_PREFIX_FACTORS}
\alias{UNIT_PREFIX_FACTORS}
\title{Factors related to unit prefixes \code{\link[units:valid_udunits]{units::valid_udunits_prefixes()}}}
\description{
named \code{numeric} vector
}
\details{
translates k, m, M, c, ... to 1000, 0.001, ...
}
\seealso{
Other UNITS: 
\code{\link{UNITS}},
\code{\link{UNIT_IS_COUNT}},
\code{\link{UNIT_PREFIXES}},
\code{\link{UNIT_SOURCES}},
\code{\link{WELL_KNOWN_META_VARIABLE_NAMES}}
}
\concept{UNITS}
