% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{UNIT_PREFIXES}
\alias{UNIT_PREFIXES}
\title{Valid unit prefixes according to \code{\link[units:valid_udunits]{units::valid_udunits_prefixes()}}}
\description{
like k, m, M, c, ...
}
\seealso{
Other UNITS: 
\code{\link{UNITS}},
\code{\link{UNIT_IS_COUNT}},
\code{\link{UNIT_PREFIX_FACTORS}},
\code{\link{UNIT_SOURCES}},
\code{\link{WELL_KNOWN_META_VARIABLE_NAMES}}
}
\concept{UNITS}
