% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{UNIT_SOURCES}
\alias{UNIT_SOURCES}
\title{Maturity stage of a unit according to \code{\link[units:valid_udunits]{units::valid_udunits()}}}
\description{
see column \code{source_xml} therein, i.e., base, derived, accepted, or common
}
\seealso{
Other UNITS: 
\code{\link{UNITS}},
\code{\link{UNIT_IS_COUNT}},
\code{\link{UNIT_PREFIXES}},
\code{\link{UNIT_PREFIX_FACTORS}},
\code{\link{WELL_KNOWN_META_VARIABLE_NAMES}}
}
\concept{UNITS}
