% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{UNITS}
\alias{UNITS}
\title{Valid unit symbols according to \code{\link[units:valid_udunits]{units::valid_udunits()}}}
\description{
like m, g, N, ...
}
\seealso{
Other UNITS: 
\code{\link{UNIT_IS_COUNT}},
\code{\link{UNIT_PREFIXES}},
\code{\link{UNIT_PREFIX_FACTORS}},
\code{\link{UNIT_SOURCES}},
\code{\link{WELL_KNOWN_META_VARIABLE_NAMES}}
}
\concept{UNITS}
