% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{UNIVARIATE_OUTLIER_CHECKTYPE}
\alias{UNIVARIATE_OUTLIER_CHECKTYPE}
\title{Item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
UNIVARIATE_OUTLIER_CHECKTYPE
}
\description{
Select, which outlier criteria to compute, see \link{acc_univariate_outlier}.
}
\details{
You can leave the cell empty, then, all checks will apply. If you enter
a set of methods, the maximum for \link{N_RULES} changes. See also
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}}.
}
\seealso{
\link{WELL_KNOWN_META_VARIABLE_NAMES}
}
\keyword{datasets}
