% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_3SD.R
\name{util_3SD}
\alias{util_3SD}
\title{Utility function for 3SD deviations rule}
\usage{
util_3SD(x)
}
\arguments{
\item{x}{\link{numeric} data to check for outliers}
}
\value{
binary vector
}
\description{
This function calculates outliers according to the rule of 3SD deviations.
}
\seealso{
Other outlier_functions: 
\code{\link{util_hubert}()},
\code{\link{util_sigmagap}()},
\code{\link{util_tukey}()}
}
\concept{outlier}
\concept{outlier_functions}
\keyword{internal}
