% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_abbreviate.R
\name{util_abbreviate}
\alias{util_abbreviate}
\title{Abbreviate snake_case function names to shortened \code{CamelCase}}
\usage{
util_abbreviate(x)
}
\arguments{
\item{x}{a vector of indicator function names}
}
\value{
abbreviations
}
\description{
Abbreviate snake_case function names to shortened \code{CamelCase}
}
