% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_adjust_data_type.R
\name{util_adjust_data_type}
\alias{util_adjust_data_type}
\title{Adjust the data types of study data, if needed}
\usage{
util_adjust_data_type(study_data, meta_data, relevant_vars_for_warnings)
}
\arguments{
\item{study_data}{\link{data.frame} the study data}

\item{meta_data}{\link{meta_data} \code{VAR_NAMES} relevant for warnings about
conversion errors}

\item{relevant_vars_for_warnings}{\link{character}}
}
\value{
\link{data.frame} modified study data
}
\description{
Adjust the data types of study data, if needed
}
\keyword{internal}
