% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_adjust_geom_text_for_plotly.R
\name{util_adjust_geom_text_for_plotly}
\alias{util_adjust_geom_text_for_plotly}
\title{Place all geom_texts also in \code{plotly} right from the x position}
\usage{
util_adjust_geom_text_for_plotly(plotly)
}
\arguments{
\item{plotly}{the \code{plotly}}
}
\value{
modified \code{plotly}-built object
}
\description{
Place all geom_texts also in \code{plotly} right from the x position
}
