% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_alias2caption.R
\name{util_alias2caption}
\alias{util_alias2caption}
\title{Create a caption from an alias name of a \code{dq_report2} result}
\usage{
util_alias2caption(alias, long = FALSE)
}
\arguments{
\item{alias}{alias name}

\item{long}{not for menu entry}
}
\value{
caption
}
\description{
Create a caption from an alias name of a \code{dq_report2} result
}
