% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_all_ind_functions.R
\name{util_all_ind_functions}
\alias{util_all_ind_functions}
\title{All indicator functions of \code{dataquieR}}
\usage{
util_all_ind_functions()
}
\value{
\link{character} names of all indicator functions
}
\description{
All indicator functions of \code{dataquieR}
}
\keyword{internal}
