% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_all_is_integer.R
\name{util_all_is_integer}
\alias{util_all_is_integer}
\title{convenience function to abbreviate all(util_is_integer(...))}
\usage{
util_all_is_integer(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
\code{TRUE}, if all entries are integer-like, \code{FALSE} otherwise
}
\description{
convenience function to abbreviate all(util_is_integer(...))
}
