% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_anytime_installed.R
\name{util_anytime_installed}
\alias{util_anytime_installed}
\title{Test, if package \code{anytime} is installed}
\usage{
util_anytime_installed()
}
\value{
\code{TRUE} if \code{anytime} is installed.
}
\description{
Test, if package \code{anytime} is installed
}
\seealso{
\link{requireNamespace}

\url{https://forum.posit.co/t/how-can-i-make-testthat-think-i-dont-have-a-package-installed/33441/2}

\link{util_ensure_suggested}
}
\keyword{internal}
