% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_app_cd.R
\name{util_app_cd}
\alias{util_app_cd}
\title{utility function for the applicability of contradiction checks}
\usage{
util_app_cd(x, dta)
}
\arguments{
\item{x}{\link{data.frame} metadata}

\item{dta}{\link{logical} vector, 1=matching data type,
0 = non-matching data type}
}
\value{
\link{factor} 0-3 for each variable in metadata
\itemize{
\item 0 data type mismatch and not applicable
\item 1 data type mismatches but applicable
\item 2 data type matches but not applicable
\item 3 data type matches and applicable
}
}
\description{
Test for applicability of contradiction checks
}
