% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_app_iav.R
\name{util_app_iav}
\alias{util_app_iav}
\title{utility function for the applicability of numeric admissibility}
\usage{
util_app_iav(x, dta)
}
\arguments{
\item{x}{\link{data.frame} metadata}

\item{dta}{\link{logical} vector, 1=matching data type,
0 = non-matching data type}
}
\value{
\link{factor} 0-3 for each variable in metadata
\itemize{
\item 0 data type mismatch and not applicable
\item 1 data type mismatches but applicable
\item 2 data type matches but not applicable
\item 3 data type matches and applicable
\item 4 not applicable because of not suitable data type
}
}
\description{
Test for applicability of numeric admissibility
}
\seealso{
\link{pro_applicability_matrix}
}
\keyword{internal}
