% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_as_color.R
\name{util_as_color}
\alias{util_as_color}
\title{convert a character to a specific ordered factor}
\usage{
util_as_color(x)
}
\arguments{
\item{x}{color name}
}
\value{
factor
}
\description{
convert a character to a specific ordered factor
}
