% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_as_numeric.R
\name{util_as_numeric}
\alias{util_as_numeric}
\title{Convert factors to label-corresponding numeric values}
\usage{
util_as_numeric(v, warn)
}
\arguments{
\item{v}{the vector}

\item{warn}{if not missing: character with error message stating conversion
error}
}
\value{
the converted vector
}
\description{
Converts a vector factor aware of numeric values not being scrambled.
}
