% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot_figure_plotly.R
\name{util_as_plotly_from_res}
\alias{util_as_plotly_from_res}
\title{Return the pre-computed \code{plotly} from a \code{dataquieR} result}
\usage{
util_as_plotly_from_res(res, ...)
}
\arguments{
\item{res}{the \code{dataquieR} result}

\item{...}{not used}
}
\value{
a \code{plotly} object
}
\description{
Return the pre-computed \code{plotly} from a \code{dataquieR} result
}
\concept{plotly_shims}
\keyword{internal}
