% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_as_valid_missing_codes.R
\name{util_as_valid_missing_codes}
\alias{util_as_valid_missing_codes}
\title{Convert \code{x} to valid missing codes}
\usage{
util_as_valid_missing_codes(x)
}
\arguments{
\item{x}{\link{character} a vector of values}
}
\value{
converted \code{x}
}
\description{
Convert \code{x} to valid missing codes
}
\seealso{
Other robustness_functions: 
\code{\link{util_check_one_unique_value}()},
\code{\link{util_correct_variable_use}()},
\code{\link{util_empty}()},
\code{\link{util_ensure_character}()},
\code{\link{util_ensure_in}()},
\code{\link{util_ensure_suggested}()},
\code{\link{util_expect_scalar}()},
\code{\link{util_fix_rstudio_bugs}()},
\code{\link{util_is_integer}()},
\code{\link{util_is_numeric_in}()},
\code{\link{util_is_valid_missing_codes}()},
\code{\link{util_match_arg}()},
\code{\link{util_observations_in_subgroups}()},
\code{\link{util_only_NAs}()},
\code{\link{util_stop_if_not}()},
\code{\link{util_warn_unordered}()}
}
\concept{robustness_functions}
\keyword{internal}
