% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_backtickQuote.R
\name{util_backtickQuote}
\alias{util_backtickQuote}
\title{utility function to set string in backticks}
\usage{
util_backtickQuote(x)
}
\arguments{
\item{x}{variable names}
}
\value{
quoted variable names
}
\description{
Quote a set of variable names with backticks
}
\seealso{
\link{util_bQuote}

Other process_functions: 
\code{\link{util_abbreviate}()},
\code{\link{util_all_is_integer}()},
\code{\link{util_attach_attr}()},
\code{\link{util_bQuote}()},
\code{\link{util_coord_flip}()},
\code{\link{util_count_NA}()},
\code{\link{util_extract_matches}()},
\code{\link{util_par_pmap}()},
\code{\link{util_setup_rstudio_job}()},
\code{\link{util_suppress_output}()}
}
\concept{data_management}
\concept{process_functions}
\keyword{internal}
