% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bQuote.R
\name{util_bQuote}
\alias{util_bQuote}
\title{Put in back-ticks}
\usage{
util_bQuote(x)
}
\arguments{
\item{x}{a string}
}
\value{
x in back-ticks
}
\description{
also escape potential back-ticks in x
}
\seealso{
\link{util_backtickQuote}

Other process_functions: 
\code{\link{util_abbreviate}()},
\code{\link{util_all_is_integer}()},
\code{\link{util_attach_attr}()},
\code{\link{util_backtickQuote}()},
\code{\link{util_coord_flip}()},
\code{\link{util_count_NA}()},
\code{\link{util_extract_matches}()},
\code{\link{util_par_pmap}()},
\code{\link{util_setup_rstudio_job}()},
\code{\link{util_suppress_output}()}
}
\concept{data_management}
\concept{process_functions}
\keyword{internal}
