% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cast_off.R
\name{util_cast_off}
\alias{util_cast_off}
\title{Data frame leaves \code{haven}}
\usage{
util_cast_off(df, symb, .dont_cast_off_cols = FALSE)
}
\arguments{
\item{df}{\link{data.frame} may have or contain non-standard classes}

\item{symb}{\link{character} name of the data frame for error messages}

\item{.dont_cast_off_cols}{\link{logical} internal use, only.}
}
\value{
\link{data.frame} having all known special things removed
}
\description{
if \code{df} is/contains a \code{haven} \code{labelled} or \code{tibble} object, convert it to
a base R data frame
}
\keyword{internal}
