% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_data_type.R
\name{util_check_data_type}
\alias{util_check_data_type}
\title{Support function to verify the data type of a value}
\usage{
util_check_data_type(x, type)
}
\arguments{
\item{x}{the value}

\item{type}{expected data type}
}
\value{
\link{logical} whether x is of the expected type
}
\description{
Function to verify the data type of a value.
}
