% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_one_unique_value.R
\name{util_check_one_unique_value}
\alias{util_check_one_unique_value}
\title{Check for one value only}
\usage{
util_check_one_unique_value(x)
}
\arguments{
\item{x}{vector with values}
}
\value{
logical(1): TRUE, if -- except NA -- exactly only one value
is observed in \code{x},
FALSE otherwise
}
\description{
utility function to identify variables with one value only.
}
