% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cll_nm2fkt_nm.R
\name{util_cll_nm2fkt_nm}
\alias{util_cll_nm2fkt_nm}
\title{Get Function called for a Call Name}
\usage{
util_cll_nm2fkt_nm(cll_name, report)
}
\arguments{
\item{cll_name}{\link{character} then systematic function call name to fetch its
function name}

\item{report}{\link{dataquieR_resultset2} the report}
}
\value{
\link{character} the function name
}
\description{
get aliases from report attributes and then replace them by the actual
function name
}
\keyword{internal}
