% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_col2rgb.R
\name{util_col2rgb}
\alias{util_col2rgb}
\title{Return hex code colors from color names or \code{STATAReporter} syntax}
\usage{
util_col2rgb(colors)
}
\arguments{
\item{colors}{the colors, e.g.,"255 0 0" or "red" or "#ff0000"}
}
\value{
character vector with colors using \code{HTML} hexadecimal encoding, e..g,
"#ff0000" for "red"
}
\description{
Return hex code colors from color names or \code{STATAReporter} syntax
}
\keyword{internal}
