% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_col_description.R
\name{util_col_description}
\alias{util_col_description}
\title{Get description for a call}
\usage{
util_col_description(cn)
}
\arguments{
\item{cn}{the call name}
}
\value{
the description
}
\description{
Get description for a call
}
\keyword{internal}
