% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_collapse_msgs.R
\name{util_collapse_msgs}
\alias{util_collapse_msgs}
\title{Collect all errors, warnings, or messages so that they are combined for a combined result}
\usage{
util_collapse_msgs(class, all_of_f)
}
\description{
Collect all errors, warnings, or messages so that they are combined for a combined result
}
\keyword{internal}
