% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_combine_list_report_summaries.R
\name{util_combine_list_report_summaries}
\alias{util_combine_list_report_summaries}
\title{Create a data frame containing all the results from summaries of reports}
\usage{
util_combine_list_report_summaries(
  to_combine,
  type = c("unique_vars", "repeated_vars")
)
}
\arguments{
\item{to_combine}{\link{vector} a list containing the summaries of reports
obtained with \code{summary(report)}}

\item{type}{\link{character} if \code{type} is \code{unique_vars} it means that the variable
names are unique and there is not need to add a prefix
to the variables and labels to specify the
report of origin.
If \code{type} is \code{repeated_vars} a prefix will be used to
specify the report of origin of each variable}
}
\value{
a summary of summaries of \code{dataquieR} reports
}
\description{
Create a data frame containing all the results from summaries of reports
}
