% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_combine_res.R
\name{util_combine_res}
\alias{util_combine_res}
\title{Combine results for Single Variables}
\usage{
util_combine_res(all_of_f)
}
\arguments{
\item{all_of_f}{all results of a function}
}
\value{
row-bound combined results
}
\description{
to, e.g., a data frame with one row per variable or a similar heat-map,
see \code{\link[=print.ReportSummaryTable]{print.ReportSummaryTable()}}.
}
