% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalize_value_labels.R
\name{util_combine_value_label_tables}
\alias{util_combine_value_label_tables}
\title{Combine two value lists}
\usage{
util_combine_value_label_tables(vlt1, vlt2)
}
\arguments{
\item{vlt1}{\link{value_label_table}}

\item{vlt2}{\link{value_label_table}}
}
\value{
\link{value_label_table}
}
\description{
Combine two value lists
}
\examples{
\dontrun{
util_combine_value_label_tables(
  tibble::tribble(~ CODE_VALUE, ~ CODE_LABEL, 17L, "Test", 19L, "Test", 17L, "TestX"),
  tibble::tribble(~ CODE_VALUE, ~ CODE_LABEL, 17L, "Test", 19L, "Test", 17L, "TestX"))
}
}
\keyword{internal}
