% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_compare_meta_with_study.R
\name{util_compare_meta_with_study}
\alias{util_compare_meta_with_study}
\title{Compares study data data types with the ones expected according to the
metadata}
\usage{
util_compare_meta_with_study(sdf, mdf, label_col)
}
\arguments{
\item{sdf}{the \link{data.frame} of study data}

\item{mdf}{the \link{data.frame} of associated static meta data}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}
}
\value{
a binary vector \verb{(0, 1)} if data type applies
}
\description{
Utility function to compare data type of study data with those defined
in metadata
}
