% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_compress_ggplots_in_res.R
\name{util_compress_ggplots_in_res}
\alias{util_compress_ggplots_in_res}
\title{Remove specific classes from a ggplot \code{plot_env} environment}
\usage{
util_compress_ggplots_in_res(r)
}
\arguments{
\item{r}{the object}
}
\description{
Useful to remove large objects before writing to disk with \code{qs} or \code{rds}.
Also deletes parent environment of the plot environment.
Also deletes unneeded variables
}
\seealso{
\href{https://github.com/tidyverse/ggplot2/issues/3619#issuecomment-628021555}{HERE}
}
\keyword{internal}
