% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_summary.R
\name{util_compute_SE_skewness}
\alias{util_compute_SE_skewness}
\title{Compute SE.Skewness}
\usage{
util_compute_SE_skewness(x, skewness = util_compute_skewness(x))
}
\arguments{
\item{x}{data}

\item{skewness}{if already known}
}
\value{
the standard error of skewness
}
\description{
Compute SE.Skewness
}
