% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_summary.R
\name{util_compute_skewness}
\alias{util_compute_skewness}
\title{Compute the Skewness}
\usage{
util_compute_skewness(x)
}
\arguments{
\item{x}{data}
}
\value{
the Skewness
}
\description{
Compute the Skewness
}
