% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_condition_constructor_factory.R
\name{util_condition_constructor_factory}
\alias{util_condition_constructor_factory}
\title{Produce a condition function}
\usage{
util_condition_constructor_factory(
  .condition_type = c("error", "warning", "message")
)
}
\arguments{
\item{.condition_type}{\link{character} the type of the conditions being created
and signaled by the function, "error", "warning", or "message"}
}
\description{
Produce a condition function
}
\seealso{
Other condition_functions: 
\code{\link{util_deparse1}()},
\code{\link{util_error}()},
\code{\link{util_find_external_functions_in_stacktrace}()},
\code{\link{util_find_first_externally_called_functions_in_stacktrace}()},
\code{\link{util_find_indicator_function_in_callers}()},
\code{\link{util_message}()},
\code{\link{util_suppress_warnings}()},
\code{\link{util_warning}()}
}
\concept{condition_functions}
\concept{process}
\keyword{internal}
