% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_is_try_error.R
\name{util_condition_from_try_error}
\alias{util_condition_from_try_error}
\title{Extract condition from try error}
\usage{
util_condition_from_try_error(x)
}
\arguments{
\item{x}{the try-error object}
}
\value{
\link{condition} of the try-error
}
\description{
Extract condition from try error
}
\keyword{internal}
