% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_copy_all_deps.R
\name{util_copy_all_deps}
\alias{util_copy_all_deps}
\title{Copy default dependencies to the report's lib directory}
\usage{
util_copy_all_deps(dir, pages, ...)
}
\arguments{
\item{dir}{report directory}

\item{pages}{all pages to write}

\item{...}{additional \code{htmltools::htmlDependency} objects to be added to all
pages, also}
}
\value{
\code{invisible(NULL)}
}
\description{
Copy default dependencies to the report's lib directory
}
\seealso{
Other reporting_functions: 
\code{\link{util_alias2caption}()},
\code{\link{util_create_page_file}()},
\code{\link{util_eval_to_dataquieR_result}()},
\code{\link{util_evaluate_calls}()},
\code{\link{util_float_index_menu}()},
\code{\link{util_generate_anchor_link}()},
\code{\link{util_generate_anchor_tag}()},
\code{\link{util_generate_calls}()},
\code{\link{util_generate_calls_for_function}()},
\code{\link{util_load_manual}()},
\code{\link{util_make_data_slot_from_table_slot}()},
\code{\link{util_make_function}()},
\code{\link{util_order_by_order}()},
\code{\link{util_set_size}()}
}
\concept{process}
\concept{reporting_functions}
\keyword{internal}
