% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_count_code_classes.R
\name{util_count_code_classes}
\alias{util_count_code_classes}
\title{count distinct realizations of missing codes of a specific class}
\usage{
util_count_code_classes(sdf, mdf, variables, name, list, warn = FALSE)
}
\arguments{
\item{sdf}{study data}

\item{mdf}{meta data}

\item{variables}{variables}

\item{name}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{list}{\link{variable attribute} \code{JUMP_LIST} or \code{MISSING_LIST}: Count which
categories.}

\item{warn}{\link{logical} emit warnings on non-numeric missing codes}
}
\value{
a vector with the number of distinct realized missing codes of the
missing class referred by \code{list} per \code{variables}.
}
\description{
count numbers of distinct codes of class missings or of class jump jumps
}
