% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_count_codes.R
\name{util_count_codes}
\alias{util_count_codes}
\title{count realizations of missing codes of any class}
\usage{
util_count_codes(sdf, mdf, variables, list, name, warn = TRUE)
}
\arguments{
\item{sdf}{study data}

\item{mdf}{meta data}

\item{variables}{variables}

\item{list}{\link{variable attribute} \code{JUMP_LIST} or \code{MISSING_LIST}: Count which
categories.}

\item{name}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{warn}{\link{logical} emit warnings on non-numeric missing codes}
}
\value{
a vector with the total number of missings of the class referred by
\code{list} per \code{variables}.
}
\description{
count total numbers of any sort of missing codes (MISSING or JUMP)
}
