% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_count_NA.R
\name{util_count_NA}
\alias{util_count_NA}
\title{Support function to count number of \code{NA}s}
\usage{
util_count_NA(x)
}
\arguments{
\item{x}{object to count \code{NA}s in}
}
\value{
number of \code{NA}s
}
\description{
Counts the number of \code{NA}s in x.
}
\seealso{
Other process_functions: 
\code{\link{util_abbreviate}()},
\code{\link{util_all_is_integer}()},
\code{\link{util_attach_attr}()},
\code{\link{util_bQuote}()},
\code{\link{util_backtickQuote}()},
\code{\link{util_coord_flip}()},
\code{\link{util_extract_matches}()},
\code{\link{util_par_pmap}()},
\code{\link{util_setup_rstudio_job}()},
\code{\link{util_suppress_output}()}
}
\concept{process}
\concept{process_functions}
\keyword{internal}
