% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_create_report_by_overview.R
\name{util_create_report_by_overview}
\alias{util_create_report_by_overview}
\title{Create an overview of the reports created with \code{dq_report_by}}
\usage{
util_create_report_by_overview(
  output_dir,
  strata_column,
  segment_column,
  strata_column_label,
  subgroup,
  mod_label
)
}
\arguments{
\item{output_dir}{\link{character} the directory in which all reports are searched
and the overview is saved}

\item{strata_column}{\link{character} name of a study variable to stratify the
report by. It can be null}

\item{segment_column}{\link{character} name of a metadata attribute
usable to split the report in
sections of variables. It can be null}

\item{strata_column_label}{\link{character} the label of the variable used as
strata_column}

\item{subgroup}{\link{character} optional, to define subgroups of cases}

\item{mod_label}{\link{list} \code{util_ensure_label()} info}
}
\value{
an overview of all \code{dataquieR} reports created with \code{dq_report_by}
}
\description{
Create an overview of the reports created with \code{dq_report_by}
}
