% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dashboard_table.R
\name{util_dashboard_table}
\alias{util_dashboard_table}
\title{Create a dashboard-table from a report summary}
\usage{
util_dashboard_table(repsum)
}
\arguments{
\item{repsum}{a report summary from \code{summary(report)}}
}
\description{
Create a dashboard-table from a report summary
}
\seealso{
Other html: 
\code{\link{util_extract_all_ids}()},
\code{\link{util_generate_pages_from_report}()},
\code{\link{util_get_hovertext}()}
}
\concept{html}
\concept{process}
\keyword{internal}
