% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data_type_conversion.R
\name{util_data_type_conversion}
\alias{util_data_type_conversion}
\title{Data type conversion}
\usage{
util_data_type_conversion(x, type)
}
\arguments{
\item{x}{the value}

\item{type}{expected data type}
}
\value{
the transformed values (if possible)
}
\description{
Utility function to convert a study variable to match the data type given in
the metadata, if possible.
}
\keyword{internal}
