% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_deparse1.R
\name{util_deparse1}
\alias{util_deparse1}
\title{Expression De-Parsing}
\arguments{
\item{expr}{any R expression.}

\item{collapse}{a string, passed to \code{paste()}}

\item{width.cutoff}{integer in [20, 500] determining the
cutoff (in bytes) at which line-breaking is tried.}

\item{...}{further arguments passed to \code{deparse()}.}
}
\value{
the deparsed expression
}
\description{
Turn unevaluated expressions into character strings.
}
\details{
This is a simple utility function for R < 4.0.0 to ensure a string
result (character vector of length one),
typically used in name construction, as \code{util_deparse1(substitute(.))}.

This avoids a dependency on \code{backports} and on R >= 4.0.0.
}
