% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_deprecate_soft.R
\name{util_deprecate_soft}
\alias{util_deprecate_soft}
\title{Deprecate functions and arguments}
\usage{
util_deprecate_soft(
  when,
  what,
  with = NULL,
  details = NULL,
  id = NULL,
  env = rlang::caller_env(),
  user_env = rlang::caller_env(2)
)
}
\arguments{
\item{when}{A string giving the version when the behavior was deprecated.}

\item{what}{A string describing what is deprecated}

\item{with}{An optional string giving a recommended replacement for the deprecated behavior. This takes the same form as what.}

\item{details}{only used, if \code{lifecycle::deprecate_soft} is available}

\item{id}{only used, if \code{lifecycle::deprecate_soft} is available}

\item{env}{only used, if \code{lifecycle::deprecate_soft} is available}

\item{user_env}{only used, if \code{lifecycle::deprecate_soft} is available}
}
\value{
NULL, invisibly.
}
\description{
if available, it calls \code{lifecycle::deprecate_soft}. otherwise, it just
shows a warning.
}
