% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_df_escape.R
\name{util_df_escape}
\alias{util_df_escape}
\title{Escape characters for HTML in a data frame}
\usage{
util_df_escape(x)
}
\arguments{
\item{x}{\link{data.frame} to be escaped}
}
\value{
\link{data.frame} with html escaped content
}
\description{
Escape characters for HTML in a data frame
}
\keyword{internal}
