% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dist_selection.R
\name{util_dist_selection}
\alias{util_dist_selection}
\title{Utility function distribution-selection}
\usage{
util_dist_selection(measurements, meta_data)
}
\arguments{
\item{measurements}{study data}

\item{meta_data}{metadata, not yet used}
}
\value{
data frame with one column for each variable in study data giving
\code{IsInteger}, \code{IsMultCat} and \code{IsNCategory}

\code{IsInteger} contains a guess, if the variable contains integer values or is a
factor

\code{IsMultCat} contains a guess, if the variable has more than two categories,
if  it is categorical or ordinal

\code{NCategory} contains the number of distinct values detected for the variable
}
\description{
This function differentiates the type of measurement variables.
}
