% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ds1_eval_env.R
\name{util_ds1_eval_env}
\alias{util_ds1_eval_env}
\title{Create an environment with several alias names for the study data variables}
\usage{
util_ds1_eval_env(study_data, meta_data = "item_level", label_col = LABEL)
}
\arguments{
\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables. If
\code{study_data} has already been mapped,
i.e., \code{util_ds1_eval_env(ds1, ...)} is
called, this will work too}
}
\description{
generates an environment similar to \code{as.environment(ds1)}, but makes
variables available by their \code{VAR_NAME}, \code{LABEL}, and \code{label_col} - names.
}
