% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_empty.R
\name{util_empty}
\alias{util_empty}
\title{Test, if values of x are empty, i.e. NA or whitespace characters}
\usage{
util_empty(x)
}
\arguments{
\item{x}{the vector to test}
}
\value{
a logical vector, same length as x; TRUE, if resp. element in x is
"empty"
}
\description{
Test, if values of x are empty, i.e. NA or whitespace characters
}
