% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ensure_character.R
\name{util_ensure_character}
\alias{util_ensure_character}
\title{convert a value to character}
\usage{
util_ensure_character(x, error = FALSE, error_msg, ...)
}
\arguments{
\item{x}{the value}

\item{error}{\link{logical} if \code{TRUE}, an error is thrown, a warning otherwise
in case of a conversion error}

\item{error_msg}{error message to be  displayed, if conversion was not
possible}

\item{...}{additional arguments passed to \link{util_error} or \link{util_warning}
respectively in case of an error, and if an \code{error_msg} has
been passed}
}
\value{
as.character(x)
}
\description{
convert a value to character
}
