% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ensure_label.R
\name{util_ensure_label}
\alias{util_ensure_label}
\title{Utility function ensuring valid labels}
\usage{
util_ensure_label(meta_data, label_col)
}
\arguments{
\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}
}
\value{
a list containing the metadata, possibly with adapted labels,
and a string informing about the changes
}
\description{
Utility function ensuring valid labels
}
