% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_error.R
\name{util_error}
\alias{util_error}
\title{Produce an error message with a useful short stack trace. Then it stops the
execution.}
\usage{
util_error(
  m,
  ...,
  applicability_problem = NA,
  integrity_indicator = "none",
  level = 0
)
}
\arguments{
\item{m}{error message or a \link{condition}}

\item{...}{arguments for \link{sprintf} on m, if m is a character}

\item{applicability_problem}{\link{logical} error indicates unsuitable resp_vars}

\item{integrity_indicator}{\link{character} the message is an integrity problem,
here is the indicator abbreviation..}

\item{level}{\link{integer} level of the error message (defaults to 0). Higher
levels are more severe.}
}
\value{
nothing, its purpose is to stop.
}
\description{
Produce an error message with a useful short stack trace. Then it stops the
execution.
}
