% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_extract_all_ids.R
\name{util_extract_all_ids}
\alias{util_extract_all_ids}
\title{Extract all ids from a list of \code{htmltools} objects}
\usage{
util_extract_all_ids(pages)
}
\arguments{
\item{pages}{the list of objects}
}
\value{
a character vector with valid targets
}
\description{
Extract all ids from a list of \code{htmltools} objects
}
\seealso{
Other html: 
\code{\link{util_generate_pages_from_report}()}
}
\concept{html}
\keyword{internal}
