% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_extract_matches.R
\name{util_extract_matches}
\alias{util_extract_matches}
\title{return all matches of an expression}
\usage{
util_extract_matches(data, pattern)
}
\arguments{
\item{data}{TODO}

\item{pattern}{TODO}
}
\value{
TODO
}
\description{
return all matches of an expression
}
\examples{
\dontrun{ # not exported, so not tested
dat0 <- list("a sentence with citation (Ref. 12), (Ref. 13), and then (Ref. 14)",
  "another sentence without reference")
pat <- "Ref. (\\\\d+)"
util_extract_matches(dat0, pat)
}
}
\seealso{
\href{https://stackoverflow.com/a/10215257}{Stack Overflow}
}
\author{
Josh O'Brien
}
