% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_filter_missing_list_table_for_rv.R
\name{util_filter_missing_list_table_for_rv}
\alias{util_filter_missing_list_table_for_rv}
\title{Filter a \code{MISSING_LIST_TABLE} for rows matching the variable \code{rv}}
\usage{
util_filter_missing_list_table_for_rv(table, rv, rv2 = rv)
}
\arguments{
\item{table}{\link{cause_label_df} a data frame with missing codes and
optionally \code{resp_vars}. It also comprises
labels and optionally an interpretation
column with \code{AAPOR} codes. Must already cover
the variable \code{rv}, i.e., item level metadata
is not checked to find the suitable missing
table for \code{rv}.}

\item{rv}{\link{variable} the response variable to filter the missing list for
specified by a label.}

\item{rv2}{\link{variable} the response variable to filter the missing list for
specified by a \code{VAR_NAMES}-name.}
}
\value{
\link{data.frame} the row-wise bound data frames as one data frame
}
\description{
In \code{MISSING_LIST_TABLE}, a column \code{resp_vars} may be specified. If so,
and if, for a row, this column is not empty, then that row only affects the
one variable specified in that cell
}
