% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_filter_repsum.R
\name{util_filter_repsum}
\alias{util_filter_repsum}
\title{Delete rows from summary table for \code{SSI} or non-\code{SSI} variables}
\usage{
util_filter_repsum(
  repsumtab,
  vars_to_include,
  meta_data,
  rownames_of_report,
  label_col
)
}
\arguments{
\item{repsumtab}{\link{data.frame} the report summary table}

\item{vars_to_include}{\code{"study"}, \code{"ssi"} or \code{c("study", "ssi")}. variables to include}

\item{meta_data}{\link{data.frame} old name for \code{item_level}}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}
}
\value{
\link{data.frame} the filtered \code{repsumtab} with attribute
\code{rownames_of_report}, also filtered
}
\description{
Delete rows from summary table for \code{SSI} or non-\code{SSI} variables
}
\keyword{internal}
