% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_find_external_functions_in_stacktrace.R
\name{util_find_external_functions_in_stacktrace}
\alias{util_find_external_functions_in_stacktrace}
\title{Find externally called function in the stack trace}
\usage{
util_find_external_functions_in_stacktrace(
  sfs = rev(sys.frames()),
  cls = rev(sys.calls())
)
}
\arguments{
\item{sfs}{reverse \link{sys.frames} to search in}

\item{cls}{reverse \link{sys.calls} to search in}
}
\value{
vector of \link{logical}s stating for each index, if it had been called
externally
}
\description{
intended use: error messages for the user
}
