% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/util_find_first_externally_called_functions_in_stacktrace.R
\name{util_find_first_externally_called_functions_in_stacktrace}
\alias{util_find_first_externally_called_functions_in_stacktrace}
\title{Find first externally called function in the stack trace}
\usage{
util_find_first_externally_called_functions_in_stacktrace(
  sfs = rev(sys.frames()),
  cls = rev(sys.calls())
)
}
\arguments{
\item{sfs}{reverse \link{sys.frames} to search in}

\item{cls}{reverse \link{sys.calls} to search in}
}
\value{
reverse \link{sys.frames} index of first non-dataquieR function in
this stack
}
\description{
intended use: error messages for the user
}
\seealso{
Other condition_functions: 
\code{\link{util_condition_constructor_factory}()},
\code{\link{util_deparse1}()},
\code{\link{util_error}()},
\code{\link{util_find_external_functions_in_stacktrace}()},
\code{\link{util_find_indicator_function_in_callers}()},
\code{\link{util_message}()},
\code{\link{util_warning}()}
}
\concept{condition_functions}
\keyword{internal}
