% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_find_free_missing_code.R
\name{util_find_free_missing_code}
\alias{util_find_free_missing_code}
\title{Check, if \code{x} contains valid missing codes}
\usage{
util_find_free_missing_code(x)
}
\arguments{
\item{x}{a vector of missing codes}
}
\value{
a missing code not in \code{x}
}
\description{
Check, if \code{x} contains valid missing codes
}
\seealso{
Other metadata_management: 
\code{\link{util_dist_selection}()},
\code{\link{util_find_var_by_meta}()},
\code{\link{util_get_var_att_names_of_level}()},
\code{\link{util_get_vars_in_segment}()},
\code{\link{util_looks_like_missing}()},
\code{\link{util_no_value_labels}()},
\code{\link{util_validate_known_meta}()},
\code{\link{util_validate_missing_lists}()}
}
\concept{metadata_management}
\keyword{internal}
