% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_find_indicator_function_in_callers.R
\name{util_find_indicator_function_in_callers}
\alias{util_find_indicator_function_in_callers}
\title{Search for a formal in the stack trace}
\usage{
util_find_indicator_function_in_callers(symbol = "resp_vars")
}
\arguments{
\item{symbol}{symbol to find}
}
\value{
value of the symbol, if available, \code{NULL} otherwise
}
\description{
Similar to \code{\link[=dynGet]{dynGet()}}, find a symbol in the closest data quality indicator
function and return its value. Can \code{stop()}, if symbol evaluation causes a
stop.
}
\seealso{
Other condition_functions: 
\code{\link{util_condition_constructor_factory}()},
\code{\link{util_deparse1}()},
\code{\link{util_error}()},
\code{\link{util_find_external_functions_in_stacktrace}()},
\code{\link{util_find_first_externally_called_functions_in_stacktrace}()},
\code{\link{util_message}()},
\code{\link{util_suppress_warnings}()},
\code{\link{util_warning}()}
}
\concept{condition_functions}
\keyword{internal}
