% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_summary.R
\name{util_first_row_to_colnames}
\alias{util_first_row_to_colnames}
\title{Move the first row of a data frame to its column names}
\usage{
util_first_row_to_colnames(dfr)
}
\arguments{
\item{dfr}{\link{data.frame}}
}
\value{
\link{data.frame} with first row as column names
}
\description{
Move the first row of a data frame to its column names
}
