% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fix_merge_dups.R
\name{util_fix_merge_dups}
\alias{util_fix_merge_dups}
\title{Fix results from merge}
\usage{
util_fix_merge_dups(dfr, stop_if_incompatible = TRUE)
}
\arguments{
\item{dfr}{data frame to fix}

\item{stop_if_incompatible}{\link{logical} stop if data frame can not be fixed}
}
\description{
this function handles the result of \code{\link[=merge]{merge()}}-calls, if \code{no.dups = TRUE} and
\code{suffixes = c("", "")}
}
\seealso{
Other data_management: 
\code{\link{util_assign_levlabs}()},
\code{\link{util_check_data_type}()},
\code{\link{util_check_group_levels}()},
\code{\link{util_compare_meta_with_study}()},
\code{\link{util_dichotomize}()},
\code{\link{util_merge_data_frame_list}()},
\code{\link{util_rbind}()},
\code{\link{util_remove_na_records}()},
\code{\link{util_replace_hard_limit_violations}()},
\code{\link{util_round_to_decimal_places}()},
\code{\link{util_study_var2factor}()},
\code{\link{util_table_of_vct}()}
}
\concept{data_management}
\concept{summary}
\keyword{internal}
