% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fix_rstudio_bugs.R
\name{util_fix_rstudio_bugs}
\alias{util_fix_rstudio_bugs}
\title{RStudio crashes on parallel calls in some versions on Darwin based operating
systems with R 4}
\usage{
util_fix_rstudio_bugs()
}
\value{
invisible null
}
\description{
RStudio crashes on parallel calls in some versions on Darwin based operating
systems with R 4
}
