% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_finalize_sizing_hints.R
\name{util_fix_sizing_hints}
\alias{util_fix_sizing_hints}
\title{Ensure, sizing hint sticks at the \code{dqr}, only}
\usage{
util_fix_sizing_hints(dqr, x)
}
\arguments{
\item{dqr}{a \code{dataquieR} result}

\item{x}{a plot object}
}
\value{
a list with \code{dqr} and \code{x}, but fixed
}
\description{
Ensure, sizing hint sticks at the \code{dqr}, only
}
\keyword{internal}
