% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_storr_factory.R
\name{util_fix_storr_object}
\alias{util_fix_storr_object}
\title{Fix a \code{storr} object, if it features the factory-attribute}
\usage{
util_fix_storr_object(my_storr_object)
}
\arguments{
\item{my_storr_object}{a \code{storr}-object}
}
\value{
a (hopefully) working \code{storr_object}
}
\description{
Fix a \code{storr} object, if it features the factory-attribute
}
\seealso{
\code{\link[=util_storr_factory]{util_storr_factory()}}
}
\keyword{internal}
