% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_float_index_menu.R
\name{util_float_index_menu}
\alias{util_float_index_menu}
\title{return a single page navigation menu floating on the right}
\usage{
util_float_index_menu(index_menu_table, object)
}
\arguments{
\item{index_menu_table}{\link{data.frame} columns: links, hovers, texts}

\item{object}{\code{htmltools} tag list, used, instead of \code{index_menu_table},
if passed}
}
\description{
if displayed in a \code{dq_report2}
}
\examples{
\dontrun{
util_float_index_menu(tibble::tribble(
   ~ links, ~ hovers, ~ texts,
   "http://www.google.de/#xxx", "This is Google", "to Google",
   "http://www.uni-giessen.de/#xxx", "This is Gießen", "cruising on the A45"
))
}
}
