% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_function_description.R
\name{util_function_description}
\alias{util_function_description}
\title{Get description for an indicator function}
\usage{
util_function_description(fname)
}
\arguments{
\item{fname}{the function name}
}
\value{
the description
}
\description{
Get description for an indicator function
}
