% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generate_calls_for_function.R
\name{util_generate_calls_for_function}
\alias{util_generate_calls_for_function}
\title{Generate function calls for a given indicator function}
\usage{
util_generate_calls_for_function(
  fkt,
  meta_data,
  label_col,
  meta_data_segment,
  meta_data_dataframe,
  meta_data_cross_item,
  specific_args,
  arg_overrides,
  resp_vars
)
}
\arguments{
\item{fkt}{the indicator function's name}

\item{meta_data}{the item level metadata data frame}

\item{label_col}{the label column}

\item{meta_data_segment}{segment level metadata}

\item{meta_data_dataframe}{data frame level metadata}

\item{meta_data_cross_item}{cross-item level metadata}

\item{specific_args}{argument overrides for specific functions}

\item{arg_overrides}{general argument overrides}

\item{resp_vars}{variables to be respected}
}
\value{
function calls for the given function
}
\description{
new reporting pipeline v2.0
}
