% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generate_table_indicators_descriptors.R
\name{util_generate_table_indicators_descriptors}
\alias{util_generate_table_indicators_descriptors}
\title{Create a table summarizing the number of indicators and descriptors
in the report}
\usage{
util_generate_table_indicators_descriptors(x)
}
\arguments{
\item{x}{a report summary (\code{summary(r)})}
}
\value{
a table containing the number of indicators and descriptors created
in the report, separated by data quality dimension.
}
\description{
Create a table summarizing the number of indicators and descriptors
in the report
}
\keyword{internal}
