% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_color_for_result.R
\name{util_get_color_for_result}
\alias{util_get_color_for_result}
\title{Return the color for a result}
\usage{
util_get_color_for_result(
  result,
  aspect = c("applicability", "error", "issue"),
  ...
)
}
\arguments{
\item{result}{a \code{dataquieR_resultset2} result}

\item{aspect}{an aspect/problem category of results
(error, applicability error or data quality issue)}

\item{...}{not used}
}
\value{
a color as a grading
}
\description{
messages do not cause any coloring, warnings are yellow, errors are red
}
