% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_colors.R
\name{util_get_colors}
\alias{util_get_colors}
\title{Get colors for each russet \code{DQ} category}
\usage{
util_get_colors()
}
\value{
named vector of colors, names are categories (e.g, "1" to "5")
values are colors as \code{HTML} \code{RGB} hexadecimal strings
}
\description{
Get colors for each russet \code{DQ} category
}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()},
\code{\link{util_as_cat}()},
\code{\link{util_as_integer_cat}()},
\code{\link{util_extract_indicator_metrics}()},
\code{\link{util_get_category_for_result}()},
\code{\link{util_get_labels_grading_class}()},
\code{\link{util_get_message_for_result}()},
\code{\link{util_get_rule_sets}()},
\code{\link{util_get_ruleset_formats}()},
\code{\link{util_get_thresholds}()},
\code{\link{util_html_table}()},
\code{\link{util_sort_by_order}()}
}
\concept{process}
\concept{summary_functions}
\keyword{internal}
