% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_concept_info.R
\name{util_get_concept_info}
\alias{util_get_concept_info}
\title{Read additional concept tables}
\usage{
util_get_concept_info(filename, ...)
}
\arguments{
\item{filename}{RDS-file name without extension to read from}

\item{...}{passed to \link{subset}}
}
\value{
a data frame
}
\description{
Read additional concept tables
}
\concept{concept_functions}
\concept{reporting}
\keyword{internal}
