% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_encoding.R
\name{util_get_encoding}
\alias{util_get_encoding}
\title{Get encoding from metadata or guess it from data}
\usage{
util_get_encoding(
  resp_vars = colnames(study_data),
  study_data,
  label_col,
  meta_data,
  meta_data_dataframe
)
}
\arguments{
\item{resp_vars}{\link{variable} the names of the measurement variables, if
missing or \code{NULL}, all variables will be checked}

\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{meta_data}{\link{data.frame} old name for \code{item_level}}

\item{meta_data_dataframe}{\link{data.frame} the data frame that contains the
metadata for the data frame level}
}
\value{
named vector of valid encoding strings matching \code{resp_vars}
}
\description{
Get encoding from metadata or guess it from data
}
