% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_fg_color.R
\name{util_get_fg_color}
\alias{util_get_fg_color}
\title{Find a foreground color for a background}
\usage{
util_get_fg_color(cl)
}
\arguments{
\item{cl}{colors}
}
\value{
black or white for each cl
}
\description{
black or white
}
\seealso{
\href{https://stackoverflow.com/a/24810681}{\code{stackoverflow.com}}
}
\keyword{internal}
