% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_hovertext.R
\name{util_get_hovertext}
\alias{util_get_hovertext}
\title{Import vector of hover text for tables in the report}
\usage{
util_get_hovertext(x)
}
\arguments{
\item{x}{name of the tables. They are \code{meta_data},
\code{meta_data_segment}, \code{meta_data_dataframe}, \code{meta_data_cross_item},
\code{meta_data_item_computation}, \code{com_item_missingness},
\code{int_datatype_matrix}, \code{con_inadmissible_categorical},
\code{rulesetformat}, \code{gradingruleset}}
}
\value{
named vector containing the hover text from the file \code{metadata-hovertext.rds}
in the inst folder. Names correspond to column names in the metadata
tables
}
\description{
Import vector of hover text for tables in the report
}
\seealso{
Other html: 
\code{\link{util_dashboard_table}()},
\code{\link{util_extract_all_ids}()},
\code{\link{util_generate_pages_from_report}()}
}
\concept{html}
\concept{reporting}
\keyword{internal}
