% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_redcap_rule_env.R
\name{util_get_redcap_rule_env}
\alias{util_get_redcap_rule_env}
\title{an environment with functions available for \code{REDcap} rules}
\usage{
util_get_redcap_rule_env()
}
\value{
environment
}
\description{
an environment with functions available for \code{REDcap} rules
}
