% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_ruleset_formats.R
\name{util_get_ruleset_formats}
\alias{util_get_ruleset_formats}
\title{Get formats for \code{DQ} categories}
\usage{
util_get_ruleset_formats()
}
\value{
\link{data.frame} columns: \code{categories} (e.g., "1" to "5"),
\code{color} (e.g., "33 102 172", "67 147 195", "227 186 20", "214 96 77",
178 23 43"), \code{label} (e.g., "OK", "unclear", "moderate", "important",
"critical" )
}
\description{
Get formats for \code{DQ} categories
}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()},
\code{\link{util_as_cat}()},
\code{\link{util_extract_indicator_metrics}()},
\code{\link{util_get_category_for_result}()},
\code{\link{util_get_colors}()},
\code{\link{util_get_html_cell_for_result}()},
\code{\link{util_get_labels_grading_class}()},
\code{\link{util_get_message_for_result}()},
\code{\link{util_get_rule_sets}()},
\code{\link{util_get_thresholds}()},
\code{\link{util_html_table}()},
\code{\link{util_melt_summary}()},
\code{\link{util_sort_by_order}()}
}
\concept{summary_functions}
\keyword{internal}
