% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR_resultset2.R
\name{util_get_storr_att_namespace}
\alias{util_get_storr_att_namespace}
\title{Get namespace for attributes}
\usage{
util_get_storr_att_namespace(my_storr_object)
}
\arguments{
\item{my_storr_object}{the \code{storr} object}
}
\value{
the namespace name
}
\description{
Get namespace for attributes
}
\keyword{internal}
