% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_storr_factory.R
\name{util_get_storr_object_from_report}
\alias{util_get_storr_object_from_report}
\title{Get the \code{storr} object backing a report}
\usage{
util_get_storr_object_from_report(r)
}
\arguments{
\item{r}{the dataquieR_resultset2 / report}
}
\value{
the \code{storr} object holding the results or \code{NULL}, if the report
lives in the memory, only
}
\description{
Get the \code{storr} object backing a report
}
\keyword{internal}
