% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_thresholds.R
\name{util_get_thresholds}
\alias{util_get_thresholds}
\title{Get the thresholds for grading}
\usage{
util_get_thresholds(indicator_metric, meta_data)
}
\arguments{
\item{indicator_metric}{which indicator metric to be classified}

\item{meta_data}{the item level metadata}
}
\value{
named list (names are \code{VAR_NAMES}, values are named vectors of
intervals, names in the vectors are the category numbers)
}
\description{
Get the thresholds for grading
}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()},
\code{\link{util_as_cat}()},
\code{\link{util_extract_indicator_metrics}()},
\code{\link{util_get_category_for_result}()},
\code{\link{util_get_colors}()},
\code{\link{util_get_html_cell_for_result}()},
\code{\link{util_get_labels_grading_class}()},
\code{\link{util_get_message_for_result}()},
\code{\link{util_get_rule_sets}()},
\code{\link{util_get_ruleset_formats}()},
\code{\link{util_html_table}()},
\code{\link{util_melt_summary}()},
\code{\link{util_sort_by_order}()}
}
\concept{summary_functions}
\keyword{internal}
