% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_vars_in_segment.R
\name{util_get_vars_in_segment}
\alias{util_get_vars_in_segment}
\title{Return all variables in the segment \code{segment}}
\usage{
util_get_vars_in_segment(segment, meta_data = "item_level", label_col = LABEL)
}
\arguments{
\item{segment}{\link{character} the segment as specified in \code{STUDY_SEGMENT}}

\item{meta_data}{\link{data.frame} the metadata}

\item{label_col}{\link{character} the metadata attribute used for naming the
variables}
}
\value{
vector of variable names
}
\description{
Return all variables in the segment \code{segment}
}
