% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_voc_tab.R
\name{util_get_voc_tab}
\alias{util_get_voc_tab}
\title{Get the Table with Known Vocabularies}
\usage{
util_get_voc_tab(.data_frame_list = .dataframe_environment())
}
\arguments{
\item{.data_frame_list}{\link{environment} cache for loaded data frames}
}
\value{
\link{data.frame} the (combined) table with known vocabularies
}
\description{
Get the Table with Known Vocabularies
}
\keyword{internal}
