% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_gg_var_label.R
\name{util_gg_var_label}
\alias{util_gg_var_label}
\title{Add labels to \code{ggplot}}
\usage{
util_gg_var_label(
  ...,
  meta_data = get("meta_data", parent.frame()),
  label_col = get("label_col", parent.frame())
)
}
\arguments{
\item{...}{EXPERIMENTAL}

\item{meta_data}{the metadata}

\item{label_col}{the label columns}
}
\value{
a modified \code{ggplot}
}
\description{
EXPERIMENTAL
}
\concept{process}
\keyword{internal}
