% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_has_no_group_vars.R
\name{util_has_no_group_vars}
\alias{util_has_no_group_vars}
\title{Utility function to check whether a variable has no grouping variable assigned}
\usage{
util_has_no_group_vars(resp_vars, label_col = LABEL, meta_data = "item_level")
}
\arguments{
\item{resp_vars}{\link{variable list} the name of a measurement variable}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{meta_data}{\link{data.frame} old name for \code{item_level}}
}
\value{
boolean
}
\description{
Utility function to check whether a variable has no grouping variable assigned
}
