% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_heatmap_1th.R
\name{util_heatmap_1th}
\alias{util_heatmap_1th}
\title{Utility Function Heatmap with 1 Threshold}
\usage{
util_heatmap_1th(
  df,
  cat_vars,
  values,
  threshold,
  right_intv,
  invert,
  cols,
  strata
)
}
\arguments{
\item{df}{\link{data.frame} with data to display as a heatmap.}

\item{cat_vars}{\link{variable list} len=1-2. Variables to group by. Up to 2
group levels supported.}

\item{values}{\link{variable} the name of the percentage variable}

\item{threshold}{\link{numeric} lowest acceptable value}

\item{right_intv}{\link{logical} len=1. If \code{FALSE} (default), intervals used
to define color ranges in the heatmap
are closed on the left side, if \code{TRUE} on
the right side, respectively.}

\item{invert}{\link{logical} len=1. If \code{TRUE}, high values are better, warning
colors are used for low values. \code{FALSE} works
vice versa.}

\item{cols}{deprecated, ignored.}

\item{strata}{\link{variable} optional, the name of a variable
used for stratification}
}
\value{
a \link{list} with:
\itemize{
\item \code{SummaryPlot}: \link{ggplot} object with the heatmap
}
}
\description{
Function to create heatmap-like plot given one threshold -- works for
percentages for now.
}
