% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_hide_file_windows.R
\name{util_hide_file_windows}
\alias{util_hide_file_windows}
\title{If on Windows, hide a file}
\usage{
util_hide_file_windows(fn)
}
\arguments{
\item{fn}{the file path + name}
}
\value{
\code{invisible(NULL)}
}
\description{
If on Windows, hide a file
}
\keyword{internal}
