% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_html_attr_quote_escape.R
\name{util_html_attr_quote_escape}
\alias{util_html_attr_quote_escape}
\title{escape \verb{"}}
\usage{
util_html_attr_quote_escape(s)
}
\arguments{
\item{s}{haystack}
}
\value{
\code{s} with \verb{"} replaced by \verb{&quot;}
}
\description{
escape \verb{"}
}
\concept{reporting}
\keyword{internal}
