% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_html_for_dims.R
\name{util_html_for_dims}
\alias{util_html_for_dims}
\title{Create a dynamic dimension related page for the report}
\usage{
util_html_for_dims(
  report,
  use_plot_ly,
  template,
  block_load_factor,
  repsum,
  dir
)
}
\arguments{
\item{report}{\link{dataquieR_resultset2} a \code{dq_report2} report}

\item{use_plot_ly}{\link{logical} use \code{plotly}, if available.}

\item{template}{\link{character} template to use for the \code{dq_report2} report.}

\item{block_load_factor}{\link{numeric} multiply size of parallel compute blocks
by this factor.}

\item{repsum}{the \code{dataquieR} summary, see \code{\link[=summary]{summary()}} and \code{\link[=dq_report2]{dq_report2()}}}

\item{dir}{\link{character} output directory for potential \code{iframes}.}
}
\value{
list of arguments for \code{append_single_page()} defined locally in
\code{util_generate_pages_from_report()}.
}
\description{
Create a dynamic dimension related page for the report
}
\keyword{internal}
