% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_html_for_var.R
\name{util_html_for_var}
\alias{util_html_for_var}
\title{Create a dynamic single variable page for the report}
\usage{
util_html_for_var(report, cur_var, use_plot_ly, template, note_meta = c())
}
\arguments{
\item{report}{\link{dataquieR_resultset2} a \code{dq_report2} report}

\item{cur_var}{\link{character} variable name for single variable pages}

\item{use_plot_ly}{\link{logical} use \code{plotly}, if available.}

\item{template}{\link{character} template to use for the \code{dq_report2} report.}

\item{note_meta}{\link{character} notes on the metadata for a single variable
(if needed)}
}
\value{
list of arguments for \code{append_single_page()} defined locally in
\code{\link[=util_generate_pages_from_report]{util_generate_pages_from_report()}}.
}
\description{
Create a dynamic single variable page for the report
}
