% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_iframe_it_if_needed.R
\name{util_iframe_it_if_needed}
\alias{util_iframe_it_if_needed}
\title{Make \code{it} scalable, if it is a figure}
\usage{
util_iframe_it_if_needed(it, dir, nm, fkt, sizing_hints, ggthumb)
}
\arguments{
\item{it}{\code{\link[htmltools:builder]{htmltools::tag()}} compatible object}

\item{dir}{\link{character} output directory for potential \code{iframe}.}

\item{nm}{\link{character} name for the \code{iframe}d file, if one is created}

\item{fkt}{\link{character} function name of the indicator function that created
\code{ìt}.}

\item{sizing_hints}{\code{object} additional metadata about the natural figure size}

\item{ggthumb}{\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} optional, underlying \code{ggplot2} object
for a preview}
}
\value{
\code{\link[htmltools:builder]{htmltools::tag()}} compatible object, maybe now in an \code{iframe}
}
\description{
this function writes figures to helper files and embeds these in a returned
object which is a scalable \code{iframe}. it does not change other objects in \code{it}.
}
\keyword{internal}
