% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_init_respum_tab.R
\name{util_init_respum_tab}
\alias{util_init_respum_tab}
\title{Extract all properties of a \code{ReportSummaryTable}}
\usage{
util_init_respum_tab(x)
}
\arguments{
\item{x}{\code{ReportSummaryTable} object}
}
\value{
\link{list} with all properties
}
\description{
Extract all properties of a \code{ReportSummaryTable}
}
\keyword{internal}
