% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_is_integer.R
\name{util_is_integer}
\alias{util_is_integer}
\title{Check for integer values}
\usage{
util_is_integer(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{the object to test}

\item{tol}{precision of the detection. Values deviating more than \code{tol} from
their closest integer value will not be deemed integer.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This function checks if a variable is integer.
}
\seealso{
is.integer

Copied from the documentation of \link[base]{is.integer}

\link[base]{is.integer} detects, if the storage mode of an R-object is
integer. Usually, users want to know, if the values are integer. As suggested
by \link[base]{is.integer}'s documentation, \code{is.wholenumber} does so.
}
