% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_is_na_0_empty_or_false.R
\name{util_is_na_0_empty_or_false}
\alias{util_is_na_0_empty_or_false}
\title{Detect falsish values}
\usage{
util_is_na_0_empty_or_false(x)
}
\arguments{
\item{x}{a value/vector of values}
}
\value{
vector of logical values:
\code{TRUE}, wherever x is somehow empty
}
\description{
Detect falsish values
}
\seealso{
Other missing_functions: 
\code{\link{util_all_intro_vars_for_rv}()},
\code{\link{util_count_expected_observations}()},
\code{\link{util_filter_missing_list_table_for_rv}()},
\code{\link{util_get_code_list}()},
\code{\link{util_observation_expected}()},
\code{\link{util_remove_empty_rows}()},
\code{\link{util_replace_codes_by_NA}()}
}
\concept{missing_functions}
\concept{robustness}
\keyword{internal}
