% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot2svg_object.R
\name{util_is_svg_object}
\alias{util_is_svg_object}
\title{Detect un-disclosed \code{ggplot}}
\usage{
util_is_svg_object(x)
}
\arguments{
\item{x}{the object to check}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Detect un-disclosed \code{ggplot}
}
\keyword{internal}
