% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_is_try_error.R
\name{util_is_try_error}
\alias{util_is_try_error}
\title{Check, if \code{x} is a try-error}
\usage{
util_is_try_error(x)
}
\arguments{
\item{x}{}
}
\value{
\code{\link[=logical]{logical()}} if it is a try-error
}
\description{
Check, if \code{x} is a try-error
}
\keyword{internal}
