% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection.R
\name{util_load_manual}
\alias{util_load_manual}
\title{being called by the active binding function for .manual}
\usage{
util_load_manual()
}
\description{
being called by the active binding function for .manual
}
