% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_make_data_slot_from_table_slot.R
\name{util_make_data_slot_from_table_slot}
\alias{util_make_data_slot_from_table_slot}
\title{Rename columns of a \code{SummaryTable} (or Segment, ...) to look nice}
\usage{
util_make_data_slot_from_table_slot(Table)
}
\arguments{
\item{Table}{\link{data.frame}, a table}
}
\value{
renamed table
}
\description{
Rename columns of a \code{SummaryTable} (or Segment, ...) to look nice
}
