% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_map_all.R
\name{util_map_all}
\alias{util_map_all}
\title{Maps label column metadata on study data variable names}
\usage{
util_map_all(label_col = VAR_NAMES, study_data, meta_data)
}
\arguments{
\item{label_col}{the variable of the metadata that contains the variable
names of the study data}

\item{study_data}{the name of the data frame that contains the measurements}

\item{meta_data}{the name of the data frame that contains metadata attributes
of study data}
}
\value{
\link{list} with slot \code{df} with a study data frame with mapped column
names
}
\description{
Maps a certain label column from the metadata
to the study data frame.
}
\seealso{
Other mapping: 
\code{\link{util_map_by_largest_prefix}()},
\code{\link{util_map_labels}()},
\code{\link{util_recode}()}
}
\concept{mapping}
\concept{metadata_management}
\keyword{internal}
