% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_match_arg.R
\name{util_match_arg}
\alias{util_match_arg}
\title{\code{dataquieR} version of match.arg}
\usage{
util_match_arg(arg, choices, several_ok = FALSE, error = TRUE)
}
\arguments{
\item{arg}{the argument}

\item{choices}{the choices}

\item{several_ok}{allow more than one entry in \code{arg}}

\item{error}{\code{stop()}, if \code{arg} is not in \code{choices}
(warns and cleans \code{arg}, otherwise)}
}
\value{
"cleaned" \code{arg}
}
\description{
does not support partial matching, but will display the most likely
match as a warning/error.
}
