% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_message.R
\name{util_message}
\alias{util_message}
\title{Produce a condition message with a useful short stack trace.}
\usage{
util_message(
  m,
  ...,
  applicability_problem = NA,
  intrinsic_applicability_problem = NA,
  integrity_indicator = "none",
  level = 0,
  immediate
)
}
\arguments{
\item{m}{a message or a \link{condition}}

\item{...}{arguments for \link{sprintf} on m, if m is a character}

\item{applicability_problem}{\link{logical} error indicates missing metadata}

\item{intrinsic_applicability_problem}{\link{logical} error unsuitable resp_vars}

\item{integrity_indicator}{\link{character} the message is an integrity problem,
here is the indicator abbreviation..}

\item{level}{\link{integer} level of the message (defaults to 0). Higher
levels are more severe.}

\item{immediate}{\link{logical} not used.}
}
\value{
\link{condition} the condition object, if the execution is not stopped
}
\description{
Produce a condition message with a useful short stack trace.
}
\seealso{
Other condition_functions: 
\code{\link{util_condition_constructor_factory}()},
\code{\link{util_deparse1}()},
\code{\link{util_error}()},
\code{\link{util_find_external_functions_in_stacktrace}()},
\code{\link{util_find_first_externally_called_functions_in_stacktrace}()},
\code{\link{util_find_indicator_function_in_callers}()},
\code{\link{util_warning}()}
}
\concept{condition_functions}
\concept{process}
\keyword{internal}
