% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalize_clt.R
\name{util_normalize_clt}
\alias{util_normalize_clt}
\title{Distribute \code{CODE_LIST_TABLE} in item level metadata}
\usage{
util_normalize_clt(meta_data)
}
\arguments{
\item{meta_data}{\link{data.frame} old name for \code{item_level}}
}
\value{
\link{meta_data}, but \code{CODE_LIST_TABLE} column is distributed to the
columns \code{VALUE_LABEL_TABLE} and \code{MISSING_LIST_TABLE}, respectively.
}
\description{
fills the columns \code{MISSING_LIST_TABLE} and \code{VALUE_LABEL_TABLE} from
\code{CODE_LIST_TABLE}, if applicable
}
\keyword{internal}
