% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalize_cross_item.R
\name{util_normalize_cross_item}
\alias{util_normalize_cross_item}
\title{Normalize and check cross-item-level metadata}
\usage{
util_normalize_cross_item(
  meta_data = "item_level",
  meta_data_cross_item = "cross-item_level",
  label_col = LABEL
)
}
\arguments{
\item{meta_data}{\link{meta_data}}

\item{meta_data_cross_item}{\link[=meta_data_cross]{cross-item-level metadata}}

\item{label_col}{\link{character} label column to use for variable naming}
}
\value{
normalized and checked \link[=meta_data_cross]{cross-item-level metadata}
}
\description{
Normalize and check cross-item-level metadata
}
\seealso{
\code{\link[=meta_data_cross]{meta_data_cross()}}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_METRIC}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{CONTRADICTION_TYPE}},
\code{\link{DATA_PREPARATION}},
\code{\link{GOLDSTANDARD}},
\code{\link{MULTIVARIATE_OUTLIER_CHECK}},
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}},
\code{\link{N_RULES}},
\code{\link{REL_VAL}},
\code{\link{VARIABLE_LIST}},
\code{\link{meta_data_cross}}
}
\concept{meta_data_cross}
\keyword{internal}
