% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_online_ref.R
\name{util_online_ref}
\alias{util_online_ref}
\title{Creates a Link to our Website}
\usage{
util_online_ref(fkt_name)
}
\arguments{
\item{fkt_name}{\link{character} function name to generate a link for}
}
\value{
\link{character} the link
}
\description{
i.e., to a vignette on the website
}
