% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_only_NAs.R
\name{util_only_NAs}
\alias{util_only_NAs}
\title{identify \code{NA}-only variables}
\usage{
util_only_NAs(x)
}
\arguments{
\item{x}{the variable to check a vector}
}
\value{
flagged binary vector
}
\description{
This utility function identifies variables with \code{NA}s values only.
}
