% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_as_numeric_with_unit.R
\name{util_op_numeric_with_unit}
\alias{util_op_numeric_with_unit}
\title{Operator caring for units}
\usage{
util_op_numeric_with_unit(e1, e2)
}
\arguments{
\item{e1}{first argument}

\item{e2}{second argument}
}
\value{
result
}
\description{
Operator caring for units
}
\keyword{internal}
