% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_open_in_excel.R
\name{util_open_in_excel}
\alias{util_open_in_excel}
\title{Open a data frame in Excel}
\usage{
util_open_in_excel(dfr)
}
\arguments{
\item{dfr}{the data frame}
}
\value{
42
}
\description{
Open a data frame in Excel
}
\keyword{internal}
