% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_optimize_sequence_across_time_var.R
\name{util_optimize_sequence_across_time_var}
\alias{util_optimize_sequence_across_time_var}
\title{Utility function to distribute points across a time variable}
\usage{
util_optimize_sequence_across_time_var(
  time_var_data,
  n_points,
  prop_grid = 0.5
)
}
\arguments{
\item{time_var_data}{vector of the data points of the time variable}

\item{n_points}{maximum number of points to distribute across the
time variable (minimum: 3)}

\item{prop_grid}{proportion of points given in \code{n_points} that should be
distributed in an equally spaced grid across the time
variable (minimum: 0.1, maximum: 1). The remaining
proportion of points will be spaced according to the
distribution of the time variable's data points.}
}
\value{
a sequence of points in datetime format
}
\description{
Utility function to distribute points across a time variable
}
\keyword{internal}
