% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sort_by_order.R
\name{util_order_by_order}
\alias{util_order_by_order}
\title{Get the order of a vector with general order given in some other vector}
\usage{
util_order_by_order(x, order, ...)
}
\arguments{
\item{x}{the vector}

\item{order}{the "order vector}

\item{...}{additional arguments passed to \code{order}}
}
\description{
Get the order of a vector with general order given in some other vector
}
\examples{
\dontrun{
util_order_by_order(c("a", "b", "a", "c", "d"), letters)
}
}
\seealso{
Other reporting_functions: 
\code{\link{util_alias2caption}()},
\code{\link{util_copy_all_deps}()},
\code{\link{util_create_page_file}()},
\code{\link{util_eval_to_dataquieR_result}()},
\code{\link{util_evaluate_calls}()},
\code{\link{util_float_index_menu}()},
\code{\link{util_generate_anchor_link}()},
\code{\link{util_generate_anchor_tag}()},
\code{\link{util_generate_calls}()},
\code{\link{util_generate_calls_for_function}()},
\code{\link{util_load_manual}()},
\code{\link{util_make_data_slot_from_table_slot}()},
\code{\link{util_set_size}()}
}
\concept{reporting_functions}
\concept{summary}
\keyword{internal}
