% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_paste_with_na.R
\name{util_paste_with_na}
\alias{util_paste_with_na}
\title{Paste strings but keep NA}
\usage{
util_paste_with_na(...)
}
\arguments{
\item{...}{other arguments passed to \code{paste}}
}
\value{
\link{character} pasted strings
}
\description{
Paste strings but keep NA
}
\keyword{internal}
