% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot_figure_plotly.R
\name{util_plot_figure_plotly}
\alias{util_plot_figure_plotly}
\title{Plot a \code{ggplot2} figure using \code{plotly}}
\usage{
util_plot_figure_plotly(x)
}
\arguments{
\item{x}{\link{ggplot2} object}
}
\value{
\code{htmltools} compatible object
}
\description{
Plot a \code{ggplot2} figure using \code{plotly}
}
\keyword{internal}
