% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot2svg_object.R
\name{util_plotly2svg_object}
\alias{util_plotly2svg_object}
\title{\code{Plotly} to un-disclosed \code{ggplot} object}
\usage{
util_plotly2svg_object(plotly, w = 21.2, h = 15.9, sizing_hints)
}
\arguments{
\item{plotly}{the object}

\item{w}{width in cm}

\item{h}{height in cm}
}
\value{
\code{ggplot} object, but rendered (no original data included)
}
\description{
\code{Plotly} to un-disclosed \code{ggplot} object
}
\keyword{internal}
