% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rbind.R
\name{util_rbind}
\alias{util_rbind}
\title{Bind data frames row-based}
\usage{
util_rbind(..., data_frames_list = list())
}
\arguments{
\item{...}{\link{data.frame} none more more data frames}

\item{data_frames_list}{\link{list} optional, a list of data frames}
}
\value{
\link{data.frame} all data frames appended
}
\description{
if not all data frames share  all columns, missing columns will be filled with
\code{NA}s.
}
\examples{
\dontrun{
util_rbind(head(cars), head(iris))
util_rbind(head(cars), tail(cars))
util_rbind(head(cars)[, "dist", FALSE], tail(cars)[, "speed", FALSE])
}

}
\seealso{
Other data_management: 
\code{\link{util_assign_levlabs}()},
\code{\link{util_check_data_type}()},
\code{\link{util_check_group_levels}()},
\code{\link{util_compare_meta_with_study}()},
\code{\link{util_dichotomize}()},
\code{\link{util_merge_data_frame_list}()},
\code{\link{util_remove_na_records}()},
\code{\link{util_replace_hard_limit_violations}()},
\code{\link{util_study_var2factor}()},
\code{\link{util_table_of_vct}()}
}
\concept{data_management}
\concept{process}
\keyword{internal}
