% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_really_rstudio.R
\name{util_really_rstudio}
\alias{util_really_rstudio}
\title{Can we really be sure to run \code{RStudio}}
\usage{
util_really_rstudio()
}
\value{
\code{TRUE}, if really sure to be \code{RStudio}, \code{FALSE}, otherwise.
}
\description{
\verb{Jetbrain's} \code{Idea} and so on fake to be \code{RStudio} by having \code{RStudio} in
\code{.Platform$GUI}.
}
\keyword{internal}
