% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_remove_empty_rows.R
\name{util_remove_empty_rows}
\alias{util_remove_empty_rows}
\title{removes empty rows from \code{x}}
\usage{
util_remove_empty_rows(x, id_vars = character(0))
}
\arguments{
\item{x}{\link{data.frame} a data frame to be cleaned}

\item{id_vars}{\link{character} column names, that will be treated as empty}
}
\value{
\link{data.frame} reduced \code{x}
}
\description{
removes empty rows from \code{x}
}
