% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_render_table_dataquieR_summary.R
\name{util_render_table_dataquieR_summary}
\alias{util_render_table_dataquieR_summary}
\title{Render a table summarizing dataquieR results}
\usage{
util_render_table_dataquieR_summary(
  x,
  grouped_by = c("call_names", "indicator_metric")
)
}
\arguments{
\item{x}{a report summary (\code{summary(r)})}

\item{grouped_by}{define the columns of the resulting matrix. It can be either
"call_names", one column per function,  or "indicator_metric", one column per indicator
or both c("call_names", "indicator_metric"). The last combination is the default}
}
\value{
something, \code{htmltools} can render
}
\description{
Render a table summarizing dataquieR results
}
\keyword{internal}
