% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_replace_hard_limit_violations.R
\name{util_replace_hard_limit_violations}
\alias{util_replace_hard_limit_violations}
\title{Replace limit violations (HARD_LIMITS) by NAs}
\usage{
util_replace_hard_limit_violations(study_data, meta_data, label_col)
}
\arguments{
\item{study_data}{\link{study_data}}

\item{meta_data}{\link{study_data}}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}
}
\value{
modified \code{study_data}
}
\description{
Replace limit violations (HARD_LIMITS) by NAs
}
