% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rio_import.R
\name{util_rio_import}
\alias{util_rio_import}
\title{Import a data frame}
\usage{
util_rio_import(fn, keep_types, ...)
}
\arguments{
\item{fn}{the file name to load.}

\item{keep_types}{\link{logical} keep types as possibly defined in the file.
set \code{TRUE} for study data.}

\item{...}{additional arguments for \link[rio:import]{rio::import}}
}
\value{
\link{data.frame} as in \link[rio:import]{rio::import}
}
\description{
see \code{\link[rio:import]{rio::import}}, but with argument \code{keep_types} and modified error
handling.
}
\keyword{internal}
