% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rio_import.R
\name{util_rio_import_list}
\alias{util_rio_import_list}
\title{Import list of data frames}
\usage{
util_rio_import_list(fn, keep_types, ...)
}
\arguments{
\item{fn}{the file name to load.}

\item{keep_types}{\link{logical} keep types as possibly defined in the file.
set \code{TRUE} for study data.}

\item{...}{additional arguments for \link[rio:import_list]{rio::import_list}}
}
\value{
\link{list} as in \link[rio:import_list]{rio::import_list}
}
\description{
see \code{\link[rio:import_list]{rio::import_list}}, but with argument \code{keep_types} and modified error
handling.
}
\keyword{internal}
