% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_round_to_decimal_places.R
\name{util_round_to_decimal_places}
\alias{util_round_to_decimal_places}
\title{Round number of decimal places to 3 if the values are between 0.001 and 9999.999
otherwise (if at least one value of the vector is outside this limits)
use scientific notation for all the values in a vector}
\usage{
util_round_to_decimal_places(x, digits = 3)
}
\arguments{
\item{x}{a numeric vector to be rounded}

\item{digits}{a numeric value indicating the number of desired decimal places}
}
\description{
Round number of decimal places to 3 if the values are between 0.001 and 9999.999
otherwise (if at least one value of the vector is outside this limits)
use scientific notation for all the values in a vector
}
\seealso{
Other data_management: 
\code{\link{util_assign_levlabs}()},
\code{\link{util_check_data_type}()},
\code{\link{util_check_group_levels}()},
\code{\link{util_compare_meta_with_study}()},
\code{\link{util_dichotomize}()},
\code{\link{util_fix_merge_dups}()},
\code{\link{util_merge_data_frame_list}()},
\code{\link{util_rbind}()},
\code{\link{util_remove_na_records}()},
\code{\link{util_replace_hard_limit_violations}()},
\code{\link{util_study_var2factor}()},
\code{\link{util_table_of_vct}()}
}
\concept{data_management}
\keyword{internal}
