% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_set_size.R
\name{util_set_size}
\alias{util_set_size}
\title{Attaches attributes about the recommended minimum
absolute sizes to the plot p}
\usage{
util_set_size(p, width_em = NA_integer_, height_em = NA_integer_)
}
\arguments{
\item{p}{\link{ggplot} the plot}

\item{width_em}{\link{numeric} len=1. the minimum width hint in \code{em}}

\item{height_em}{\link{numeric} len=1. the minimum height in \code{em}}
}
\value{
p the modified plot
}
\description{
Attaches attributes about the recommended minimum
absolute sizes to the plot p
}
