% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_set_singlequotes.R
\name{util_set_sQuoteString}
\alias{util_set_sQuoteString}
\title{Utility function single quote string}
\usage{
util_set_sQuoteString(string)
}
\arguments{
\item{string}{Character vector}
}
\value{
quoted string
}
\description{
This function generates usual single-quotes for each element of the character
vector.
}
