% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sigmagap.R
\name{util_sigmagap}
\alias{util_sigmagap}
\title{Utility function outliers according to the rule of Huber et al.}
\usage{
util_sigmagap(x)
}
\arguments{
\item{x}{\link{numeric} data to check for outliers}
}
\value{
binary vector
}
\description{
This function calculates outliers according to the rule of Huber et al. T
his function requires the package \code{robustbase}.
}
