% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sixsigma.R
\name{util_sixsigma}
\alias{util_sixsigma}
\title{Utility function for six sigma deviations rule}
\usage{
util_sixsigma(x)
}
\arguments{
\item{x}{\link{numeric} data to check for outliers}
}
\value{
binary vector
}
\description{
This function calculates outliers according to the rule of six sigma
deviations.
}
