% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_split_val_tab.R
\name{util_split_val_tab}
\alias{util_split_val_tab}
\title{Split table with mixed code/missing lists to single tables}
\usage{
util_split_val_tab(val_tab = CODE_LIST_TABLE)
}
\arguments{
\item{val_tab}{\link{data.frame} tables in one long data frame.}
}
\value{
\code{invisible(NULL)}
}
\description{
resulting tables are populated to the data frame cache.
}
\keyword{internal}
