% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_standardise_ordinal_codes.R
\name{util_standardise_ordinal_codes}
\alias{util_standardise_ordinal_codes}
\title{Compute something comparable from an \code{ordered}}
\usage{
util_standardise_ordinal_codes(codes, maxlevel_old, maxlevel_new)
}
\arguments{
\item{codes}{\code{\link[=numeric]{numeric()}} \code{n} values}

\item{maxlevel_old}{\code{\link[=integer]{integer()}} number of categories of \code{codes}}

\item{maxlevel_new}{\code{\link[=integer]{integer()}} number of categories for output}
}
\value{
\code{\link[=integer]{integer()}} \code{n} values in \verb{\{1, ..., maxlevel_new\}}
}
\description{
interpolates categories of an ordinal variable
}
\keyword{internal}
