% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_startsWith_prefix._or_equals_prefix.R
\name{util_startsWith_prefix._or_equals_prefix}
\alias{util_startsWith_prefix._or_equals_prefix}
\title{String check for results/combined results}
\usage{
util_startsWith_prefix._or_equals_prefix(x, prefix)
}
\arguments{
\item{x}{\link{character} haystack}

\item{prefix}{\link{character} needle}
}
\value{
\link{logical} if entries in x start with prefix-DOT/equal to prefix
}
\description{
detect, if \code{x} starts with \verb{<prefix>.} or equals \verb{<prefix>},
if results have been combined
}
\keyword{internal}
