% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stop_if_not.R
\name{util_stop_if_not}
\alias{util_stop_if_not}
\title{Verify assumptions made by the code, that must be TRUE}
\usage{
util_stop_if_not(..., label, label_only)
}
\arguments{
\item{...}{see \code{\link{stopifnot}}}

\item{label}{\link{character} a label for the assumptions, can be missing}

\item{label_only}{\link{logical} if \code{TRUE} and \code{label} is given, the condition
will not be displayed, if \code{FALSE}}
}
\value{
\code{invisible(FALSE)}, if not stopped.
}
\description{
Verify assumptions made by the code, that must be TRUE
}
\seealso{
\code{\link{stopifnot}}
}
