% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_storr_factory.R
\name{util_storr_factory}
\alias{util_storr_factory}
\title{Create a \code{storr} object with a \code{storr_factory} attribute}
\usage{
util_storr_factory(my_storr_object, my_storr_factory)
}
\arguments{
\item{my_storr_object}{a \code{storr}-object}

\item{my_storr_factory}{a function creating the/a \code{storr_object}}
}
\value{
\code{storr}-object with the factory attribute and (hopefully) valid.
}
\description{
also does basic validity checks
}
\keyword{internal}
