% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_storr_factory.R
\name{util_storr_object}
\alias{util_storr_object}
\title{Create a \code{storr}-object using the factory}
\usage{
util_storr_object(
  my_storr_factory = function() {
     storr::storr_environment()
 }
)
}
\arguments{
\item{my_storr_factory}{a function returning a \code{storr} object}
}
\value{
a \code{storr} object
}
\description{
also performs checks.
}
\keyword{internal}
