% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_string_is_not_categorical.R
\name{util_string_is_not_categorical}
\alias{util_string_is_not_categorical}
\title{Utility function for judging whether a character vector does not appear to
be a categorical variable}
\usage{
util_string_is_not_categorical(vec)
}
\arguments{
\item{vec}{a character vector}
}
\value{
TRUE or FALSE
}
\description{
The function considers the following properties:
\itemize{
\item the maximum number of characters (to identify free text fields with
long entries),
\item the relative frequency of punctuation and space characters per element (to
identify, e.g., JSON or XML elements, which are structured by those
characters),
\item the relative frequency of elements (categorical variables would have
a low proportion of unique values in comparison to other variables).
}
}
\keyword{internal}
