% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_suppress_output.R
\name{util_suppress_output}
\alias{util_suppress_output}
\title{Suppress any output to \code{stdout} using \code{\link[=sink]{sink()}}}
\usage{
util_suppress_output(expr)
}
\arguments{
\item{expr}{\code{\link[=expression]{expression()}}}
}
\value{
\code{\link[=invisible]{invisible()}} result of \code{expr}
}
\description{
Suppress any output to \code{stdout} using \code{\link[=sink]{sink()}}
}
\seealso{
Other process_functions: 
\code{\link{util_abbreviate}()},
\code{\link{util_all_is_integer}()},
\code{\link{util_attach_attr}()},
\code{\link{util_bQuote}()},
\code{\link{util_backtickQuote}()},
\code{\link{util_coord_flip}()},
\code{\link{util_extract_matches}()},
\code{\link{util_par_pmap}()},
\code{\link{util_setup_rstudio_job}()}
}
\concept{process_functions}
\keyword{internal}
