% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_suppress_warnings.R
\name{util_suppress_warnings}
\alias{util_suppress_warnings}
\title{Suppress warnings conditionally}
\usage{
util_suppress_warnings(expr, classes = "warning")
}
\arguments{
\item{expr}{\link{expression} to evaluate}

\item{classes}{\link{character} classes of warning-conditions to suppress}
}
\value{
the result of \code{expr}
}
\description{
Suppress warnings conditionally
}
\seealso{
Other condition_functions: 
\code{\link{util_condition_constructor_factory}()},
\code{\link{util_deparse1}()},
\code{\link{util_error}()},
\code{\link{util_find_external_functions_in_stacktrace}()},
\code{\link{util_find_first_externally_called_functions_in_stacktrace}()},
\code{\link{util_find_indicator_function_in_callers}()},
\code{\link{util_message}()},
\code{\link{util_warning}()}
}
\concept{condition_functions}
\concept{process}
\keyword{internal}
