% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_table_of_vct.R
\name{util_table_of_vct}
\alias{util_table_of_vct}
\title{Tabulate a vector}
\usage{
util_table_of_vct(Var1)
}
\arguments{
\item{Var1}{vector to ta tabulate}
}
\value{
a data frame with columns \code{Var1} and \code{Freq}
}
\description{
does the same as \code{as.data.frame(table(x))} but
guarantees a data frame with two columns is returned
}
