% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_table_rotator.R
\name{util_table_rotator}
\alias{util_table_rotator}
\title{Rotate 1-row data frames to key-value data frames}
\usage{
util_table_rotator(tb)
}
\arguments{
\item{tb}{\link{data.frame} a data frame}
}
\value{
\link{data.frame} but transposed
}
\description{
if \code{nrow(tb) > 1}, \code{util_table_rotator} just returns \code{tb}.
}
