% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_translate.R
\name{util_translate}
\alias{util_translate}
\title{Get a translation}
\usage{
util_translate(keys, ns = "general", lang = getOption("dataquieR.lang", ""))
}
\arguments{
\item{keys}{\link{character} translation keys}

\item{ns}{\link{character} translation namespace}

\item{lang}{\link{character} language to translate to}
}
\value{
\link{character} translations
}
\description{
Get a translation
}
\seealso{
Other string_functions: 
\code{\link{util_abbreviate_unique}()},
\code{\link{util_filter_names_by_regexps}()},
\code{\link{util_pretty_vector_string}()},
\code{\link{util_set_dQuoteString}()},
\code{\link{util_set_sQuoteString}()},
\code{\link{util_sub_string_left_from_.}()},
\code{\link{util_sub_string_right_from_.}()}
}
\concept{reporting}
\concept{string_functions}
\keyword{internal}
