% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_translate_indicator_metrics.R
\name{util_translate_indicator_metrics}
\alias{util_translate_indicator_metrics}
\title{Translate standard column names to readable ones}
\usage{
util_translate_indicator_metrics(colnames, short = FALSE)
}
\arguments{
\item{colnames}{\link{character} the names to translate}

\item{short}{\link{logical} include unit in output}
}
\value{
translated names
}
\description{
Translate standard column names to readable ones
}
