% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tukey.R
\name{util_tukey}
\alias{util_tukey}
\title{Utility function Tukey outlier rule}
\usage{
util_tukey(x)
}
\arguments{
\item{x}{\link{numeric} data to check for outliers}
}
\value{
binary vector
}
\description{
This function calculates outliers according to the rule of Tukey.
}
\seealso{
Other outlier_functions: 
\code{\link{util_3SD}()},
\code{\link{util_hubert}()},
\code{\link{util_sigmagap}()}
}
\concept{outlier}
\concept{outlier_functions}
\keyword{internal}
