% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_undisclose.R
\name{util_undisclose}
\alias{util_undisclose}
\title{Remove tables referred to by metadata and use \code{SVG} for most figures}
\usage{
util_undisclose(x, ...)
}
\arguments{
\item{x}{an object to un-disclose}

\item{...}{further arguments, used for pointing to the \code{dataquieR_result}
object, if called recursively}
}
\value{
undisclosed object
}
\description{
Remove tables referred to by metadata and use \code{SVG} for most figures
}
\keyword{internal}
