% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_user_hint.R
\name{util_user_hint}
\alias{util_user_hint}
\title{Save a hint to the user during package load}
\usage{
util_user_hint(x)
}
\arguments{
\item{x}{\link{character} the hint}
}
\value{
\code{invisible(NULL)}
}
\description{
Save a hint to the user during package load
}
