% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validate_known_meta.R
\name{util_validate_known_meta}
\alias{util_validate_known_meta}
\title{Utility function verifying syntax of known metadata columns}
\usage{
util_validate_known_meta(meta_data)
}
\arguments{
\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data}
}
\value{
invisible(NULL)
}
\description{
This function goes through metadata columns, \code{dataquieR} supports
and verifies for these, that they follow its metadata conventions.
}
