% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validate_report_summary_table.R
\name{util_validate_report_summary_table}
\alias{util_validate_report_summary_table}
\title{Verify the class \code{ReportSummaryTable}}
\usage{
util_validate_report_summary_table(tb, meta_data, label_col)
}
\arguments{
\item{tb}{\link{data.frame} object to be a \code{ReportSummaryTable}}

\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data. Used to translate
variable names, if given.}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}
}
\value{
\link{data.frame} maybe fixed \code{ReportSummaryTable}
}
\description{
Verify the class \code{ReportSummaryTable}
}
\keyword{internal}
