% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_variable_references.R
\name{util_variable_references}
\alias{util_variable_references}
\title{Find all columns in item-level-metadata, that refer to some other variable}
\usage{
util_variable_references(meta_data = "item_level")
}
\arguments{
\item{meta_data}{\link{data.frame} the metadata}
}
\value{
\link{character} all column names referring to variables from item-level
metadata
}
\description{
Find all columns in item-level-metadata, that refer to some other variable
}
\seealso{
Other lookup_functions: 
\code{\link{util_prep_location_check}()},
\code{\link{util_prep_proportion_check}()}
}
\concept{lookup_functions}
\concept{metadata_management}
\keyword{internal}
