% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_verify_names.R
\name{util_verify_names}
\alias{util_verify_names}
\title{Test for likely misspelled data frame references}
\usage{
util_verify_names(name_of_study_data = character(0))
}
\arguments{
\item{name_of_study_data}{\link{character} names of study data, such are expected}
}
\value{
\code{invisible(NULL)}, messages / warns only.
}
\description{
checks, if some data frame names may have typos in their names
}
\keyword{internal}
