% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_view_file.R
\name{util_view_file}
\alias{util_view_file}
\title{View a file in most suitable viewer}
\usage{
util_view_file(file)
}
\arguments{
\item{file}{the file to view}
}
\value{
\code{invisible(file)}
}
\description{
View a file in most suitable viewer
}
