% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_warn_unordered.R
\name{util_warn_unordered}
\alias{util_warn_unordered}
\title{Warn about a problem in \code{varname}, if \code{x} has no natural order}
\usage{
util_warn_unordered(x, varname)
}
\arguments{
\item{x}{\link{vector} of data}

\item{varname}{\link{character} len=1. Variable name for warning messages}
}
\value{
invisible(NULL)
}
\description{
Also warns, if R does not have a comparison operator for \code{x}.
}
