% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{VARATT_REQUIRE_LEVELS}
\alias{VARATT_REQUIRE_LEVELS}
\alias{COMPATIBILITY}
\alias{REQUIRED}
\alias{RECOMMENDED}
\alias{OPTIONAL}
\alias{TECHNICAL}
\alias{UNKNOWN}
\title{Requirement levels of certain metadata columns}
\format{
An object of class \code{list} of length 5.
}
\usage{
VARATT_REQUIRE_LEVELS
}
\description{
These levels are cumulatively used by the function \link{prep_create_meta} and
related in the argument \code{level} therein.
}
\details{
currently available:
\itemize{
\item 'COMPATIBILITY' = "compatibility"
\item 'REQUIRED' = "required"
\item 'RECOMMENDED' = "recommended"
\item 'OPTIONAL' = "optional"
\item 'TECHNICAL' = "technical"
}
}
\keyword{datasets}
