% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.dataquieR_resultset.R
\name{as.data.frame.dataquieR_resultset}
\alias{as.data.frame.dataquieR_resultset}
\title{Convert a full \code{dataquieR} report to a \code{data.frame}}
\usage{
\method{as.data.frame}{dataquieR_resultset}(x, ...)
}
\arguments{
\item{x}{\link[=dq_report]{dataquieR report}}

\item{...}{not used}
}
\value{
a \link{data.frame} with one row per indicator call, one column
\code{implementationform} naming the called indicator function, one column per
function argument and one additional column containing the results of each
call as a list.
}
\description{
converts a \link[=dq_report]{dataquieR report} to a \link{data.frame}. Intended for
use in pipelines.
}
