% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.dataquieR_resultset.R
\name{as.list.dataquieR_resultset}
\alias{as.list.dataquieR_resultset}
\title{Convert a full \code{dataquieR} report to a \code{list}}
\usage{
\method{as.list}{dataquieR_resultset}(x, ...)
}
\arguments{
\item{x}{Deprecated}

\item{...}{Deprecated}
}
\value{
Deprecated
}
\description{
Deprecated
}
