% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_qualified_segment_missingness.R
\name{com_qualified_segment_missingness}
\alias{com_qualified_segment_missingness}
\title{Compute Indicators for Qualified Segment Missingness}
\usage{
com_qualified_segment_missingness(
  study_data,
  meta_data,
  label_col = NULL,
  meta_data_segment,
  expected_observations = c("HIERARCHY", "ALL", "SEGMENT")
)
}
\arguments{
\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{meta_data_segment}{\link{data.frame} Segment level metadata}

\item{expected_observations}{\link{enum} HIERARCHY | ALL | SEGMENT. Report the
number of observations expected using
the old \code{PART_VAR} concept. See
\link{com_item_missingness} for an
explanation.}
}
\value{
\link{list} list with entries:
}
\description{
Compute Indicators for Qualified Segment Missingness
}
