% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{COMPUTATION_RULE}
\alias{COMPUTATION_RULE}
\title{Cross-item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
COMPUTATION_RULE
}
\description{
Cross-item level metadata attribute name
}
\seealso{
\link{meta_data_computation}
}
\concept{meta_data_computation}
\keyword{datasets}
