% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_contradiction_rules.R
\docType{data}
\name{contradiction_functions_descriptions}
\alias{contradiction_functions_descriptions}
\title{description of the contradiction functions}
\format{
An object of class \code{list} of length 11.
}
\usage{
contradiction_functions_descriptions
}
\description{
description of the contradiction functions
}
\keyword{datasets}
