% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{DATA_TYPES_OF_R_TYPE}
\alias{DATA_TYPES_OF_R_TYPE}
\title{All available data types, mapped from their respective
R types}
\format{
An object of class \code{list} of length 13.
}
\usage{
DATA_TYPES_OF_R_TYPE
}
\description{
All available data types, mapped from their respective
R types
}
\keyword{datasets}
