% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquieR-package.R
\docType{package}
\name{dataquieR}
\alias{dataquieR}
\alias{dataquieR-package}
\title{The \code{dataquieR} package about Data Quality in
Epidemiological Research}
\description{
For a quick start please read \link{dq_report} and maybe the vignettes
or the package's \href{https://dfg-qa.ship-med.uni-greifswald.de/}{website}.
}
