% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{DF_RECORD_CHECK}
\alias{DF_RECORD_CHECK}
\title{Data frame level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
DF_RECORD_CHECK
}
\description{
The type of check to be conducted when comparing the reference
ID table with the IDs delivered in the study data files.
}
\seealso{
\link{meta_data_dataframe}
}
\keyword{datasets}
