% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{dims}
\alias{dims}
\title{Dimension Titles for Prefixes}
\format{
An object of class \code{character} of length 4.
}
\usage{
dims
}
\description{
order does matter, because it defines the order in the \code{dq_report2}.
}
\seealso{
\code{\link[=util_html_for_var]{util_html_for_var()}}

\code{\link[=util_html_for_dims]{util_html_for_dims()}}
}
\keyword{datasets}
