% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{DISTRIBUTIONS}
\alias{DISTRIBUTIONS}
\title{All available probability distributions for
\link{acc_shape_or_scale}}
\format{
An object of class \code{list} of length 3.
}
\usage{
DISTRIBUTIONS
}
\description{
\itemize{
\item \code{uniform} For uniform distribution
\item \code{normal} For Gaussian distribution
\item \code{GAMMA} For a gamma distribution
}
}
\keyword{datasets}
