% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_meta_data_v1_to_item_level_meta_data.R
\name{.util_internal_normalize_meta_data}
\alias{.util_internal_normalize_meta_data}
\title{Make normalizations of v2.0 item_level metadata.}
\usage{
.util_internal_normalize_meta_data(
  meta_data = "item_level",
  label_col = LABEL,
  verbose = TRUE
)
}
\arguments{
\item{meta_data}{\link{data.frame} the old item-level-metadata}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}

\item{verbose}{\link{logical} display all estimated decisions, defaults to \code{TRUE},
except if called in a \link{dq_report2} pipeline.}
}
\description{
Requires referred missing-tables being available by
\code{prep_get_data_frame}.
}
