% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_unexp_records_dataframe.R
\name{int_unexp_records_dataframe}
\alias{int_unexp_records_dataframe}
\title{Check for unexpected data record count at the data frame level}
\usage{
int_unexp_records_dataframe(identifier_name_list, data_record_count)
}
\arguments{
\item{identifier_name_list}{\link{character} a character vector indicating the name of each study data frame, mandatory.}

\item{data_record_count}{\link{integer}  an integer vector with the number of expected data records per study data frame, mandatory.}
}
\value{
a \link{list} with
\itemize{
\item \code{DataframeData}: data frame with the results of the quality check for unexpected data elements
\item \code{DataframeTable}: data frame with selected unexpected data elements check results, used for the data quality report.
}
}
\description{
This function contrasts the expected record number in each study in
the metadata with the actual record number in each study data frame.
}
