% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\name{meta_data_cross}
\alias{meta_data_cross}
\title{Well known columns on the \code{meta_data_cross-item} sheet}
\description{
Metadata describing groups of variables, e.g., for their multivariate
distribution or for defining contradiction rules.
}
\seealso{
check_table
}
