% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dataquieR_summary.R
\name{plot.dataquieR_summary}
\alias{plot.dataquieR_summary}
\title{Plot a \code{dataquieR} summary}
\usage{
\method{plot}{dataquieR_summary}(x, y, ..., filter, dont_plot = FALSE, stratify_by)
}
\arguments{
\item{x}{the \code{dataquieR} summary, see \code{\link[=summary]{summary()}} and \code{\link[=dq_report2]{dq_report2()}}}

\item{y}{not yet used}

\item{...}{not yet used}

\item{filter}{if given, this filters the summary, e.g.,
\code{filter = call_names == "com_qualified_item_missingness"}}

\item{dont_plot}{suppress the actual plotting, just return a printable
object derived from \code{x}}

\item{stratify_by}{column to stratify the summary, may be one string.}
}
\value{
invisible html object
}
\description{
Plot a \code{dataquieR} summary
}
