% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_add_to_meta.R
\name{prep_add_to_meta}
\alias{prep_add_to_meta}
\title{Support function to augment metadata during data quality reporting}
\usage{
prep_add_to_meta(
  VAR_NAMES,
  DATA_TYPE,
  LABEL,
  VALUE_LABELS,
  meta_data = "item_level",
  ...
)
}
\arguments{
\item{VAR_NAMES}{\link{character} Names of the Variables to add}

\item{DATA_TYPE}{\link{character} Data type for the added variables}

\item{LABEL}{\link{character} Labels for these variables}

\item{VALUE_LABELS}{\link{character} Value labels for the values of the variables
as usually pipe separated and assigned with
\code{=}: \code{1 = male | 2 = female}}

\item{meta_data}{\link{data.frame} the metadata to extend}

\item{...}{Further defined variable attributes, see
\link{prep_create_meta}}
}
\value{
a data frame with amended meta data.
}
\description{
adds an annotation to static metadata
}
\details{
Add metadata e.g. of transformed/new variable
This function is not yet considered stable, but we already export it,
because it could help. Therefore, we have some inconsistencies in the
formals still.
}
