% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_combine_report_summaries.R
\name{prep_combine_report_summaries}
\alias{prep_combine_report_summaries}
\title{Combine two report summaries}
\usage{
prep_combine_report_summaries(..., summaries_list, amend_segment_names = FALSE)
}
\arguments{
\item{...}{objects returned by \link{prep_extract_summary}}

\item{summaries_list}{if given, \link{list} of objects returned by
\link{prep_extract_summary}}

\item{amend_segment_names}{\link{logical} use names of the \code{summaries_list} and
argument names as segment prefixes}
}
\value{
combined summaries
}
\description{
Combine two report summaries
}
\seealso{
Other summary_functions: 
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()}
}
\concept{summary_functions}
