% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dq_data_type_of.R
\name{prep_dq_data_type_of}
\alias{prep_dq_data_type_of}
\title{Get the dataquieR \code{DATA_TYPE} of \code{x}}
\usage{
prep_dq_data_type_of(x)
}
\arguments{
\item{x}{object to define the dataquieR data type of}
}
\value{
the dataquieR data type as listed in \code{DATA_TYPES}
}
\description{
Get the dataquieR \code{DATA_TYPE} of \code{x}
}
\seealso{
\code{\link{DATA_TYPES_OF_R_TYPE}}
}
