% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_extract_cause_label_df.R
\name{prep_extract_cause_label_df}
\alias{prep_extract_cause_label_df}
\title{Extract all missing/jump codes from metadata and export a
cause-label-data-frame}
\usage{
prep_extract_cause_label_df(meta_data = "item_level", label_col = VAR_NAMES)
}
\arguments{
\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}
}
\value{
\link{list} with the entries
\itemize{
\item \code{meta_data} \link{data.frame} a data frame that contains updated metadata
\item \code{cause_label_df} \link{data.frame} missing code table. If missing codes have
labels the respective data frame are
specified here, see \link{cause_label_df}
}
}
\description{
Extract all missing/jump codes from metadata and export a
cause-label-data-frame
}
\seealso{
\link{prep_add_cause_label_df}
}
