% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_extract_summary.R
\name{prep_extract_summary}
\alias{prep_extract_summary}
\title{Extract summary from data quality results}
\usage{
prep_extract_summary(r, ...)
}
\arguments{
\item{r}{\link{dq_report2} or \link{dataquieR_result} object}

\item{...}{further arguments, maybe needed for some implementations}
}
\value{
\link{list} with two slots \code{Data} and \code{Table} with \link{data.frame}s
featuring all metrics columns
from the report or result in \code{x},
the \link{STUDY_SEGMENT} and the \link{VAR_NAMES}.
In case of \code{Data}, the columns are formatted nicely but still
with the standardized column names -- use
\code{\link[=util_translate_indicator_metrics]{util_translate_indicator_metrics()}} to rename them nicely. In
case of \code{Table}, just as they are.
}
\description{
Generic function, currently supports \link{dq_report2} and \link{dataquieR_result}
}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()},
\code{\link{util_as_cat}()},
\code{\link{util_extract_indicator_metrics}()},
\code{\link{util_get_category_for_result}()},
\code{\link{util_get_colors}()},
\code{\link{util_get_html_cell_for_result}()},
\code{\link{util_get_labels_grading_class}()},
\code{\link{util_get_message_for_result}()},
\code{\link{util_get_rule_sets}()},
\code{\link{util_get_ruleset_formats}()},
\code{\link{util_get_thresholds}()},
\code{\link{util_html_table}()},
\code{\link{util_melt_summary}()},
\code{\link{util_sort_by_order}()}
}
\concept{summary_functions}
