% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scalelevel_from_data_and_metadata.R
\name{prep_scalelevel_from_data_and_metadata}
\alias{prep_scalelevel_from_data_and_metadata}
\title{Heuristics to amend a SCALE_LEVEL column and a UNIT column in the metadata}
\usage{
prep_scalelevel_from_data_and_metadata(
  resp_vars = NULL,
  study_data,
  meta_data = "item_level",
  label_col = LABEL
)
}
\arguments{
\item{resp_vars}{\link{variable list} the names of the measurement variables}

\item{study_data}{\link{data.frame} the data frame that contains the measurements}

\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}
}
\value{
\link{data.frame} modified metadata
}
\description{
...if missing
}
\examples{
\dontrun{
  prep_load_workbook_like_file("meta_data_v2")
  prep_scalelevel_from_data_and_metadata(study_data = "study_data")
}
}
