% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_result.R
\name{print.dataquieR_result}
\alias{print.dataquieR_result}
\alias{dataquieR_result}
\title{Print a \link{dataquieR} result returned by \link{dq_report2}}
\usage{
\method{print}{dataquieR_result}(x, ...)
}
\arguments{
\item{x}{\link{list} a dataquieR result from \link{dq_report2} or
\code{util_eval_to_dataquieR_result}}

\item{...}{passed to print. Additionally, the argument \code{slot} may be passed
to print only specific sub-results.}
}
\value{
see print
}
\description{
Print a \link{dataquieR} result returned by \link{dq_report2}
}
\seealso{
\code{util_pretty_print()}
}
