% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_result.R
\name{print.dataquieR_result}
\alias{print.dataquieR_result}
\alias{dataquieR_result}
\title{Print a \link{dataquieR} result returned by pipeline_vectorized}
\usage{
\method{print}{dataquieR_result}(x, ...)
}
\arguments{
\item{x}{\link{list} a dataquieR result from \link{pipeline_vectorized} or
\link{util_eval_to_dataquieR_result}}

\item{...}{passed to print. Additionally, the argument \code{slot} may be passed
to print only specific sub-results.}
}
\value{
see print
}
\description{
Print a \link{dataquieR} result returned by pipeline_vectorized
}
\seealso{
\code{\link[=util_pretty_print]{util_pretty_print()}}
}
