% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_resultset.R
\name{print.dataquieR_resultset}
\alias{print.dataquieR_resultset}
\title{Generate a RMarkdown-based report from a \link{dataquieR} report}
\usage{
\method{print}{dataquieR_resultset}(x, ...)
}
\arguments{
\item{x}{\link[=dq_report]{dataquieR report}.}

\item{...}{additional arguments:
\itemize{
\item template: Report template to use, not yet supported.
\item chunk_error: display error messages in report
\item chunk_warning: display warnings in report
\item output_format: output format to use, see \link[rmarkdown:render]{rmarkdown::render} -- currently,
html based formats are supported by the default template.
\item chunk_echo: display R code in report
\item chunk_message: display \link{message} outputs in report
}}
}
\value{
file name of the generated report
}
\description{
Generate a RMarkdown-based report from a \link{dataquieR} report
}
