% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_resultset2.R
\name{print.dataquieR_resultset2}
\alias{print.dataquieR_resultset2}
\title{Generate a RMarkdown-based report from a \link{dataquieR} report}
\usage{
\method{print}{dataquieR_resultset2}(x, dir, view = TRUE, disable_plotly = FALSE, ...)
}
\arguments{
\item{x}{\link[=dq_report2]{dataquieR report v2}.}

\item{dir}{\link{character} directory to store the rendered report's files,
a temporary one,
if omitted. Directory will be created, if missing, files may be
overwritten inside that directory}

\item{view}{\link{logical} display the report}

\item{disable_plotly}{\link{logical} do not use \code{plotly}, even if installed}

\item{...}{additional arguments:}
}
\value{
file names of the generated report's HTML files
}
\description{
Generate a RMarkdown-based report from a \link{dataquieR} report
}
