% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataquieR_result.R
\name{print.master_result}
\alias{print.master_result}
\title{Print a \code{master_result} object}
\usage{
\method{print}{master_result}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{not used}
}
\value{
\code{invisible(NULL)}
}
\description{
Print a \code{master_result} object
}
