% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{SEGMENT_ID_TABLE}
\alias{SEGMENT_ID_TABLE}
\title{Segment level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
SEGMENT_ID_TABLE
}
\description{
The name of the data frame containing the reference IDs to be compared
with the IDs in the targeted segment.
}
\seealso{
\link{meta_data_segment}
}
\keyword{datasets}
