% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{SEGMENT_MISS}
\alias{SEGMENT_MISS}
\title{Segment level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
SEGMENT_MISS
}
\description{
\code{true} or \code{false} to suppress crude segment missingness output
(\verb{Completeness/Misg. Segments} in the report). Defaults to compute
the output, if more than one segment is available in the item-level
metadata.
}
\seealso{
\link{meta_data_segment}
}
\keyword{datasets}
