% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{SEGMENT_PART_VARS}
\alias{SEGMENT_PART_VARS}
\title{Segment level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
SEGMENT_PART_VARS
}
\description{
The name of the segment participation status variable
}
\seealso{
\link{meta_data_segment}
}
\keyword{datasets}
