% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{SEGMENT_RECORD_CHECK}
\alias{SEGMENT_RECORD_CHECK}
\title{Segment level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
SEGMENT_RECORD_CHECK
}
\description{
The type of check to be conducted when comparing the reference ID
table with the IDs in a segment.
}
\seealso{
\link{meta_data_segment}
}
\keyword{datasets}
