% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{SEGMENT_RECORD_COUNT}
\alias{SEGMENT_RECORD_COUNT}
\title{Segment level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
SEGMENT_RECORD_COUNT
}
\description{
Number of expected data records in each segment.
\link{numeric}. Check only conducted if number entered
}
\seealso{
\link{meta_data_segment}
}
\keyword{datasets}
