% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{SEGMENT_UNIQUE_ROWS}
\alias{SEGMENT_UNIQUE_ROWS}
\title{Segment level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
SEGMENT_UNIQUE_ROWS
}
\description{
Specifies whether identical data is permitted across rows in a
segment (excluding ID variables)
}
\seealso{
\link{meta_data_segment}
}
\keyword{datasets}
