% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{SPLIT_CHAR}
\alias{SPLIT_CHAR}
\title{Character used  by default as a separator in meta data such as
missing codes}
\format{
An object of class \code{character} of length 1.
}
\usage{
SPLIT_CHAR
}
\description{
This 1 character is according to our metadata concept “|”.
}
\keyword{datasets}
