% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_abbreviate.R
\name{util_abbreviate}
\alias{util_abbreviate}
\title{Abbreviate snake_case function names to shortened \code{CamelCase}}
\usage{
util_abbreviate(x)
}
\arguments{
\item{x}{a vector of indicator function names}
}
\value{
abbreviations
}
\description{
Abbreviate snake_case function names to shortened \code{CamelCase}
}
\seealso{
\link[base:abbreviate]{base::abbreviate}

Other process_functions: 
\code{\link{util_all_is_integer}()},
\code{\link{util_attach_attr}()},
\code{\link{util_bQuote}()},
\code{\link{util_backtickQuote}()},
\code{\link{util_coord_flip}()},
\code{\link{util_count_NA}()},
\code{\link{util_extract_matches}()},
\code{\link{util_par_pmap}()},
\code{\link{util_setup_rstudio_job}()},
\code{\link{util_suppress_output}()}
}
\concept{process_functions}
\concept{reporting}
\keyword{internal}
