% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_alias2caption.R
\name{util_alias2caption}
\alias{util_alias2caption}
\title{Create a caption from an alias name of a \code{dq_report2} result}
\usage{
util_alias2caption(alias, long = FALSE)
}
\arguments{
\item{alias}{alias name}

\item{long}{not for menu entry}
}
\value{
caption
}
\description{
Create a caption from an alias name of a \code{dq_report2} result
}
\seealso{
\link{util_html_table}

Other reporting_functions: 
\code{\link{util_copy_all_deps}()},
\code{\link{util_create_page_file}()},
\code{\link{util_eval_to_dataquieR_result}()},
\code{\link{util_evaluate_calls}()},
\code{\link{util_float_index_menu}()},
\code{\link{util_generate_anchor_link}()},
\code{\link{util_generate_anchor_tag}()},
\code{\link{util_generate_calls}()},
\code{\link{util_generate_calls_for_function}()},
\code{\link{util_load_manual}()},
\code{\link{util_make_data_slot_from_table_slot}()},
\code{\link{util_make_function}()},
\code{\link{util_order_by_order}()},
\code{\link{util_set_size}()}
}
\concept{html}
\concept{reporting_functions}
\keyword{internal}
