% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_all_is_integer.R
\name{util_all_is_integer}
\alias{util_all_is_integer}
\title{convenience function to abbreviate all(util_is_integer(...))}
\usage{
util_all_is_integer(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
\code{TRUE}, if all entries are integer-like, \code{FALSE} otherwise
}
\description{
convenience function to abbreviate all(util_is_integer(...))
}
\seealso{
\link{util_is_integer}

Other process_functions: 
\code{\link{util_abbreviate}()},
\code{\link{util_attach_attr}()},
\code{\link{util_bQuote}()},
\code{\link{util_backtickQuote}()},
\code{\link{util_coord_flip}()},
\code{\link{util_extract_matches}()},
\code{\link{util_par_pmap}()},
\code{\link{util_setup_rstudio_job}()},
\code{\link{util_suppress_output}()}
}
\concept{process}
\concept{process_functions}
\keyword{internal}
