% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_attach_attr.R
\name{util_attach_attr}
\alias{util_attach_attr}
\title{Attach attributes to an object and return it}
\usage{
util_attach_attr(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{named arguments, each becomes an attributes}
}
\value{
\code{x}, having the desired attributes attached
}
\description{
Attach attributes to an object and return it
}
