% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_backtickQuote.R
\name{util_backtickQuote}
\alias{util_backtickQuote}
\title{utility function to set string in backticks}
\usage{
util_backtickQuote(x)
}
\arguments{
\item{x}{variable names}
}
\value{
quoted variable names
}
\description{
Quote a set of variable names with backticks
}
