% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_one_unique_value.R
\name{util_check_one_unique_value}
\alias{util_check_one_unique_value}
\title{Check for one value only}
\usage{
util_check_one_unique_value(x)
}
\arguments{
\item{x}{vector with values}
}
\value{
logical(1): TRUE, if -- except NA -- exactly only one value
is observed in \code{x},
FALSE otherwise
}
\description{
utility function to identify variables with one value only.
}
\seealso{
Other robustness_functions: 
\code{\link{util_as_valid_missing_codes}()},
\code{\link{util_correct_variable_use}()},
\code{\link{util_empty}()},
\code{\link{util_ensure_character}()},
\code{\link{util_ensure_in}()},
\code{\link{util_ensure_suggested}()},
\code{\link{util_expect_scalar}()},
\code{\link{util_fix_rstudio_bugs}()},
\code{\link{util_is_integer}()},
\code{\link{util_is_numeric_in}()},
\code{\link{util_is_valid_missing_codes}()},
\code{\link{util_match_arg}()},
\code{\link{util_observations_in_subgroups}()},
\code{\link{util_only_NAs}()},
\code{\link{util_stop_if_not}()},
\code{\link{util_warn_unordered}()}
}
\concept{data_management}
\concept{robustness_functions}
\keyword{internal}
