% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_condition_constructor_factory.R
\name{util_condition_constructor_factory}
\alias{util_condition_constructor_factory}
\title{Produce a condition function}
\usage{
util_condition_constructor_factory(
  .condition_type = c("error", "warning", "message")
)
}
\arguments{
\item{.condition_type}{\link{character} the type of the conditions being created
and signaled by the function, "error", "warning", or "message"}
}
\description{
Produce a condition function
}
