% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_copy_all_deps.R
\name{util_copy_all_deps}
\alias{util_copy_all_deps}
\title{Copy default dependencies to the report's lib directory}
\usage{
util_copy_all_deps(dir, pages, ...)
}
\arguments{
\item{dir}{report directory}

\item{pages}{all pages to write}

\item{...}{additional \code{htmltools::htmlDependency} objects to be added to all
pages, also}
}
\value{
\code{invisible(NULL)}
}
\description{
Copy default dependencies to the report's lib directory
}
