% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_count_NA.R
\name{util_count_NA}
\alias{util_count_NA}
\title{Support function to count number of \code{NA}s}
\usage{
util_count_NA(x)
}
\arguments{
\item{x}{object to count \code{NA}s in}
}
\value{
number of \code{NA}s
}
\description{
Counts the number of \code{NA}s in x.
}
