% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_create_page_file.R
\name{util_create_page_file}
\alias{util_create_page_file}
\title{Create an HTML file for the \link{dq_report2}}
\usage{
util_create_page_file(
  page,
  pages,
  rendered_pages,
  dir,
  template_file,
  report,
  logo,
  loading,
  packageName,
  deps
)
}
\arguments{
\item{page}{the name of the page-file being created}

\item{pages}{list with all page-contents named by their desired file names}

\item{rendered_pages}{list with all rendered (\code{htmltools::renderTags}) page-contents named by their desired file names}

\item{dir}{target directory}

\item{template_file}{the report template file to use}

\item{report}{the output of \link{dq_report2}}

\item{logo}{logo \code{PNG} file}

\item{loading}{loading animation div}

\item{packageName}{the name of the current package}

\item{deps}{dependencies, as pre-processed by
\code{htmltools::copyDependencyToDir} and
\code{htmltools::renderDependencies}}
}
\value{
\code{invisible(file_name)}
}
\description{
Create an HTML file for the \link{dq_report2}
}
