% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_detect_cores.R
\name{util_detect_cores}
\alias{util_detect_cores}
\title{Detect cores}
\usage{
util_detect_cores()
}
\value{
number of available CPU cores.
}
\description{
See \code{parallel::detectCores} for further details.
}
