% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_empty.R
\name{util_empty}
\alias{util_empty}
\title{Test, if values of x are empty, i.e. NA or whitespace characters}
\usage{
util_empty(x)
}
\arguments{
\item{x}{the vector to test}
}
\value{
a logical vector, same length as x; TRUE, if resp. element in x is
"empty"
}
\description{
Test, if values of x are empty, i.e. NA or whitespace characters
}
\seealso{
Other robustness_functions: 
\code{\link{util_as_valid_missing_codes}()},
\code{\link{util_check_one_unique_value}()},
\code{\link{util_correct_variable_use}()},
\code{\link{util_ensure_character}()},
\code{\link{util_ensure_in}()},
\code{\link{util_ensure_suggested}()},
\code{\link{util_expect_scalar}()},
\code{\link{util_fix_rstudio_bugs}()},
\code{\link{util_is_integer}()},
\code{\link{util_is_numeric_in}()},
\code{\link{util_is_valid_missing_codes}()},
\code{\link{util_match_arg}()},
\code{\link{util_observations_in_subgroups}()},
\code{\link{util_stop_if_not}()},
\code{\link{util_warn_unordered}()}
}
\concept{process}
\concept{robustness_functions}
\keyword{internal}
