% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ensure_suggested.R
\name{util_ensure_suggested}
\alias{util_ensure_suggested}
\title{Support function to stop, if an optional package is not installed}
\usage{
util_ensure_suggested(pkg, goal)
}
\arguments{
\item{pkg}{needed package}

\item{goal}{feature description for error message.}
}
\description{
This function stops, if a package is not installed but needed for using an
optional feature of \code{dataquieR}.
}
