% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_eval_to_dataquieR_result.R
\name{util_eval_to_dataquieR_result}
\alias{util_eval_to_dataquieR_result}
\title{Evaluate an expression and create a \code{dataquieR_result} object from
it's evaluated value}
\usage{
util_eval_to_dataquieR_result(
  expression,
  env = parent.frame(),
  filter_result_slots
)
}
\arguments{
\item{expression}{the expression}

\item{env}{the environment to evaluate the expression in}

\item{filter_result_slots}{\link{character} regular expressions, only
if an indicator function's
result's name
matches one of these, it'll
be used for the report. If
of length zero, no filtering
is performed.}
}
\value{
a \code{dataquieR_result} object
}
\description{
if an error occurs, the function will return a corresponding object
representing that error. all conditions will be recorded and replayed,
whenever the result is printed by \link{print.dataquieR_result}.
}
