% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_filter_names_by_regexps.R
\name{util_filter_names_by_regexps}
\alias{util_filter_names_by_regexps}
\title{Filter collection based on its \code{names()} using regular expressions}
\usage{
util_filter_names_by_regexps(collection, regexps)
}
\arguments{
\item{collection}{a named collection (list, vector, ...)}

\item{regexps}{\link{character} a vector of regular expressions}
}
\value{
\code{collection} reduced to entries, that's names match at least any
expression from \code{regexps}
}
\description{
Filter collection based on its \code{names()} using regular expressions
}
\examples{
\dontrun{ # internal function
util_filter_names_by_regexps(iris, c("epa", "eta"))
}

}
\seealso{
Other string_functions: 
\code{\link{util_pretty_vector_string}()},
\code{\link{util_set_dQuoteString}()},
\code{\link{util_set_sQuoteString}()},
\code{\link{util_sub_string_left_from_.}()},
\code{\link{util_sub_string_right_from_.}()}
}
\concept{process}
\concept{string_functions}
\keyword{internal}
