% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fix_rstudio_bugs.R
\name{util_fix_rstudio_bugs}
\alias{util_fix_rstudio_bugs}
\title{RStudio crashes on parallel calls in some versions on Darwin based operating
systems with R 4}
\usage{
util_fix_rstudio_bugs()
}
\value{
invisible null
}
\description{
RStudio crashes on parallel calls in some versions on Darwin based operating
systems with R 4
}
\seealso{
Other robustness_functions: 
\code{\link{util_as_valid_missing_codes}()},
\code{\link{util_check_one_unique_value}()},
\code{\link{util_correct_variable_use}()},
\code{\link{util_empty}()},
\code{\link{util_ensure_character}()},
\code{\link{util_ensure_in}()},
\code{\link{util_ensure_suggested}()},
\code{\link{util_expect_scalar}()},
\code{\link{util_is_integer}()},
\code{\link{util_is_numeric_in}()},
\code{\link{util_is_valid_missing_codes}()},
\code{\link{util_match_arg}()},
\code{\link{util_observations_in_subgroups}()},
\code{\link{util_only_NAs}()},
\code{\link{util_stop_if_not}()},
\code{\link{util_warn_unordered}()}
}
\concept{robustness_functions}
\concept{system}
\keyword{internal}
