% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_float_index_menu.R
\name{util_float_index_menu}
\alias{util_float_index_menu}
\title{return a single page navigation menu floating on the right}
\usage{
util_float_index_menu(index_menu_table, object)
}
\arguments{
\item{index_menu_table}{\link{data.frame} columns: links, hovers, texts}

\item{object}{\code{htmltools} tag list, used, instead of \code{index_menu_table},
if passed}
}
\description{
if displayed in a \code{dq_report2}
}
\examples{
\dontrun{
util_float_index_menu(tibble::tribble(
   ~ links, ~ hovers, ~ texts,
   "http://www.google.de/#xxx", "This is Google", "to Google",
   "http://www.uni-giessen.de/#xxx", "This is Gießen", "cruising on the A45"
))
}


}
\seealso{
Other reporting_functions: 
\code{\link{util_alias2caption}()},
\code{\link{util_copy_all_deps}()},
\code{\link{util_create_page_file}()},
\code{\link{util_eval_to_dataquieR_result}()},
\code{\link{util_evaluate_calls}()},
\code{\link{util_generate_anchor_link}()},
\code{\link{util_generate_anchor_tag}()},
\code{\link{util_generate_calls}()},
\code{\link{util_generate_calls_for_function}()},
\code{\link{util_load_manual}()},
\code{\link{util_make_data_slot_from_table_slot}()},
\code{\link{util_order_by_order}()},
\code{\link{util_set_size}()}
}
\concept{html}
\concept{reporting_functions}
\keyword{internal}
