% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generate_anchor_tag.R
\name{util_generate_anchor_link}
\alias{util_generate_anchor_link}
\title{Generate a link to a specific result}
\usage{
util_generate_anchor_link(
  varname,
  callname,
  order_context = c("variable", "indicator"),
  name,
  title
)
}
\arguments{
\item{varname}{variable to create a link to}

\item{callname}{function call to create a link to}

\item{order_context}{link created to variable overview or indicator
overview page}

\item{name}{replaces \code{varname} and \code{callname}, must contain the \code{.}
separator, then}

\item{title}{optional, replaces auto-generated link title}
}
\value{
the \code{htmltools} tag
}
\description{
for \code{dq_report2}
}
