% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generate_anchor_tag.R
\name{util_generate_anchor_tag}
\alias{util_generate_anchor_tag}
\title{Generate a tag for a specific result}
\usage{
util_generate_anchor_tag(
  varname,
  callname,
  order_context = c("variable", "indicator"),
  name
)
}
\arguments{
\item{varname}{variable to create an anchor for}

\item{callname}{function call to create an anchor for}

\item{order_context}{anchor created on variable overview or indicator
overview page}

\item{name}{replaces \code{varname} and \code{callname}, must contain the \code{.}
separator, then}
}
\value{
the \code{htmltools} tag
}
\description{
for \code{dq_report2}
}
