% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generate_calls_for_function.R
\name{util_generate_calls_for_function}
\alias{util_generate_calls_for_function}
\title{Generate function calls for a given indicator function}
\usage{
util_generate_calls_for_function(
  fkt,
  meta_data,
  label_col,
  meta_data_segment,
  meta_data_dataframe,
  meta_data_cross_item,
  specific_args,
  arg_overrides,
  resp_vars
)
}
\arguments{
\item{fkt}{the indicator function's name}

\item{meta_data}{the item level metadata data frame}

\item{label_col}{the label column}

\item{meta_data_segment}{segment level metadata}

\item{meta_data_dataframe}{data frame level metadata}

\item{meta_data_cross_item}{cross-item level metadata}

\item{specific_args}{argument overrides for specific functions}

\item{arg_overrides}{general argument overrides}

\item{resp_vars}{variables to be respected}
}
\value{
function calls for the given function
}
\description{
new reporting pipeline v2.0
}
\seealso{
Other reporting_functions: 
\code{\link{util_alias2caption}()},
\code{\link{util_copy_all_deps}()},
\code{\link{util_create_page_file}()},
\code{\link{util_eval_to_dataquieR_result}()},
\code{\link{util_evaluate_calls}()},
\code{\link{util_float_index_menu}()},
\code{\link{util_generate_anchor_link}()},
\code{\link{util_generate_anchor_tag}()},
\code{\link{util_generate_calls}()},
\code{\link{util_load_manual}()},
\code{\link{util_make_data_slot_from_table_slot}()},
\code{\link{util_make_function}()},
\code{\link{util_order_by_order}()},
\code{\link{util_set_size}()}
}
\concept{process}
\concept{reporting_functions}
\keyword{internal}
