% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_html_cell_for_result.R
\name{util_get_html_cell_for_result}
\alias{util_get_html_cell_for_result}
\title{Return the html summary table cell for a result}
\usage{
util_get_html_cell_for_result(
  result,
  aspect = c("applicability", "error", "issue"),
  rn,
  cn,
  ...
)
}
\arguments{
\item{result}{a \code{dataquieR_resultset2} result}

\item{aspect}{an aspect/problem category of results}

\item{rn}{row name for links inside a \code{dq_report2} report}

\item{cn}{column name for links inside a \code{dq_report2} report}

\item{...}{not used}
}
\value{
\link{character}\code{(1)} \code{html} result for a \code{DT} cell
}
\description{
Return the html summary table cell for a result
}
