% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_message_for_result.R
\name{util_get_message_for_result}
\alias{util_get_message_for_result}
\title{Return messages/warnings/notes/error messages for a result}
\usage{
util_get_message_for_result(
  result,
  aspect = c("applicability", "error", "issue"),
  collapse = "\\n<br />\\n",
  ...
)
}
\arguments{
\item{result}{a \code{dataquieR_resultset2} result}

\item{aspect}{an aspect/problem category of results}

\item{collapse}{either a lambda function or a separator for combining
multiple messages for the same result}

\item{...}{not used}
}
\value{
hover texts for results with data quality issues,
run-time errors, warnings or notes (aka messages)
}
\description{
Return messages/warnings/notes/error messages for a result
}
