% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_var_att_names_of_level.R
\name{util_get_var_att_names_of_level}
\alias{util_get_var_att_names_of_level}
\title{Get variable attributes of a certain provision level}
\usage{
util_get_var_att_names_of_level(level, cumulative = TRUE)
}
\arguments{
\item{level}{level(s) of requirement}

\item{cumulative}{include all names from more basic levels}
}
\value{
all matching variable attribute names
}
\description{
This function returns all variable attribute names of a certain
metadata provision level or of more than one level.
}
\seealso{
Other metadata_management: 
\code{\link{CAN_THIS_BE_REMOVED_util_combine_missing_lists}()},
\code{\link{util_dist_selection}()},
\code{\link{util_find_free_missing_code}()},
\code{\link{util_find_var_by_meta}()},
\code{\link{util_get_vars_in_segment}()},
\code{\link{util_looks_like_missing}()},
\code{\link{util_no_value_labels}()},
\code{\link{util_validate_known_meta}()},
\code{\link{util_validate_missing_lists}()}
}
\concept{metadata_management}
\keyword{internal}
