% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_var_att_names_of_level.R
\name{util_get_var_att_names_of_level}
\alias{util_get_var_att_names_of_level}
\title{Get variable attributes of a certain provision level}
\usage{
util_get_var_att_names_of_level(level, cumulative = TRUE)
}
\arguments{
\item{level}{level(s) of requirement}

\item{cumulative}{include all names from more basic levels}
}
\value{
all matching variable attribute names
}
\description{
This function returns all variable attribute names of a certain
meta data provision level or of more than one level.
}
