% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_hubert.R
\name{util_hubert}
\alias{util_hubert}
\title{utility function for the outliers rule of Huber et al.}
\usage{
util_hubert(x)
}
\arguments{
\item{x}{\link{numeric} data to check for outliers}
}
\value{
binary vector
}
\description{
function to calculate outliers according to the rule of Huber et al. This
function requires the package \code{robustbase}
}
