% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_interpret_limits.R
\name{util_interpret_limits}
\alias{util_interpret_limits}
\title{Utility function to interpret mathematical interval notation}
\usage{
util_interpret_limits(mdata)
}
\arguments{
\item{mdata}{\link{data.frame} the data frame that contains metadata
attributes of study data}
}
\value{
augments metadata by interpretable limit columns
}
\description{
Utility function to split limit definitions into interpretable elements
}
\seealso{
\link{util_validate_known_meta}

Other parser_functions: 
\code{\link{util_parse_assignments}()},
\code{\link{util_parse_interval}()},
\code{\link{util_parse_redcap_rule}()}
}
\concept{parser_functions}
\concept{robustness}
\keyword{internal}
