% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_interpret_range.R
\name{util_interpret_range}
\alias{util_interpret_range}
\title{Utility function to interpret mathematical interval notation for numeric
ranges}
\usage{
util_interpret_range(mdata)
}
\arguments{
\item{mdata}{\link{data.frame} the data frame that contains metadata
attributes of study data}
}
\value{
augments metadata by interpretable limit columns
}
\description{
Utility function to split range definitions into interpretable elements
}
