% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_is_na_0_empty_or_false.R
\name{util_is_na_0_empty_or_false}
\alias{util_is_na_0_empty_or_false}
\title{Detect falsish values}
\usage{
util_is_na_0_empty_or_false(x)
}
\arguments{
\item{x}{a value/vector of values}
}
\value{
vector of logical values:
\code{TRUE}, wherever x is somehow empty
}
\description{
Detect falsish values
}
