% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection.R
\name{util_load_manual}
\alias{util_load_manual}
\title{being called by the active binding function for .manual}
\usage{
util_load_manual(
  rebuild = FALSE,
  target = "inst/manual.RData",
  target2 = "inst/indicator_or_descriptor.RData",
  man_hash = ""
)
}
\arguments{
\item{rebuild}{rebuild the cache}

\item{target}{file for \code{..manual}}

\item{target2}{file for \code{..indicator_or_descriptor}}

\item{man_hash}{internal use: hash-sum over the manual to prevent rebuild
if not changed.}
}
\description{
being called by the active binding function for .manual
}
\seealso{
Other reporting_functions: 
\code{\link{util_alias2caption}()},
\code{\link{util_copy_all_deps}()},
\code{\link{util_create_page_file}()},
\code{\link{util_eval_to_dataquieR_result}()},
\code{\link{util_evaluate_calls}()},
\code{\link{util_float_index_menu}()},
\code{\link{util_generate_anchor_link}()},
\code{\link{util_generate_anchor_tag}()},
\code{\link{util_generate_calls}()},
\code{\link{util_generate_calls_for_function}()},
\code{\link{util_make_data_slot_from_table_slot}()},
\code{\link{util_order_by_order}()},
\code{\link{util_set_size}()}
}
\concept{reporting_functions}
\concept{system}
\keyword{internal}
