% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_looks_like_missing.R
\name{util_looks_like_missing}
\alias{util_looks_like_missing}
\title{Check for repetitive values using the digits 8 or 9 only}
\usage{
util_looks_like_missing(x, n_rules = 3)
}
\arguments{
\item{x}{\code{\link{numeric}} vector to test}

\item{n_rules}{\code{\link{numeric}} Only outlying values can be missing
codes; at least \code{n_rules} rules in
\code{\link{acc_univariate_outlier}} match}
}
\value{
\code{\link{logical}} indicates for each value in \code{x}, if it looks like a
missing code
}
\description{
Values not being finite (see \code{\link{is.finite}}) are also reported as missing
codes.
}
\seealso{
\code{\link{acc_univariate_outlier}}
}
