% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_make_function.R
\name{util_make_function}
\alias{util_make_function}
\title{Make a function capturing errors and other conditions for parallelization}
\usage{
util_make_function(fct)
}
\arguments{
\item{fct}{\link{function} to prepare}
}
\value{
decorated \link{function}
}
\description{
Make a function capturing errors and other conditions for parallelization
}
