% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_make_function.R
\name{util_make_function}
\alias{util_make_function}
\title{Make a function capturing errors and other conditions for parallelization}
\usage{
util_make_function(fct, caller.)
}
\arguments{
\item{fct}{\link{function} to prepare}

\item{caller.}{\link{call} for error messages, default is the caller of
\code{util_make_function}}
}
\value{
decorated \link{function}
}
\description{
Make a function capturing errors and other conditions for parallelization
}
\seealso{
Other reporting_functions: 
\code{\link{util_alias2caption}()},
\code{\link{util_copy_all_deps}()},
\code{\link{util_create_page_file}()},
\code{\link{util_eval_to_dataquieR_result}()},
\code{\link{util_evaluate_calls}()},
\code{\link{util_float_index_menu}()},
\code{\link{util_generate_anchor_link}()},
\code{\link{util_generate_anchor_tag}()},
\code{\link{util_generate_calls}()},
\code{\link{util_generate_calls_for_function}()},
\code{\link{util_load_manual}()},
\code{\link{util_make_data_slot_from_table_slot}()},
\code{\link{util_order_by_order}()},
\code{\link{util_set_size}()}
}
\concept{reporting_functions}
\concept{system}
\keyword{internal}
