% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_match_arg.R
\name{util_match_arg}
\alias{util_match_arg}
\title{\code{dataquieR} version of match.arg}
\usage{
util_match_arg(arg, choices, several_ok = FALSE, error = TRUE)
}
\arguments{
\item{arg}{the argument}

\item{choices}{the choices}

\item{several_ok}{allow more than one entry in \code{arg}}

\item{error}{\code{stop()}, if \code{arg} is not in \code{choices}
(warns and cleans \code{arg}, otherwise)}
}
\value{
"cleaned" \code{arg}
}
\description{
does not support partial matching, but will display the most likely
match as a warning/error.
}
\seealso{
Other robustness_functions: 
\code{\link{util_as_valid_missing_codes}()},
\code{\link{util_check_one_unique_value}()},
\code{\link{util_correct_variable_use}()},
\code{\link{util_empty}()},
\code{\link{util_ensure_character}()},
\code{\link{util_ensure_in}()},
\code{\link{util_ensure_suggested}()},
\code{\link{util_expect_scalar}()},
\code{\link{util_fix_rstudio_bugs}()},
\code{\link{util_is_integer}()},
\code{\link{util_is_numeric_in}()},
\code{\link{util_is_valid_missing_codes}()},
\code{\link{util_observations_in_subgroups}()},
\code{\link{util_stop_if_not}()},
\code{\link{util_warn_unordered}()}
}
\concept{robustness_functions}
\concept{string}
\keyword{internal}
