% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_data_frame_list.R
\name{util_merge_data_frame_list}
\alias{util_merge_data_frame_list}
\title{Combine data frames by merging}
\usage{
util_merge_data_frame_list(data_frames, id_vars)
}
\arguments{
\item{data_frames}{\link{list} of \link{data.frame}s}

\item{id_vars}{\link{character} the variable(s) to merge the data frames by.
each of them must exist in all data frames.}
}
\value{
\link{data.frame} combination of data frames
}
\description{
This is an extension of \code{merge} working for a list of data frames.
}
\seealso{
\link{prep_merge_study_data}

Other data_management: 
\code{\link{util_assign_levlabs}()},
\code{\link{util_check_data_type}()},
\code{\link{util_check_group_levels}()},
\code{\link{util_compare_meta_with_study}()},
\code{\link{util_dichotomize}()},
\code{\link{util_rbind}()},
\code{\link{util_remove_na_records}()},
\code{\link{util_replace_hard_limit_violations}()},
\code{\link{util_study_var2factor}()},
\code{\link{util_table_of_vct}()}
}
\concept{data_management}
\keyword{internal}
