% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_observations_in_subgroups.R
\name{util_observations_in_subgroups}
\alias{util_observations_in_subgroups}
\title{Utility function observations in subgroups}
\usage{
util_observations_in_subgroups(x, rvs)
}
\arguments{
\item{x}{data frame}

\item{rvs}{variable names}
}
\value{
matrix of flags
}
\description{
This function uses \code{!is.na} to count the number of non-missing observations in subgroups of
the data (list) and in a set of user defined response variables. In some applications it is
required that the number of observations per e.g. factor level is higher than a user-defined
minimum number.
}
