% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_only_NAs.R
\name{util_only_NAs}
\alias{util_only_NAs}
\title{identify \code{NA}-only variables}
\usage{
util_only_NAs(x)
}
\arguments{
\item{x}{the variable to check a vector}
}
\value{
flagged binary vector
}
\description{
This utility function identifies variables with \code{NA}s values only.
}
\seealso{
\link{util_correct_variable_use}

Other robustness_functions: 
\code{\link{util_as_valid_missing_codes}()},
\code{\link{util_check_one_unique_value}()},
\code{\link{util_correct_variable_use}()},
\code{\link{util_empty}()},
\code{\link{util_ensure_character}()},
\code{\link{util_ensure_in}()},
\code{\link{util_ensure_suggested}()},
\code{\link{util_expect_scalar}()},
\code{\link{util_fix_rstudio_bugs}()},
\code{\link{util_is_integer}()},
\code{\link{util_is_numeric_in}()},
\code{\link{util_is_valid_missing_codes}()},
\code{\link{util_match_arg}()},
\code{\link{util_observations_in_subgroups}()},
\code{\link{util_stop_if_not}()},
\code{\link{util_warn_unordered}()}
}
\concept{data_management}
\concept{robustness_functions}
\keyword{internal}
