% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_parse_assignments.R
\name{util_parse_assignments}
\alias{util_parse_assignments}
\title{Utility function to parse assignments}
\usage{
util_parse_assignments(text, split_char = SPLIT_CHAR)
}
\arguments{
\item{text}{Text to be parsed}

\item{split_char}{Character separating assignments}
}
\value{
the parsed assignments as a named list
}
\description{
This function parses labels & level assignments in the format
\code{1 = male | 2 = female}. The function also handles \code{m = male | f = female},
but this would not match the metadata concept. The split-character can
be given, if not the default from \link{SPLIT_CHAR} is to be used, but this
would also violate the metadata concept.
}
