% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_parse_interval.R
\name{util_parse_interval}
\alias{util_parse_interval}
\title{Utility function to parse intervals}
\usage{
util_parse_interval(int)
}
\arguments{
\item{int}{an interval as string, e.g., "[0;Inf)"}
}
\value{
the parsed interval with elements \code{inc_l} (Is the lower limit
included?), \code{low} (the value of the lower limit), \code{inc_u} (Is the upper
limit included?), \code{upp} (the value of the upper limit)
}
\description{
Utility function to parse intervals
}
