% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot_figure_no_plotly.R
\name{util_plot_figure_no_plotly}
\alias{util_plot_figure_no_plotly}
\title{Plot a \code{ggplot2} figure without \code{plotly}}
\usage{
util_plot_figure_no_plotly(x, sizing_hints = NULL)
}
\arguments{
\item{x}{\link[ggplot2:ggplot2-package]{ggplot2::ggplot2} object}

\item{sizing_hints}{\code{object} additional metadata about the natural figure size}
}
\value{
\code{htmltools} compatible object
}
\description{
Plot a \code{ggplot2} figure without \code{plotly}
}
\keyword{internal}
