% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_prep_proportion_check.R
\name{util_prep_proportion_check}
\alias{util_prep_proportion_check}
\title{Utility function to prepare the metadata for proportion checks}
\usage{
util_prep_proportion_check(
  resp_vars,
  meta_data,
  study_data,
  report_problems = c("error", "warning", "message")
)
}
\arguments{
\item{resp_vars}{\link{variable list} the names of the measurement variables}

\item{meta_data}{\link{data.frame} the data frame that contains metadata
attributes of study data}

\item{study_data}{\link{data.frame} the data frame that contains the measurements
(hint: missing value codes should be excluded,
so the function should be called with \code{ds1}, if
available)}

\item{report_problems}{\link{enum} Should missing metadata information be reported
as error, warning or message?}
}
\value{
a \link{list} with the expected range for the proportion check
}
\description{
Utility function to prepare the metadata for proportion checks
}
