% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pretty_vector_string.R
\name{util_pretty_vector_string}
\alias{util_pretty_vector_string}
\title{Prepare a vector four output}
\usage{
util_pretty_vector_string(v, quote = dQuote)
}
\arguments{
\item{v}{the vector}

\item{quote}{function, used for quoting -- \code{sQuote} or \code{dQuote}}
}
\value{
the "pretty" collapsed vector as a string.
}
\description{
Prepare a vector four output
}
