% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_replace_codes_by_na.R
\name{util_replace_codes_by_NA}
\alias{util_replace_codes_by_NA}
\title{Utility function to replace missing codes by \code{NA}s}
\usage{
util_replace_codes_by_NA(study_data, meta_data, split_char = SPLIT_CHAR)
}
\arguments{
\item{study_data}{Study data including jump/missing codes as specified in the
code conventions}

\item{meta_data}{Metadata as specified in the code conventions}

\item{split_char}{Character separating missing codes

Codes are expected to be numeric.}
}
\description{
Substitute all missing codes in a \link{data.frame} by \code{NA}.
}
