% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_replace_hard_limit_violations.R
\name{util_replace_hard_limit_violations}
\alias{util_replace_hard_limit_violations}
\title{Replace limit violations (HARD_LIMITS) by NAs}
\usage{
util_replace_hard_limit_violations(study_data, meta_data, label_col)
}
\arguments{
\item{study_data}{\link{study_data}}

\item{meta_data}{\link{study_data}}

\item{label_col}{\link{variable attribute} the name of the column in the metadata
with labels of variables}
}
\value{
modified \code{study_data}
}
\description{
Replace limit violations (HARD_LIMITS) by NAs
}
\seealso{
Other data_management: 
\code{\link{util_assign_levlabs}()},
\code{\link{util_check_data_type}()},
\code{\link{util_check_group_levels}()},
\code{\link{util_compare_meta_with_study}()},
\code{\link{util_dichotomize}()},
\code{\link{util_fix_merge_dups}()},
\code{\link{util_merge_data_frame_list}()},
\code{\link{util_rbind}()},
\code{\link{util_remove_na_records}()},
\code{\link{util_round_to_decimal_places}()},
\code{\link{util_study_var2factor}()},
\code{\link{util_table_of_vct}()}
}
\concept{data_management}
\concept{metadata_management}
\keyword{internal}
