% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_set_doublequotes.R
\name{util_set_dQuoteString}
\alias{util_set_dQuoteString}
\title{Utility function to put strings in quotes}
\usage{
util_set_dQuoteString(string)
}
\arguments{
\item{string}{Character vector}
}
\value{
quoted string
}
\description{
This function generates usual double-quotes for each element of the character
vector
}
