% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_set_size.R
\name{util_set_size}
\alias{util_set_size}
\title{Attaches attributes about the recommended minimum
absolute sizes to the plot p}
\usage{
util_set_size(p, width_em = NA_integer_, height_em = NA_integer_)
}
\arguments{
\item{p}{\link{ggplot} the plot}

\item{width_em}{\link{numeric} len=1. the minimum width hint in \code{em}}

\item{height_em}{\link{numeric} len=1. the minimum height in \code{em}}
}
\value{
p the modified plot
}
\description{
Attaches attributes about the recommended minimum
absolute sizes to the plot p
}
\seealso{
Other reporting_functions: 
\code{\link{util_alias2caption}()},
\code{\link{util_copy_all_deps}()},
\code{\link{util_create_page_file}()},
\code{\link{util_eval_to_dataquieR_result}()},
\code{\link{util_evaluate_calls}()},
\code{\link{util_float_index_menu}()},
\code{\link{util_generate_anchor_link}()},
\code{\link{util_generate_anchor_tag}()},
\code{\link{util_generate_calls}()},
\code{\link{util_generate_calls_for_function}()},
\code{\link{util_load_manual}()},
\code{\link{util_make_data_slot_from_table_slot}()},
\code{\link{util_make_function}()},
\code{\link{util_order_by_order}()}
}
\concept{figure}
\concept{reporting_functions}
\keyword{internal}
