% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_set_singlequotes.R
\name{util_set_sQuoteString}
\alias{util_set_sQuoteString}
\title{Utility function single quote string}
\usage{
util_set_sQuoteString(string)
}
\arguments{
\item{string}{Character vector}
}
\value{
quoted string
}
\description{
This function generates usual single-quotes for each element of the character
vector.
}
\seealso{
Other string_functions: 
\code{\link{util_abbreviate_unique}()},
\code{\link{util_filter_names_by_regexps}()},
\code{\link{util_pretty_vector_string}()},
\code{\link{util_set_dQuoteString}()},
\code{\link{util_sub_string_left_from_.}()},
\code{\link{util_sub_string_right_from_.}()},
\code{\link{util_translate}()}
}
\concept{process}
\concept{string_functions}
\keyword{internal}
