% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sort_by_order.R
\name{util_sort_by_order}
\alias{util_sort_by_order}
\title{Sort a vector by order given in some other vector}
\usage{
util_sort_by_order(x, order, ...)
}
\arguments{
\item{x}{the vector}

\item{order}{the "order vector}

\item{...}{additional arguments passed to \code{sort}}
}
\description{
Sort a vector by order given in some other vector
}
\examples{
\dontrun{
util_sort_by_order(c("a", "b", "a", "c", "d"), letters)
}

}
\seealso{
Other summary_functions: 
\code{\link{prep_combine_report_summaries}()},
\code{\link{prep_extract_classes_by_functions}()},
\code{\link{prep_extract_summary}()},
\code{\link{prep_extract_summary.dataquieR_result}()},
\code{\link{prep_extract_summary.dataquieR_resultset2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_ggplot2}()},
\code{\link{prep_render_pie_chart_from_summaryclasses_plotly}()},
\code{\link{prep_summary_to_classes}()},
\code{\link{util_as_cat}()},
\code{\link{util_as_integer_cat}()},
\code{\link{util_extract_indicator_metrics}()},
\code{\link{util_get_category_for_result}()},
\code{\link{util_get_colors}()},
\code{\link{util_get_labels_grading_class}()},
\code{\link{util_get_message_for_result}()},
\code{\link{util_get_rule_sets}()},
\code{\link{util_get_ruleset_formats}()},
\code{\link{util_get_thresholds}()},
\code{\link{util_html_table}()}
}
\concept{reporting}
\concept{summary_functions}
\keyword{internal}
