% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stop_if_not.R
\name{util_stop_if_not}
\alias{util_stop_if_not}
\title{Verify assumptions made by the code, that must be TRUE}
\usage{
util_stop_if_not(..., label, label_only)
}
\arguments{
\item{...}{see \code{\link{stopifnot}}}

\item{label}{\link{character} a label for the assumptions, can be missing}

\item{label_only}{\link{logical} if \code{TRUE} and \code{label} is given, the condition
will not be displayed, if \code{FALSE}}
}
\value{
\code{invisible(FALSE)}, if not stopped.
}
\description{
Verify assumptions made by the code, that must be TRUE
}
\seealso{
\code{\link{stopifnot}}

Other robustness_functions: 
\code{\link{util_as_valid_missing_codes}()},
\code{\link{util_check_one_unique_value}()},
\code{\link{util_correct_variable_use}()},
\code{\link{util_empty}()},
\code{\link{util_ensure_character}()},
\code{\link{util_ensure_in}()},
\code{\link{util_ensure_suggested}()},
\code{\link{util_expect_scalar}()},
\code{\link{util_fix_rstudio_bugs}()},
\code{\link{util_is_integer}()},
\code{\link{util_is_numeric_in}()},
\code{\link{util_is_valid_missing_codes}()},
\code{\link{util_match_arg}()},
\code{\link{util_observations_in_subgroups}()},
\code{\link{util_only_NAs}()},
\code{\link{util_warn_unordered}()}
}
\concept{condition}
\concept{robustness_functions}
\keyword{internal}
