% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_sort_by_order.R
\name{util_sub_string_right_from_.}
\alias{util_sub_string_right_from_.}
\title{Get sub-string right from first \code{.}}
\usage{
util_sub_string_right_from_.(x)
}
\arguments{
\item{x}{the string with a least one \code{.}}
}
\description{
Get sub-string right from first \code{.}
}
\examples{
\dontrun{
util_sub_string_right_from_.(c("a.b", "asdf.xyz"))
util_sub_string_right_from_.(c("a.b", "asdf.xy.z"))
util_sub_string_right_from_.(c("ab", "asdxy.z"))
}
}
