% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_table_of_vct.R
\name{util_table_of_vct}
\alias{util_table_of_vct}
\title{Tabulate a vector}
\usage{
util_table_of_vct(Var1)
}
\arguments{
\item{Var1}{vector to ta tabulate}
}
\value{
a data frame with columns \code{Var1} and \code{Freq}
}
\description{
does the same as \code{as.data.frame(table(x))} but
guarantees a data frame with two columns is returned
}
\seealso{
Other data_management: 
\code{\link{util_assign_levlabs}()},
\code{\link{util_check_data_type}()},
\code{\link{util_check_group_levels}()},
\code{\link{util_compare_meta_with_study}()},
\code{\link{util_dichotomize}()},
\code{\link{util_merge_data_frame_list}()},
\code{\link{util_rbind}()},
\code{\link{util_remove_na_records}()},
\code{\link{util_replace_hard_limit_violations}()},
\code{\link{util_study_var2factor}()}
}
\concept{data_management}
\concept{robustness}
\keyword{internal}
