% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_translate_indicator_metrics.R
\name{util_translate_indicator_metrics}
\alias{util_translate_indicator_metrics}
\title{Translate standard column names to readable ones}
\usage{
util_translate_indicator_metrics(
  colnames,
  short = FALSE,
  long = TRUE,
  ignore_unknown = FALSE
)
}
\arguments{
\item{colnames}{\link{character} the names to translate}

\item{short}{\link{logical} include unit letter in output}

\item{long}{\link{logical} include unit description in output}

\item{ignore_unknown}{\link{logical} do not replace unknown indicator metrics
by \code{NA}, keep them}
}
\value{
translated names
}
\description{
TODO: Duplicate of util_make_data_slot_from_table_slot ??
}
\keyword{internal}
