% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_view_file.R
\name{util_view_file}
\alias{util_view_file}
\title{View a file in most suitable viewer}
\usage{
util_view_file(file)
}
\arguments{
\item{file}{the file to view}
}
\value{
\code{invisible(file)}
}
\description{
View a file in most suitable viewer
}
\seealso{
Other system_functions: 
\code{\link{util_detect_cores}()},
\code{\link{util_user_hint}()}
}
\concept{reporting}
\concept{system_functions}
\keyword{internal}
