% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_warning.R
\name{util_warning}
\alias{util_warning}
\title{Produce a warning message with a useful short stack trace.}
\usage{
util_warning(m, ..., immediate. = FALSE)
}
\arguments{
\item{m}{warning message or a \link{simpleWarning}}

\item{...}{arguments for \link{sprintf} on m, if m is a character}

\item{immediate.}{logical, indicating if the call should be output
immediately, even if \code{getOption("warn") <= 0}.
See also \link[base:warning]{base::warning}.}
}
\value{
invisible(NULL).
}
\description{
Produce a warning message with a useful short stack trace.
}
