% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globs.R
\docType{data}
\name{VARIABLE_LIST}
\alias{VARIABLE_LIST}
\title{Cross-item level metadata attribute name}
\format{
An object of class \code{character} of length 1.
}
\usage{
VARIABLE_LIST
}
\description{
Specifies a group of variables for multivariate analyses. Separated
by |, please use variable names from \link{VAR_NAMES} or
a label as specified in \code{label_col}, usually \link{LABEL} or \link{LONG_LABEL}.
}
\details{
if missing, \code{dataquieR} will create such IDs from \link{CONTRADICTION_TERM},
if specified.
}
\seealso{
\link{meta_data_cross}

Other meta_data_cross: 
\code{\link{ASSOCIATION_DIRECTION}},
\code{\link{ASSOCIATION_FORM}},
\code{\link{ASSOCIATION_METRIC}},
\code{\link{ASSOCIATION_RANGE}},
\code{\link{CHECK_ID}},
\code{\link{CHECK_LABEL}},
\code{\link{CONTRADICTION_TERM}},
\code{\link{CONTRADICTION_TYPE}},
\code{\link{DATA_PREPARATION}},
\code{\link{GOLDSTANDARD}},
\code{\link{MULTIVARIATE_OUTLIER_CHECKTYPE}},
\code{\link{N_RULES}},
\code{\link{REL_VAL}},
\code{\link{util_normalize_cross_item}()}
}
\concept{meta_data_cross}
\keyword{datasets}
